### 2020-10-02 #401 @wlandau
targets and tarchetypes
Submitting Author: Name (@wlandau)
Repository:
* `targets`: https://github.com/wlandau/targets
* `tarchetypes`: https://github.com/wlandau/tarchetypes. (`tarchetypes` is a small companion package that only serves to extend `targets`, so thought it appropriate to submit it for review along with `targets`.)
* The manual: https://github.com/wlandau/targets-manual. (For https://github.com/ropensci-books.)

Version submitted:
* 0.0.0.9002 (`targets`)
* 0.0.0.9000 (`tarchetypes` and the manual)

Version accepted: TBD

---

## DESCRIPTION files

* `targets`

```
Package: targets
Title: Dynamic Function-Oriented Make-Like Declarative Pipelines for R
Description: The targets package is a pipeline toolkit that brings together
  function-oriented programming and Make-like declarative workflows for
  Statistics and data science in R. It implements a workflow as collection of
  interconnected tasks, analyzes the dependency relationships among these
  tasks, skips steps that are already up to date, runs the necessary
  computations with optional parallel workers, abstracts files as
  R objects, and provides tangible evidence that the results match
  the underlying code and data. The methodology in this package
  borrows from GNU Make by Richard Stallman (2015, ISBN:978-9881443519)
  and drake by Will Landau (2018) <doi:10.21105/joss.00550>.
Version: 0.0.0.9001
  person(
    given = c("William", "Michael"),
    family = "Landau",
    email = "will.landau@gmail.com",
    comment = c(ORCID = "0000-0003-1878-3253")
  ),
  person(
    family = "Eli Lilly and Company",
    role = "cph"
  ),
  person(
    given = c("Matthew", "T."),
    family = "Warkentin",
    role = "ctb"
  ))
  callr (>= 3.4.3),
  cli (>= 2.0.2),
  codetools (>= 0.2.16),
  data.table (>= 1.12.8),
  digest (>= 0.6.25),
  igraph (>= 1.2.5),
  R6 (>= 2.4.1),
  rlang (>= 0.4.5),
  tibble (>= 3.0.1),
  tidyselect (>= 1.1.0),
  vctrs (>= 0.2.4),
  withr (>= 2.1.2)
  aws.s3 (>= 0.3.21),
  clustermq (>= 0.8.9),
  curl (>= 4.3),
  fst (>= 0.9.2),
  future (>= 1.17.0),
  keras (>= 2.2.5.0),
  knitr (>= 1.30),
  rmarkdown (>= 2.4),
  qs (>= 0.23.2),
  rstudioapi (>= 0.11),
  torch (>= 0.1.0),
  usethis (>= 1.6.3),
  visNetwork (>= 2.0.9)
RoxygenNote: 7.1.1.9000
```

* `tarchetypes`

```
Package: tarchetypes
Title: Archetypes for Targets
Description: The targets package is a pipeline toolkit that brings together
  function-oriented programming and Make-like declarative workflows for
  Statistics and data science in R. The tarchetypes package provides
  convenient user-side functions to create specialized targets,
  making pipelines easier to create and read. The methods in this package
  were influenced by the drake R package by Will Landau (2018)
  <doi:10.21105/joss.00550>.
Version: 0.0.0.9000
  person(
    given = c("William", "Michael"),
    family = "Landau",
    email = "will.landau@gmail.com",
    comment = c(ORCID = "0000-0003-1878-3253")
  ),
  person(
    family = "Eli Lilly and Company",
    role = "cph"
  ))
  fs (>= 1.4.2),
  rlang (>= 0.4.7),
  targets,
  tidyselect (>= 1.1.0),
  vctrs (>= 0.3.4),
  withr (>= 2.1.2)
  digest (>= 0.6.25),
  knitr (>= 1.28),
  rmarkdown (>= 2.1),
  wlandau/targets
RoxygenNote: 7.1.1.9000
```

* Manual

```
Package: targets.manual
Title: Targets R Package User Manual
Description: This repository contains the source files of the targets R
  package user manual.
Version: 0.0.0.9000
  https://github.com/wlandau/targets-manual
  person(
    given = c("William", "Michael"),
    family = "Landau",
    email = "will.landau@gmail.com",
    comment = c(ORCID = "0000-0003-1878-3253")
  ),
  person(
    family = "Eli Lilly and Company",
    role = "cph"
  ))
  biglm (>= 0.9.2),
  bookdown (>= 0.19),
  fs (>= 1.4.1),
  purrr (>= 0.3.4),
  tarchetypes,
  targets,
  tidyverse (>= 1.3.0),
  visNetwork (>= 2.0.9),
  withr (>= 2.2.0)
  wlandau/tarchetypes,
  wlandau/targets
```

`targets` is an R-focused pipeline toolkit for Make-like declarative workflows. It resolves the dependency relationships among steps of a data analysis workflow and skips steps that are already up to date.

`targets` is for R users who maintain computationally intense function-oriented data analysis projects (with large codebases and/or long runtimes). Such projects may include but are not limited to Bayesian statistics, simulation, machine learning, PK/PD, and spatial statistics.

`targets` is the long-term successor to [`drake`](https://github.com/ropensci/drake). After four years of development, [`drake`](https://github.com/ropensci/drake) has improved so much that its insurmountable problems have become its most pressing ones. A new package is necessary to advance the capability further. So while I still believe [`drake`](https://github.com/ropensci/drake) is thriving, and even though I will continue to maintain [`drake`](https://github.com/ropensci/drake) indefinitely, I created `targets` to try to break new ground. At <https://wlandau.github.io/targets/articles/need.html#drake>, I take a detailed dive into the ways that `targets` surpasses [`drake`](https://github.com/ropensci/drake)'s permanent limitations.

N/A

N/A

- [x] contains a vignette with examples of its essential functions and uses. **To increase modularity and reduce package check time, all the user-side vignettes actually live at https://github.com/wlandau/targets-manual (deployed to https://wlandau.github.io/targets-manual). Repos at https://github.com/wlandau/targets-minimal, https://github.com/wlandau/targets-stan, and https://github.com/wlandau/targets-keras have the complete source code for example use cases. To avoid encumbering core `targets` and to avoid maintaining duplicated documentation, the vignettes of the actual package only include the statement of need and design documents. The README is deliberately short and links to all this existing documentation.**
- [x] has a [test suite](https://ropensci.github.io/dev_guide/building.html#testing). **Not all of the tests in `targets` can be automated, especially when it comes to visualization and profiling, so many of the tests live in non-`testthat` folders in https://github.com/wlandau/targets/tree/master/tests. Whenever I use a `#nocov` block, I always include a comment with a justification and/or a reference to one of these semi-automated tests.**

    - [x] The package contains a `paper.md` matching [JOSS's requirements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) with a high-level description in the package root or in `inst/`. ~~**I have written a paper.md, but I need to run it through my company's scientific disclosure process before I share it. That could take a few weeks.**~~ `paper.md` and `paper.bib` now disclosed and included inside `inst/`.
