### 2020-10-03 ae09ee3
Enforce type guarantees in record_init()
