### 2020-10-03 bf92824
Standardize internal types
