### 2020-10-03 ec56545
Align internal NA's
