+++ tests/testthat/test-class_stem.R
@@ -158,10 +158,10 @@ tar_test("stem$produce_record() of a successful stem", {
   expect_true(record$time > 0)
   expect_equal(record$format, "rds")
   expect_equal(record$iteration, "vector")
-  expect_equal(record$children, character(0))
+  expect_equal(record$children, NA_character_)
   expect_true(is.numeric(record$seconds))
-  expect_equal(record$warnings, character(0))
-  expect_equal(record$error, character(0))
+  expect_equal(record$warnings, NA_character_)
+  expect_equal(record$error, NA_character_)
 })
 
 tar_test("stem$produce_record() of a errored stem", {
@@ -177,15 +177,15 @@ tar_test("stem$produce_record() of a errored stem", {
   expect_equal(record$type, "stem")
   expect_equal(nchar(record$command), 16L)
   expect_equal(nchar(record$depend), 16L)
-  expect_equal(record$path, character(0))
+  expect_equal(record$path, NA_character_)
   expect_equal(record$data, NA_character_)
   expect_equal(record$bytes, 0)
   expect_equal(record$time, -Inf)
   expect_equal(record$format, "rds")
   expect_equal(record$iteration, "vector")
-  expect_equal(record$children, character(0))
+  expect_equal(record$children, NA_character_)
   expect_true(is.numeric(record$seconds))
-  expect_equal(record$warnings, character(0))
+  expect_equal(record$warnings, NA_character_)
   expect_equal(record$error, "123 .")
 })
 
