### 2020-10-03 2d9e118
Improve error messages and docs
