### 2020-10-03 ceb2603
Write comparisons more defensively
