### 2020-10-03 39fd9c6
Compare more defensively
