### 2020-10-03 ffbd209
Compare more defensively
