+++ man/tar_option_set.Rd
@@ -67,8 +67,8 @@ such as a tensor or neural network module.
 \item \code{"file"}: A dynamic file. To use this format,
 the target needs to manually identify or save some data
 and return a character vector of paths
-to the data. Then, \code{targets} automatically checks those files and cues the
-appropriate build decisions if those files are out of date.
+to the data. Then, \code{targets} automatically checks those files and cues
+the appropriate build decisions if those files are out of date.
 Those paths must point to files or directories,
 and they must not contain characters \code{|} or \code{*}.
 All the files and directories you return must actually exist,
