+++ tests/performance/test-parallel.R
@@ -17,7 +17,7 @@ tar_script({
 tar_make_clustermq(workers = 3L)
 tar_destroy()
 
-# If there are problems, use the example below
+# If there are problems, inspect the example below.
 envir <- globalenv()
 pipeline <- pipeline_init(
   list(
