### 2020-10-03 cda7b46
Use timestamp reporter
