+++ tests/aws/test-class_aws_s3_file.R
@@ -1,3 +1,4 @@
+# Use sparingly. We do not want to max out any AWS quotas.
 tar_test("aws_file format file gets stored", {
   skip_if_no_aws()
   bucket_name <- random_bucket_name()
