### 2020-10-03 9b8f33b
Fix timestamp behavior
