+++ R/class_file.R
@@ -44,8 +44,8 @@ file_update_hash <- function(file) {
 
 file_should_rehash <- function(file, bytes, time) {
   small <- bytes < 1e5
-  touched <- time > (file$time - 1e-9)
-  resized <- abs(bytes - file$bytes) > 1e-9
+  touched <- time > (file$time + 1e-6)
+  resized <- abs(bytes - file$bytes) > 1e-6
   small || touched || resized
 }
 
@@ -122,9 +122,9 @@ file_info <- function(files) {
 
 file_bytes <- function(info) {
   # Cannot be integer because of large value.
-  round(sum(replace_na(info$size, 0)), 9)
+  round(sum(replace_na(info$size, 0)), 6)
 }
 
 file_time <- function(info) {
-  round(max(replace_na(as.numeric(info$mtime), -Inf) %|||% -Inf), 9)
+  round(max(replace_na(as.numeric(info$mtime), -Inf) %|||% -Inf), 6)
 }
