+++ R/class_active.R
@@ -57,6 +57,16 @@ active_class <- R6::R6Class(
     unserialize_target = function(target) {
       builder_unserialize_value(target)
     },
+    process_target = function(name) {
+      target <- pipeline_get_target(self$pipeline, name)
+      target_debug(target)
+      target_update_depend(target, meta)
+      trn(
+        target_should_run(target, self$meta),
+        self$run_target(name),
+        self$skip_target(target)
+      )
+    },
     start = function() {
       pipeline_prune_names(self$pipeline, self$names)
       self$update_scheduler()
