+++ R/class_clustermq.R
@@ -139,16 +139,6 @@ clustermq_class <- R6::R6Class(
         self$meta
       )
     },
-    process_target = function(name) {
-      target <- pipeline_get_target(self$pipeline, name)
-      target_debug(target)
-      target_update_depend(target, meta)
-      trn(
-        target_should_run(target, self$meta),
-        self$run_target(name),
-        self$skip_target(target)
-      )
-    },
     wait = function() {
       self$crew$send_wait()
       Sys.sleep(0.001)
