### 2020-10-03 16f564a
Fix method name and add placeholder for store_sync_timestamp()
