+++ NAMESPACE
@@ -81,6 +81,9 @@ S3method(store_read_path,tar_torch)
 S3method(store_read_path,tar_url)
 S3method(store_serialize_value,tar_keras)
 S3method(store_serialize_value,tar_torch)
+S3method(store_sync_timestamp,default)
+S3method(store_sync_timestamp,tar_aws_s3)
+S3method(store_sync_timestamp,tar_url)
 S3method(store_unserialize_value,default)
 S3method(store_unserialize_value,tar_keras)
 S3method(store_unserialize_value,tar_torch)
@@ -93,6 +96,7 @@ S3method(store_validate_packages,tar_fst_tbl)
 S3method(store_validate_packages,tar_keras)
 S3method(store_validate_packages,tar_qs)
 S3method(store_validate_packages,tar_torch)
+S3method(store_validate_packages,tar_url)
 S3method(store_write_object,default)
 S3method(store_write_object,tar_aws_s3)
 S3method(store_write_object,tar_store_file)
