### 2020-10-03 87193c3
Need to start using hashes for time stamps
