+++ NAMESPACE
@@ -81,9 +81,9 @@ S3method(store_read_path,tar_torch)
 S3method(store_read_path,tar_url)
 S3method(store_serialize_value,tar_keras)
 S3method(store_serialize_value,tar_torch)
-S3method(store_sync_timestamp,default)
-S3method(store_sync_timestamp,tar_aws_s3)
-S3method(store_sync_timestamp,tar_url)
+S3method(store_sync_file_meta,default)
+S3method(store_sync_file_meta,tar_aws_s3)
+S3method(store_sync_file_meta,tar_url)
 S3method(store_unserialize_value,default)
 S3method(store_unserialize_value,tar_keras)
 S3method(store_unserialize_value,tar_torch)
