### 2020-10-03 87e486e
Change timestamp to character value
