+++ R/class_branch.R
@@ -74,8 +74,8 @@ target_produce_record.tar_branch <- function(target, meta) {
     depend = meta$get_depend(target_get_name(target)),
     path = file$path,
     data = file$hash,
-    bytes = file$bytes,
     time = file$time,
+    bytes = file$bytes,
     format = target$settings$format,
     iteration = target$settings$iteration,
     seconds = target$metrics$seconds,
