+++ R/class_stem.R
@@ -65,8 +65,8 @@ target_produce_record.tar_stem <- function(target, meta) {
     depend = meta$get_depend(target_get_name(target)),
     path = file$path,
     data = file$hash,
-    bytes = file$bytes,
     time = file$time,
+    bytes = file$bytes,
     format = target$settings$format,
     iteration = target$settings$iteration,
     children = target_get_children(target),
