+++ R/class_file.R
@@ -2,15 +2,15 @@ file_init <- function(
   path = character(0),
   stage = character(0),
   hash = NA_character_,
-  bytes = 0,
-  time = NA_character_
+  time = NA_character_,
+  bytes = 0
 ) {
   file_new(
     path = path,
     stage = stage,
     hash = hash,
-    bytes = bytes,
-    time = time
+    time = time,
+    bytes = bytes
   )
 }
 
@@ -18,14 +18,14 @@ file_new <- function(
   path = NULL,
   stage = NULL,
   hash = NULL,
-  bytes = NULL,
-  time = NULL
+  time = NULL,
+  bytes = NULL
 ) {
   force(path)
   force(stage)
   force(hash)
-  force(bytes)
   force(time)
+  force(bytes)
   environment()
 }
 
@@ -42,7 +42,7 @@ file_update_hash <- function(file) {
   invisible()
 }
 
-file_should_rehash <- function(file, bytes, time) {
+file_should_rehash <- function(file, time, bytes) {
   small <- bytes < file_bound_bytes
   touched <- !identical(time, file$time)
   resized <- abs(bytes - file$bytes) > file_tol_bytes
@@ -58,7 +58,7 @@ file_ensure_hash <- function(file) {
   info <- file_info(files)
   bytes <- file_bytes(info)
   time <- file_time(info)
-  do <- file_should_rehash(file, bytes, time)
+  do <- file_should_rehash(file = file, time = time, bytes = bytes)
   hash <- trn(do, file_hash(files), file$hash)
   file$hash <- hash
   file$bytes <- bytes
@@ -71,7 +71,7 @@ file_has_correct_hash <- function(file) {
   bytes <- file_bytes(info)
   time <- file_time(info)
   trn(
-    file_should_rehash(file, bytes, time),
+    file_should_rehash(file = file, time = time, bytes = bytes),
     identical(file$hash, file_hash(files)),
     TRUE
   )
