### 2020-10-03 82ba135
Add size hash
