+++ R/class_branch.R
@@ -75,6 +75,7 @@ target_produce_record.tar_branch <- function(target, meta) {
     path = file$path,
     data = file$hash,
     time = file$time,
+    size = file$size,
     bytes = file$bytes,
     format = target$settings$format,
     iteration = target$settings$iteration,
