+++ R/class_stem.R
@@ -66,6 +66,7 @@ target_produce_record.tar_stem <- function(target, meta) {
     path = file$path,
     data = file$hash,
     time = file$time,
+    size = file$size,
     bytes = file$bytes,
     format = target$settings$format,
     iteration = target$settings$iteration,
