### 2020-10-03 902a5b1
Use size in cue_file()
