### 2020-10-03 d1030f5
Check bytes too for rehashing
