### 2020-10-03 5763a0e
Only sync metadata if target is skipped
