+++ R/class_target.R
@@ -318,6 +318,14 @@ target_debug <- function(target) {
 target_debug.default <- function(target) {
 }
 
+target_sync_file_meta <- function(target, meta) {
+  UseMethod("target_sync_file_meta")
+}
+
+#' @export
+target_sync_file_meta.default <- function(target, meta) {
+}
+
 target_validate <- function(target) {
   UseMethod("target_validate")
 }
