+++ R/class_builder.R
@@ -145,6 +145,11 @@ target_debug.tar_builder <- function(target) {
 }
 
 #' @export
+target_sync_file_meta.tar_builder <- function(target, meta) {
+  store_sync_file_meta(target$store, target, meta)
+}
+
+#' @export
 target_validate.tar_builder <- function(target) {
   NextMethod()
   if (!is.null(target$store)) {
