+++ NEWS.md
@@ -1,6 +1,12 @@
 # targets 0.0.0.9002
 
-* Breaking change: add a new `size` field to the metadata to allow `targets` to make better judgments about when to rehash files (#180). We now compare hashes to check file size differences instead of doing messy floating point comparisons with ad hoc tolerances. It breaks back compatibility with old projects, but the error message is informative, and this is all still before the first official release.
+## Breaking changes
+
+* Add a new `size` field to the metadata to allow `targets` to make better judgments about when to rehash files (#180). We now compare hashes to check file size differences instead of doing messy floating point comparisons with ad hoc tolerances. It breaks back compatibility with old projects, but the error message is informative, and this is all still before the first official release.
+
+# Bug fixes
+
+* Stop sending target return values over the network when `storage` is `"remote"` (#182, @mattwarkentin).
 
 # targets 0.0.0.9000
 
