### 2020-10-05 daa7d43
Avoid sending workspace data over the network
