+++ R/tar_option_get.R
@@ -35,7 +35,7 @@ tar_option_default <- function(option) {
     iteration = "vector",
     error = "stop",
     memory = "persistent",
-    deployment = "remote",
+    deployment = "worker",
     priority = 0,
     resources = list(),
     storage = "local",
