+++ R/class_settings.R
@@ -8,8 +8,8 @@ settings_init <- function(
   deployment = "worker",
   priority = 0,
   resources = list(),
-  storage = "local",
-  retrieval = "local"
+  storage = "master",
+  retrieval = "master"
 ) {
   growth <- all.vars(pattern, functions = TRUE, max.names = 1L) %|||% "none"
   dimensions <- all.vars(pattern, functions = FALSE)
@@ -99,12 +99,12 @@ settings_validate <- function(settings) {
   assert_chr(settings$iteration)
   assert_in(settings$error, c("stop", "continue", "save"))
   assert_in(settings$memory, c("persistent", "transient"))
-  assert_in(settings$deployment, c("local", "worker"))
+  assert_in(settings$deployment, c("master", "worker"))
   assert_scalar(settings$priority)
   assert_ge(settings$priority, 0)
   assert_le(settings$priority, 1)
   assert_list(settings$resources)
-  assert_in(settings$storage, c("local", "worker"))
-  assert_in(settings$retrieval, c("local", "worker"))
+  assert_in(settings$storage, c("master", "worker"))
+  assert_in(settings$retrieval, c("master", "worker"))
   invisible()
 }
