+++ man/tar_manifest.Rd
@@ -1,0 +1,91 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/tar_manifest.R
+\name{tar_manifest}
+\alias{tar_manifest}
+\title{Produce a data frame of information about your targets.}
+\usage{
+tar_manifest(
+  names = NULL,
+  fields = c("name", "command", "type", "dimensions"),
+  callr_function = callr::r,
+  callr_arguments = list()
+)
+}
+\arguments{
+\item{names}{Names of the targets to show. Set to \code{NULL} to
+show all the targets (default). Otherwise, you can supply
+symbols, a character vector, or \code{tidyselect} helpers like \code{\link[=starts_with]{starts_with()}}.}
+
+\item{fields}{Names of the fields, or columns, to show. Set to \code{NULL} to
+show all the fields (default). Otherwise, you can supply
+symbols, a character vector, or \code{tidyselect} helpers like \code{\link[=starts_with]{starts_with()}}.
+Set to \code{NULL} to print all the fields.
+The name of the target is always included as the first column
+regardless of the selection.
+Possible fields are below. All of them can be set in \code{\link[=tar_target]{tar_target()}},
+\code{\link[=tar_target_raw]{tar_target_raw()}}, or \code{\link[=tar_option_set]{tar_option_set()}}.
+\itemize{
+\item \code{name}: Name of the target.
+\item \code{command}: the R command that runs when the target builds.
+\item \code{type}: Type of the target, either \code{"stem"} (non-pattern) or \code{"map"}
+or \code{"cross"} (patterns).
+\item \code{dimensions}: A list column to name the targets that each pattern
+branches over. Only applies to patterns.
+\item \code{format}: Storage format.
+\item \code{iteration}: Iteration mode for branching.
+\item \code{error}: Error mode, what to do when the target fails.
+\item \code{memory}: Memory mode, when to keep targets in memory.
+\item \code{storage}: Storage mode for high-performance computing scenarios.
+\item \code{retrieval}: Retrieval mode for high-performance computing scenarios.
+\item \code{deployment}: Where/whether to deploy the target in high-performance
+computing scenarios.
+\item \code{resources}: A list of target-specific resource requirements for
+\code{\link[=tar_make_future]{tar_make_future()}}.
+\item \code{cue_mode}: Cue mode from \code{\link[=tar_cue]{tar_cue()}}.
+\item \code{cue_depend}: Depend cue from \code{\link[=tar_cue]{tar_cue()}}.
+\item \code{cue_expr}: Command cue from \code{\link[=tar_cue]{tar_cue()}}.
+\item \code{cue_file}: File cue from \code{\link[=tar_cue]{tar_cue()}}.
+\item \code{cue_format}: Format cue from \code{\link[=tar_cue]{tar_cue()}}.
+\item \code{cue_iteration}: Iteration cue from \code{\link[=tar_cue]{tar_cue()}}.
+\item \code{packages}: List columns of packages loaded before building the target.
+\item \code{library}: List column of library paths to load the packages.
+}}
+
+\item{callr_function}{A function from \code{callr} to start a fresh clean R
+process to do the work. Set to \code{NULL} to run in the current session
+instead of an external process (but restart your R session just before
+you do in order to clear debris out of the global environment).
+\code{callr_function} needs to be \code{NULL} for interactive debugging,
+e.g. \code{tar_option_set(debug = "your_target")}.
+However, \code{callr_function} should not be \code{NULL} for serious
+reproducible work.}
+
[…]
