+++ man/tar_dir.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/tar_dir.R
+\name{tar_dir}
+\alias{tar_dir}
+\title{Execute code in a temporary directory.}
+\usage{
+tar_dir(code)
+}
+\arguments{
+\item{code}{User-defined code.}
+}
+\value{
+Return value of the user-defined code.
+}
+\description{
+Creates a new directory at \code{tempfile()} and runs code
+inside with \code{withr::with_dir()}
+}
+\examples{
+tar_dir(file.create("x"))
+file.exists("x") # Should be false.
+}
+\keyword{internal}

[…]
