### 2020-10-06 #401 @wlandau
`tarchetypes` is a collection of user-level utilities for `targets`, and it does not stand on its own. Its only purpose is to make `targets` easier to use, and it requires prior familiarity with the latter. I originally considered implementing both in a single package, and I still view their physical separation as a minor implementation detail. So I thought that separate submissions would potentially duplicate the overall reviewer-side overhead of learning `targets` in the first place, especially compared to the relatively small size of `tarchetypes`. In addition, `tarchetypes` helpers like `tar_render()` also appear in the [manual](https://github.com/wlandau/targets-manual).
