### 2020-10-06 0004d50
Align internals with API changes
