+++ tests/hpc/test-future.R
@@ -1,4 +1,4 @@
-test_that("packages are actually loaded for remote targets", {
+test_that("packages are actually loaded", {
   # Needs sge_batchtools.tmpl (in current directory).
   unlink("_targets", recursive = TRUE)
   on.exit(unlink("_targets", recursive = TRUE))
@@ -108,7 +108,7 @@ test_that("branching plan on SGE", {
   }
 })
 
-test_that("Same with remote storage", {
+test_that("Same with worker-side storage", {
   # Needs sge_batchtools.tmpl (in current directory).
   unlink("_targets", recursive = TRUE)
   on.exit(unlink("_targets", recursive = TRUE))
