### 2020-10-06 5736b3a
Lighten blue color of outdated targets
