### 2020-10-06 cfc1120
minor spelling fix
