+++ DESCRIPTION
@@ -1,11 +1,11 @@
 Package: targets
-Title: Dynamic Function-Oriented 'Make'-Like Declarative Workflows for R
+Title: Dynamic Function-Oriented 'Make'-Like Declarative Workflows
 Description: The 'targets' package is a pipeline toolkit that brings together
   function-oriented programming and 'Make'-like declarative workflows for
   Statistics and data science in R. It implements a workflow as collection of
   interconnected tasks, analyzes the dependency relationships among these
   tasks, skips steps that are already up to date, runs the necessary
-  computations with optional parallel workers, abstracts files as
+  computation with optional parallel workers, abstracts files as
   R objects, and provides tangible evidence that the results match
   the underlying code and data. The methodology in this package
   borrows from GNU 'Make' by Richard Stallman (2015, ISBN:978-9881443519)
