### 2020-10-07 532bbac
Exclude tar_pipeline and tar_target objects from dependency detection
