+++ tests/testthat/test-utils_imports.R
@@ -104,3 +104,28 @@ tar_test("hash_imports() after trivial formatting change", {
   hashes2 <- hashes2[order(hashes2$name), ]
   expect_equivalent(hashes1, hashes2)
 })
+
+tar_test("exclude special objects from imports", {
+  tar_script({
+    envir <- new.env(parent = globalenv())
+    tar_option_set(envir = envir)
+    envir$regular_object <- "object"
+    envir$target_object <- tar_target(x, regular_object)
+    envir$pipeline_object <- tar_pipeline(
+      envir$target_object,
+      tar_target(y, x)
+    )
+    envir$pipeline_object
+  })
+  out <- tar_network(callr_function = NULL, targets_only = FALSE)
+  expect_true(length(out$vertices$name) > 0L)
+  expect_true(length(out$edges$from) > 0L)
+  expect_true(length(out$edges$to) > 0L)
+  for (exclude in c("pipeline_object", "target_object")) {
+    expect_false(exclude %in% out$vertices$name)
+    expect_false(exclude %in% out$edges$from)
+    expect_false(exclude %in% out$edges$to)
+  }
+  expect_true("regular_object" %in% out$vertices$name)
+  expect_true("regular_object" %in% out$edges$from)
+})
