### 2020-10-07 99bca27
Give pipeline objects their own environments
inevitable
