+++ R/class_glimpse.R
@@ -66,7 +66,7 @@ glimpse_class <- R6::R6Class(
       )
     },
     update_imports = function() {
-      envir <- pipeline_get_envir(self$pipeline)
+      envir <- self$pipeline$envir
       graph <- graph_envir(envir)
       edges <- lapply(as_data_frame(igraph::get.edgelist(graph)), as.character)
       edges <- data_frame(from = edges[[1]], to = edges[[2]])
@@ -99,7 +99,7 @@ glimpse_class <- R6::R6Class(
         bytes = rep(NA_real_, length(names)),
         branches = rep(NA_integer_, length(names))
       )
-      names <- c(names, names(pipeline_get_envir(self$pipeline)))
+      names <- c(names, names(self$pipeline$envir))
       edges <- pipeline_upstream_edges(self$pipeline, targets_only = FALSE)
       edges <- edges[edges$from %in% names & edges$to %in% names,, drop = FALSE] # nolint
       edges <- edges[edges$from != edges$to,, drop = FALSE] # nolint
