+++ R/class_future.R
@@ -69,7 +69,7 @@ future_class <- R6::R6Class(
       self$crew <- memory_init()
     },
     update_globals = function() {
-      self$globals <- self$produce_exports(pipeline_get_envir(self$pipeline))
+      self$globals <- self$produce_exports(self$pipeline$envir)
       self$globals$.targets_gc_5048826d <- self$garbage_collection
     },
     ensure_globals = function() {
