+++ .github/workflows/pkgdown.yaml
@@ -10,6 +10,7 @@ jobs:
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       TORCH_INSTALL: 1
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: false
     steps:
       - uses: actions/checkout@v2
 
