+++ .github/workflows/lint.yaml
@@ -8,6 +8,7 @@ jobs:
     env:
       GITHUB_PAT: ${{ secrets.GITHUBPAT }}
       TORCH_INSTALL: 1
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: false
     steps:
       - uses: actions/checkout@v2
 
