+++ .github/workflows/cover.yaml
@@ -39,7 +39,7 @@ jobs:
       - name: Install packages
         run: |
           install.packages(c("remotes"))
-          remotes::install_deps(dependencies = TRUE)
+          remotes::install_deps(dependencies = TRUE, type = ifelse(tolower(Sys.info()["sysname"]) == "linux", "source", "binary"))
           remotes::install_cran("covr")
         shell: Rscript {0}
 
