+++ .github/workflows/pkgdown.yaml
@@ -39,7 +39,7 @@ jobs:
       - name: Install dependencies
         run: |
           install.packages("remotes")
-          remotes::install_deps(dependencies = TRUE)
+          remotes::install_deps(dependencies = TRUE, type = ifelse(tolower(Sys.info()["sysname"]) == "linux", "source", "binary"))
           remotes::install_dev("pkgdown")
         shell: Rscript {0}
 
