+++ .github/workflows/lint.yaml
@@ -34,7 +34,7 @@ jobs:
       - name: Install packages
         run: |
           install.packages(c("remotes"))
-          remotes::install_deps(dependencies = TRUE)
+          remotes::install_deps(dependencies = TRUE, type = ifelse(tolower(Sys.info()["sysname"]) == "linux", "source", "binary"))
           remotes::install_cran("lintr")
         shell: Rscript {0}
 
