### 2020-10-09 f15face
[WARNING: invalidates old workflows] Sort function deps before hashing.
Helps make sure functions stay up to date even when igraph is updated.
Unfortuantely, this commit does invalidate old workflows.
