### 2020-10-09 6a68891
Sanitize errors and warnings more thoroughly
