+++ man/tar_target.Rd
@@ -154,7 +154,10 @@ the target builds on the host machine / process managing the pipeline.}
 
 \item{priority}{Numeric of length 1 between 0 and 1. Controls which
 targets get deployed first when multiple competing targets are ready
-simultaneously. Targets with priorities closer to 1 get built earlier.}
+simultaneously. Targets with priorities closer to 1 get built earlier.
+Only applies to \code{\link[=tar_make_future]{tar_make_future()}} and \code{\link[=tar_make_clustermq]{tar_make_clustermq()}}
+(not \code{\link[=tar_make]{tar_make()}}). \code{\link[=tar_make_future]{tar_make_future()}} with no extra settings is
+a drop-in replacement for \code{\link[=tar_make]{tar_make()}} in this case.}
 
 \item{resources}{A named list of computing resources. Uses:
 \itemize{
