+++ tests/testthat/test-class_future.R
@@ -5,17 +5,17 @@ tar_test("future$workers", {
 
 tar_test("workerless deployment works", {
   skip_if_not_installed("future")
-  x <- tar_target_raw("x", quote(1L), deployment = "master")
-  y <- tar_target_raw("y", quote(x), deployment = "master")
-  z <- tar_target_raw("z", quote(x + 1L), deployment = "master")
+  x <- tar_target_raw("x", quote(1L), deployment = "main")
+  y <- tar_target_raw("y", quote(x), deployment = "main")
+  z <- tar_target_raw("z", quote(x + 1L), deployment = "main")
   pipeline <- tar_pipeline(list(x, y, z))
   future_init(pipeline)$run()
   expect_equal(target_read_value(x)$object, 1L)
   expect_equal(target_read_value(y)$object, 1L)
   expect_equal(target_read_value(z)$object, 2L)
-  x <- tar_target_raw("x", quote(1L), deployment = "master")
-  y <- tar_target_raw("y", quote(x), deployment = "master")
-  z <- tar_target_raw("z", quote(x + 1L), deployment = "master")
+  x <- tar_target_raw("x", quote(1L), deployment = "main")
+  y <- tar_target_raw("y", quote(x), deployment = "main")
+  z <- tar_target_raw("z", quote(x + 1L), deployment = "main")
   pipeline <- tar_pipeline(list(x, y, z))
   out <- future_init(pipeline)
   out$run()
