+++ R/tar_option_get.R
@@ -38,8 +38,8 @@ tar_option_default <- function(option) {
     deployment = "worker",
     priority = 0,
     resources = list(),
-    storage = "master",
-    retrieval = "master",
+    storage = "main",
+    retrieval = "main",
     cue = targets::tar_cue(),
     debug = character(0)
   )
