### 2020-10-13 6c475a5
Set seed arg of future()
