### 2020-10-14 803dc5a
Add an AWS S3 test
