+++ tests/aws/test-class_aws_s3_qs.R
@@ -157,6 +157,9 @@ tar_test("aws_qs format and dynamic branching", {
       tar_target(z, sum(y))
     )
   })
+  expr <- tidy_eval(expr, environment(), TRUE)
+  eval(as.call(list(`tar_script`, expr, ask = FALSE)))
+  tar_make(callr_function = NULL)
   expect_equal(tar_read(x), seq_len(2))
   expect_equal(tar_read(y, branches = 1), 10L)
   expect_equal(tar_read(y, branches = 2), 20L)
