### 2020-10-15 24ab562
Implement tar_bind()
