+++ NEWS.md
@@ -6,10 +6,14 @@
 * Change "local" to "main" and "remote" to "worker" in the `storage`, `retrieval`, and `deployment` settings (#183, @mattwarkentin).
 * Ensure function dependencies are sorted before computing the function hash (GitHub commit f15face7d72c15c2d1098da959492bdbfcddb425).
 
-# Bug fixes
+## Bug fixes
 
 * Stop sending target return values over the network when `storage` is `"remote"` (#182, @mattwarkentin).
 * Shorten lengths of warnings and error messages to 128 characters (#186, @gorkang).
+
+## New features
+
+* Add a new function `tar_bind()` to combine pipeline objects.
 
 # targets 0.0.0.9000
 
