+++ NAMESPACE
@@ -202,6 +202,7 @@ export(rstudio_addin_tar_read)
 export(rstudio_addin_tar_target)
 export(rstudio_addin_tar_visnetwork)
 export(starts_with)
+export(tar_bind)
 export(tar_cancel)
 export(tar_cue)
 export(tar_deduplicate)
