+++ R/class_passive.R
@@ -24,6 +24,7 @@ passive_class <- R6::R6Class(
     ensure_meta = function() {
       self$meta$database$ensure_preprocessed(write = FALSE)
       self$meta$set_imports(self$pipeline$envir, self$pipeline)
+      self$meta$restrict_records(self$pipeline)
     },
     start = function() {
       pipeline_prune_names(self$pipeline, self$names)
