+++ NEWS.md
@@ -10,6 +10,7 @@
 
 * Stop sending target return values over the network when `storage` is `"remote"` (#182, @mattwarkentin).
 * Shorten lengths of warnings and error messages to 128 characters (#186, @gorkang).
+* Restrict in-memory metadata to avoid incorrectly recycling deleted targets (#191).
 
 ## New features
 
