+++ R/class_meta.R
@@ -48,7 +48,7 @@ meta_class <- R6::R6Class(
       names_records <- self$list_records()
       names_targets <- pipeline_get_names(pipeline)
       names_parents <- intersect(names_records, names_targets)
-      names_children <- map(names_parents, ~self$get_record(.x)$children)
+      names_children <- map(names_parents, ~self$database$get_row(.x)$children)
       names_children <- as.character(unlist(names_children))
       names_children <- names_children[!is.na(names_children)]
       names_current <- c(names_envir, names_targets, names_children)
