### 2020-10-17 #401 @maurolepore
\@wlandau, thanks for the high quality of this submission. I'm pleased to be the
editor.

Here I first give an overview of what I see, then ask for more information to
better understand if this or another structure is the best for this submission,
and finally comment on the editor checks. Please note the comments preceded by
a bullet point and respond to those preceded by a check-box; you may refer to their
labels.

In this submission I see three packages: `targets`, `tarchetypes`, and
`targets.manual`. I understand that (a) `targets` supersedes the popular package
`drake`, also maintained by you; (b) `tarchetyes` extends `targets`; and (c)
`targets.manual` would be published as an rOpenSci book containing the manual
for both `targets` and `tarchetypes`. Please correct me if necessary.

With three packages, this submission is more complex than the typical 1-package
submission. This may add extra pressure on resources such as rOpenSci reviewers,
and rOpenSci and CRAN servers. Please help me understand why this is the best
alternative (if possible, point me to the source of information that guided your
decision, e.g. books, articles, conversations with other developers):

- [ ] (ml01) How did you decide to supersede `drake` instead of extend it? Did
you considered any other option, such as the "edition" approach of testthat 3?

- [ ] (ml02) How did you decide that `targets` and `tarchetypes` should be two
separate packages?

- [ ] (ml03) Who would you suggest as a reviewer for this submission? Please
consider people with skills that complement yours, specially if they have
experience restructuring popular R packages.

Your answers will clarify the fit -- which I leave unchecked.

##### `targets`

README

- [ ] (ml04) Under [What about
drake?](https://github.com/wlandau/targets#what-about-drake) I see "It [is] has
become ...". You may remove "is".

- [ ] (ml05) Under [Help](https://github.com/wlandau/targets#help), where it
says "Post to the [GitHub issue tracker]", the link seems incorrect:
<https://github.com/wlandau/issues>.

* (ml06) In general, I see an opportunity to remove duplication and thus make
the package easier to maintain. For example, the file "README.Rmd" links to
multiple pages under the same website; a single link to the website might be
enough and easier to maintain. Similarly, under
[Help](https://github.com/wlandau/targets#examples) I see links to repositories
other than `targets`. This may be hard to maintain. In isolation, this may seem
like a minor issue, but minor issues can compound quickly.

CI

- [ ] (ml07) In the file "DESCRIPTION" I see `Depends: R (>= 3.5.0)`, but the
file ".github/workflows/check.yaml" covers R-release only. Please extend the
continuous-integration workflow to run R CMD check from R-devel to the R
version stated in the file "DESCRIPTION"
(`?usethis::use_github_action_check_full()`).

##### `tarchetypes`

* (ml08) Same as ml06.

- [ ] (ml09) Same as ml07.

##### `tarchetypes`

* (ml10) The name of the package and GitHub repository don't match: "targets.manual" (with ".") and "target-manual" (with "-"). Because the symbol "-" is invalid in the name of an R package, consider renaming the repository to "targets.manual". The inconsistency might confuse some users.

* (ml11) `devtools::install_github()` failed with `upgrade = "default" and passed
with `upgrade = "never"`. The error seems unrelated to the package but please check.

``` r
devtools::install_github("wlandau/targets-manual", dependencies = TRUE,
    build_vignettes = TRUE)
#> Using github PAT from envvar GITHUB_PAT
#> Downloading GitHub repo wlandau/targets-manual@HEAD
#> readr (1.3.1 -> 1.4.0) [CRAN]
#> Installing 1 packages: readr
#> Installing package into '/home/mauro/R/x86_64-pc-linux-gnu-library/4.0'
#> (as 'lib' is unspecified)
#> Error: Failed to install 'targets.manual' from GitHub:
#>   (converted from warning) installation of package 'readr' had non-zero exit status
```

- [ ] (ml12) R CMD check found 1 ERROR. Please fix the error or explain if R CMD check is not applicable to this type of package.

```r
#> > checking package dependencies ... ERROR
#>   VignetteBuilder package not declared: ‘knitr’
```

* (ml13) Tests seem not applicable to the book manual. I see no test in other
books, e.g. <https://github.com/ropensci-books/http-testing> and
<https://github.com/ropensci-books/taxize>.

- [ ] (ml14) I see a mismatch between the copy right holder in DESCRIPTION and
LICENSE (but the copy right holder in DESCRIPTION and LICENSE.md do match). Please ensure they match.

- [ ] (ml15) Please update the words list.

```r
devtools::spell_check()
#>   WORD        FOUND IN
#> ’s        index.Rmd:55
#> indivdual   index.Rmd:91
#> learnings   index.Rmd:55
#> repo        README.md:5
#> toolkits    index.Rmd:45,47,51
#> zenodo      README.md:3
```

---
