### 2020-10-17 b1a3261
Check R-3.5 and R-3.6 in CI
