+++ README.Rmd
@@ -23,7 +23,7 @@ The `targets` package is a [Make](https://www.gnu.org/software/make/)-like pipel
 
 ## What about `drake`?
 
-The [`drake`](https://github.com/ropensci/drake) package is an older and more established R-focused pipeline toolkit. It is has become a key piece of the R ecosystem, and development will continue. However, nearly four years of community feedback have exposed major user-side limitations regarding data management, collaboration, and parallel efficiency. Unfortunately, these limitations are permanent. Solutions in [`drake`](https://github.com/ropensci/drake) itself would make the package incompatible with existing projects that use it, and the internal architecture is too copious, elaborate, and mature for such extreme refactoring. That is why `targets` was created. The `targets` package borrows from past learnings, user suggestions, discussions, complaints, success stories, and feature requests, and it improves the user experience in ways that will never be possible in [`drake`](https://github.com/ropensci/drake). Please see the [statement of need](https://wlandau.github.io/targets/articles/need.html) for details.
+The [`drake`](https://github.com/ropensci/drake) package is an older and more established R-focused pipeline toolkit. It has become a key piece of the R ecosystem, and development will continue. However, nearly four years of community feedback have exposed major user-side limitations regarding data management, collaboration, and parallel efficiency. Unfortunately, these limitations are permanent. Solutions in [`drake`](https://github.com/ropensci/drake) itself would make the package incompatible with existing projects that use it, and the internal architecture is too copious, elaborate, and mature for such extreme refactoring. That is why `targets` was created. The `targets` package borrows from past learnings, user suggestions, discussions, complaints, success stories, and feature requests, and it improves the user experience in ways that will never be possible in [`drake`](https://github.com/ropensci/drake). Please see the [statement of need](https://wlandau.github.io/targets/articles/need.html) for details.
 
 ## Installation
 
