### 2020-10-17 #401 @wlandau
Thank you for your insightful feedback, @maurolepore. Below, I address ml04-ml15. (I will respond to ml01-ml03 in a separate post.)

* ml04: odd, I thought I removed that grammatical typo earlier. Fixed in https://github.com/wlandau/targets/commit/454ddf5d0c083894856369fb47f240f210b99d97.
* ml05: link fixed in https://github.com/wlandau/targets/commit/a3491b1b84a828221403d69b9c4f7c821cbff2df.
* ml06: I agree, and avoiding duplication is one of my major documentation goals for `targets`. Addressed in https://github.com/wlandau/targets/commit/a3491b1b84a828221403d69b9c4f7c821cbff2df.
* ml07: I added new CI checks under R 3.5 and 3.6 in https://github.com/wlandau/targets/commit/b1a326177c727d6b4f56bf1fe0e9b39a21dbcd89
* ml08: I deduplicated some links in https://github.com/wlandau/tarchetypes/commit/d2915e96bbc4590ff7934036f8a675ca691cbeb6. I prefer to keep the existing links to individual package functions.
* ml09: addressed in https://github.com/wlandau/tarchetypes/commit/c2e94eab18083a5021a6a2433f18a35445651116.
* ml10-ml13: I intend [targets-manual](https://github.com/wlandau/targets-manual) to just be a `bookdown` manual, not an installable package. The only purpose of the `DESCRIPTION` was to make CI easier. To reduce confusion, in https://github.com/wlandau/targets-manual/commit/a2092f974ca6099dc970575c4df7f1f196479d53 and https://github.com/wlandau/targets-manual/commit/5a69d0c33882f4d5e15a469203ee933712af1a98, I removed the `DESCRIPTION` and moved the relevant content into `README.md` and `packages.R`.
* ml14: also fixed in https://github.com/wlandau/targets-manual/commit/a2092f974ca6099dc970575c4df7f1f196479d53 and https://github.com/wlandau/targets-manual/commit/5a69d0c33882f4d5e15a469203ee933712af1a98. `DESCRIPTION`  and `LICENSE` are gone, and `README.md` is consistent with `LICENSE.md`.
* ml15: without a `DESCRIPTION`, `devtools::spell_check()` no longer works. However, I did fix a spelling issue in https://github.com/wlandau/targets-manual/commit/7ed038840f69b8f7c3d9b849f654effffd69abc5 ("indivdual" => "individual").
