### 2020-10-17 #401 @maurolepore
Thanks @wlandau for addressing my comments.

* ml10-13 and ml15:

> The only purpose of the DESCRIPTION was to make CI easier. To reduce confusion, in wlandau/targets-manual@a2092f9 and wlandau/targets-manual@5a69d0c, I removed the DESCRIPTION and moved the relevant content into README.md and packages.R.

- [ ] (ml16) I support this (more on ml18 below). Now that I understand it, I prefer your original take -- it aligns with the idea of "research compendium". Please revert to it, then explain the goal, structure, and usage of the repository in the field "Description" of the file DESCRIPTION.

* (ml17) Even if `targets.manual` is not meant to be build and installed as a package, consider adding the minimum infrastructure to the repository to get R CMD to pass cleanly -- which seems like the path of least resistance. If this is a bad idea, maybe write a broken test with the error message you want someone like me to read.

--

* (ml18) I support anything that makes your work less error prone and easier to maintain. My comments ml01-ml03 focused on external resources, but I forgot to stress that my main concern is maintainability. Your work is outstanding; the R community will benefit from you investing time in the important things instead of mundane tasks. If those mundane tasks are avoidable (e.g. with continuous integration, an "edition" approach, or rethinking the architecture of you system), then my goal is to ensure you consider those options. Collectively, the rOpenSci community has a lot of experience that should help you make informed, good choices.
