### 2020-10-17 #401 @wlandau
Below, I answer ml02 and ml03, and I address the physical separation of [`targets-manual`](https://github.com/wlandau/targets-manual) and other documentation. ml01 requires a much deeper answer, so I will address it in its own post.

# ml02

I have had more time to reflect since @melvidoni first brought this up, and I will try to elaborate.

I could have easily implemented `targets` and `tarchetypes` in the same package. Both are essentially the same system, and `tarchetypes` is currently very small, supporting only a limited handful of superficial extensions to `targets`' deeper capabilities. So for the present, the physical separation is a trivial implementation detail, and I do not believe it will exacerbate the burden of software review.

The separation of `tarchetypes` is about planning for the future. From what I learned maintaining `drake` and observing use cases, `tarchetypes` will enhance maintainability, the quality of the infrastructure and user-side freedom for years to come.

## Sustainable infrastructure

Interface development incurs additional challenges, code volume, bugs, tests, and documentation. I learned this the hard way while developing [static branching in `drake`](https://books.ropensci.org/drake/static.html). Because `drake`'s design did not allow me to implement static branching in a separate package, `drake` itself became more difficult to maintain, more prone to feature creep, and more prone to errors (see https://github.com/ropensci/drake/issues/1199, https://github.com/ropensci/drake/issues/1262, https://github.com/ropensci/drake/issues/1010, https://github.com/ropensci/drake/issues/1009, https://github.com/ropensci/drake/issues/1008, and many more issues like these). And although I fixed all the known and reported bugs, the deeper underlying causes only worsened as the years went by, and the fundamental design of the internals and the interface made it impossible to eliminate these problems in `drake` itself.

As its own package, `tarchetypes` has plenty of room to explore alternative interfaces and shorthand for targets and pipelines. That way,`targets` itself will stay light, elegant, and sharply delineated in scope, and the architecture will remain clean, resilient, and sustainable in the long run.

## A precedent for extensibility

In addition, `tarchetypes` deliberately sets a precedent for enhanced automation through third-party interface development. `tarchetypes` functions like `tar_render()` and `tar_map()` are easy to implement using `targets::tar_target_raw()`, and even now, developers are already borrowing the pattern to develop their own specialized interfaces. Internally at work, for example, one of my colleagues maintains a package for clinical trial simulation, and it has custom target archetypes tailored to the specific needs of clinical pipelines. An earlier version of this same package used `drake`, and the metaprogramming on top of `drake_plan()` was messy and nearly impossible.

# `targets-manual`

In the early days of `drake`, all the documentation lived in the README and vignettes. Over time, the vignettes steadily accrued volume, length, and R package dependencies, and `drake` became difficult to install. So I moved the vignettes to [their own `bookdown` project](https://github.com/ropensci-books/drake), and the problem was solved.

With [`targets-manual`](https://github.com/wlandau/targets-manual), I am trying to do the same thing right from the start, even though it is much shorter than [`drake`'s manual](https://books.ropensci.org/drake) at the moment. So [`targets-manual`](https://github.com/wlandau/targets-manual) is just a collection of user-side vignettes that happens to be in its own GitHub repository. As with [`tarchetypes`](https://github.com/wlandau/tarchetypes), I do not believe the physical separation of `targets` and [`targets-manual`](https://github.com/wlandau/targets-manual) increases the burden of software review relative to unifying them in a single package. Conceptually, they still belong together.

# Other documentation

`targets` has additional repositories with extra documentation. They help people learn the package, but I am not submitting them to rOpenSci.

1. <https://github.com/wlandau/targets-minimal>
2. <https://github.com/wlandau/targets-stan>
3. <https://github.com/wlandau/targets-keras>
4. <https://github.com/wlandau/targets-tutorial>

(1)-(3) are example projects that use `targets`, and (4) is a half-day short course. Again, I find the physical separation to be important. For `drake`, I made the mistake of curating multiple example projects in <https://github.com/wlandau/drake-examples>, and the repo grew without bound. With different projects in different repos now, everything is easier to maintain. The code is easier to fix, and it is easier to set up RStudio Cloud workspaces to help people learn. In addition, the pattern demonstrates to users that each `targets`-powered data analysis workflow should live in its own repo, which is exactly how I have observed most successful projects to be implemented in practice.

# ml03

My top reviewer recommendations are power users who have already provided extensive feedback and expressed strong enthusiasm about `targets`. Their prior familiarity could reduce the workload.

* @MilesMcBain
* @noamross
* @kendonB
* @mattwarkentin

Otherwise, it may help to reach out to folks who have spread `drake` throughout the community. Here are some names that come to mind.

* @krlmlr
* @thebioengineer
* @aedobbyn
* @matt-dray
* @matthewstrasiotto
* @tjmahr
* @sinarueeger
* @cstawitz
* @bpbond
* @gadenbuie
* @billdenney
* @tiernanmartin
* @b-rodrigues
* @DominikRafacz
* @pat-s
* @maelle (editor of `drake`'s software review: https://github.com/ropensci/software-review/issues/156)
* @karthik (though I am not sure if co-founders also review packages)
