### 2020-10-17 #401 @mattwarkentin
Happy to be considered as a possible reviewer of `targets` and `tarchetypes`. I use both packages on a near-daily basis for many on-going projects, including projects that are developed and built as purely local, purely remote (live and run on HPC) and a mixture of local and remote (live local and selectively run locally or on HPC via SSH).
