### 2020-10-17 #401 @wlandau
Thank you both for your eagerness to help!

@maurolepore, I addressed ml16 and ml17 in commits https://github.com/wlandau/targets-manual/commit/f792c14e7031c5c3900ad4f426a58dfc554621e0 through https://github.com/wlandau/targets-manual/commit/2ee3725943514e330d982c3e320c7e31eff3e1dc.

Do you still think it is necessary to move https://github.com/wlandau/targets-manual to https://github.com/wlandau/targets.manual? If the manual is accepted into rOpenSci, I expect the hyphen to go away on its own when the URLs to move to https://github.com/ropensci-books/targets and https://books.ropensci.org/targets.

Other than that, I believe only ml01 remains for now.
