### 2020-10-17 #401 @matthewstrasiotto
I'm deeply appreciative to be considered to review `targets`.

 I've been following it's development with interest as a daily drake user on remote systems/hpc environments with an interest in usability and discoverability in data pipelines.

I've been hesitant to migrate my workflow over. From your description, it'll be valuable to do so and I look forward to trying it out.

I'll also be making an effort to update my extension package for drake, [mandrake](https://github.com/matthewstrasiotto/mandrake) to support `targets`.
