### 2020-10-17 #401 @wlandau
Almost forgot: @mattwarkentin is actually a [formal contributor](https://github.com/wlandau/targets/blob/a3491b1b84a828221403d69b9c4f7c821cbff2df/DESCRIPTION#L25-L29) because of https://github.com/wlandau/targets/pull/170, and @noamross has a PR at https://github.com/wlandau/tarchetypes/pull/9 that I meant to keep open. @melvidoni and @maurolepore, does this mean they might have to recuse themselves?
