### 2020-10-17 d73d2ea
Compat with testthat edition 3
