+++ tests/testthat/test-class_timestamp.R
@@ -19,7 +19,9 @@ tar_test("run timestamp reporter with a error and saved workspace", {
 tar_test("run timestamp reporter with a warning", {
   pipeline <- pipeline_init(list(target_init("x", quote(warning(123)))))
   local <- local_init(pipeline, reporter = "timestamp")
-  expect_warning(local$run(), class = "condition_run")
+  suppressWarnings(
+    expect_warning(local$run(), class = "condition_run")
+  )
 })
 
 tar_test("run timestamp reporter with a cancellation", {
