+++ tests/testthat/test-class_cross.R
@@ -78,28 +78,28 @@ tar_test("cross pipeline gives correct values", {
   branches <- target_get_children(pipeline_get_target(pipeline, "map1"))
   children <- target_get_children(pipeline_get_target(pipeline, "map1"))
   out <- map_int(children, value)
-  expect_equivalent(out, seq(11, 13))
+  expect_equiv(out, seq(11, 13))
   children <- target_get_children(pipeline_get_target(pipeline, "cross1"))
   out1 <- map_int(children, value)
   exp1 <- seq(2, 4)
-  expect_equivalent(out1, exp1)
+  expect_equiv(out1, exp1)
   children <- target_get_children(pipeline_get_target(pipeline, "cross2"))
   out2 <- map_int(children, value)
   exp2 <- rowSums(expand_grid(x = seq_len(3L), y = rev(seq_len(3L))))
-  expect_equivalent(out2, exp2)
+  expect_equiv(out2, exp2)
   children <- target_get_children(pipeline_get_target(pipeline, "cross3"))
   out3 <- map_int(children, value)
   exp3 <- rowSums(expand_grid(x = seq_len(3L), y = seq(11, 13)))
-  expect_equivalent(out3, exp3)
+  expect_equiv(out3, exp3)
   expect_equal(value("out1"), sum(exp1))
   expect_equal(value("out2"), sum(exp2))
   expect_equal(value("out3"), sum(exp3))
   expect_equal(value("out4"), sum(exp1) + sum(exp2))
   expect_equal(value("out5"), sum(exp2) + sum(exp3))
-  expect_equivalent(value("out6"), c(12, 25))
+  expect_equiv(value("out6"), c(12, 25))
   children <- target_get_children(pipeline_get_target(pipeline, "map2"))
   out <- map_int(children, value)
-  expect_equivalent(out, c(144, 625))
+  expect_equiv(out, c(144, 625))
 })
 
 tar_test("empty mapping variable", {
