### 2020-10-18 #401 @wlandau
Now, I will address ml01. I think it is the most pressing question from this thread so far, and it is definitely the most difficult to explain in complete depth. So please let me know if you are skeptical or if anything remains unclear or unresolved.

# What all this means for drake

`drake` is in a position of strength. In fact, its wellness is exactly what allowed me to create `targets` in the first place. After four years of steady development, I feel that we have solved most of the solvable problems worth mentioning, which finally forced me to reckon with the *unsolvable* ones. `targets` is not about remediation, it is about breaking new ground.

# Maintenance

I will maintain `drake` indefinitely. I will continue to provide one-on-one help with use cases, fix known bugs, address known inefficiencies, and consider new community-requested features. However, since I am not going to propose new feature ideas of my own or strike out on Odyssean refactoring adventures, maintenance will be far easier and less time-consuming. In fact, because `targets` has a much cleaner design, the combined maintenance of `drake`, `targets`, `tarchetypes`, and the docs could prove far less demanding than that of `drake` alone up to this point.

# Why not an edition?

I have been hearing quite a lot about `testthat` editions lately. I read the [vignette](https://testthat.r-lib.org/articles/third-edition.html), and I just updated `targets` and `tarchetypes` so their test suites are compatible with both editions 2 and 3. So although I did not actually inspect the changes to the `testthat` code base, I think I get the idea. From the [last section of the vignette](https://testthat.r-lib.org/articles/third-edition.html#alternatives):

> You might wonder why we came up with the idea of an “edition”, rather than creating a new package like testthat3. We decided against making a new package because the 2nd and 3rd edition share a very large amount of code, so making a new package would have substantially increased the maintenance burden: the majority of bugs would’ve needed to be fixed in two places.

I think Hadley says it well. The breaking changes in `testthat` would ordinarily be disruptive to users because of the thousands of packages that rely on it. However, the impact on the `testthat` package itself is minor, and a totally new package would be overkill.

With `drake`, similar changes have occurred a small handful of times in the past. In version 6, I restructured the cache in order to improve efficiency. The change was not back-compatible, and I configured `drake` to error out at the right time, preserve existing data, and walk users through their available options. Similarly, version 7 removed all the obsolete parallel backends and threw a deprecation warning. These small but disruptive changes were manageable, and they did not require a new package.

This time, however, the change is tectonic. `drake`'s goals have transformed, and the package is about to outgrow its own architecture and design principles. More than that, my entire philosophy of programming has shifted, and I think completely differently about what a pipeline tool should do and how it should be designed. A new package seemed like the only way to achieve this, and I am glad I went that direction.

# Motivation

`drake` is the largest and most fruitful software development undertaking I have ever attempted. Just working through the technical issues more than doubled my proficiency in R, and the success of the tool granted me access to a fountain of community wisdom. To say I learned a lot would be an understatement. I am not the same developer I was four years ago, and I think totally differently about software.

For the purposes of `targets`, if I had to identify a single tipping point, it would probably be a book recommendation from Jim Hester: [Design Patterns: Elements of Reusable Object-Oriented Software](https://www.amazon.com/Design-Patterns-Elements-Reusable-Object-Oriented/dp/0201633612), the classic text by Gamma, Helm, Johnson, and Vlissides. At the time, I already had some experience with traditional message-passing OOP, but I was still confused about the problems it solved, and I was unsure how to apply it properly. The first few chapters finally answered questions about when to use object composition instead of inheritance, the importance of mutable objects for certain use cases, and how to choose a mental model that fits the task at hand.

I began to see the limitations of `drake`'s design. `drake` uses [simple immutable data structures](https://books.ropensci.org/drake/design.html) such as ad hoc lists, some of them [god objects](https://en.wikipedia.org/wiki/God_object). As a result, it has serious trouble expressing features such as [static branching](https://books.ropensci.org/drake/static.html), [dynamic branching](https://books.ropensci.org/drake/dynamic.html), and [dynamic files](https://books.ropensci.org/drake/plans.html#dynamic-files). [Seamless `targets`-style cloud integration](https://wlandau.github.io/targets-manual/cloud.html) was beyond reach, and compatibility between [dynamic files](https://books.ropensci.org/drake/plans.html#dynamic-files) and [data recovery](https://books.ropensci.org/drake/walkthrough.html#reproducible-data-recovery-and-renaming) was nearly impossible and super messy to enforce.

Finally, I realized I would no longer be able to make nontrivial improvements to `drake` as long as it relied on the old infrastructure. The path forward needed not only a complete rewrite and a complete reevaluation of every feature, but also an entirely different programming paradigm and philosophy.

# A new design

Internally, `targets` is a collection of interdependent, carefully constructed, sharply defined, lightweight R6 and S3 classes with mutable instances. I built these 74 classes from the ground up in order to express the complicated reasoning that happens both within and among targets, and the mental model easily captures the behavior of a dynamic pipeline tool. This internal harmonization paved the way for several improvements that were impossible in `drake`, some of which appear in the [statement of need](https://wlandau.github.io/targets/articles/need.html), notably cloud storage integration, interface extensibility, and parallel-efficient flexible dynamic branching.

Let's take dynamic branching as an example. `drake` struggles with this because it does not have a formal structure to express what it means to be a target. `targets`, on the other hand, supports a formal inheritance hierarchy of classes to express the unique role of each target in the dynamic branching process. For example, "stems" are unbranched targets capable of producing "buds", and "patterns" are special templates that generate the actual branches. ([This design document](https://wlandau.github.io/targets/articles/orchestration.html#target-class) overviews the different kinds of targets and their role in dynamic branching.) As a result, `targets` overcomes the map-reduce efficiency limitation of `drake`.
