### 2020-10-18 4e13f33
Repurpose memory strategy for dynamic cloud files
