+++ R/tar_target.R
@@ -117,6 +117,13 @@
 #'   after every new target completes.
 #'   Either way, the target gets automatically loaded into memory
 #'   whenever another target needs the value.
+#'   For cloud-based dynamic files such as `format = "aws_file"`,
+#'   this memory policy applies to
+#'   temporary local copies of the file in `_targets/scratch/"`:
+#'   `"persistent"` means they remain until the end of the pipeline,
+#'   and `"transient"` means they get deleted from the file system
+#'   as soon as possible. The former conserves bandwidth,
+#'   and the latter conserves local storage.
 #' @param deployment Character of length 1, only relevant to
 #'   [tar_make_clustermq()] and [tar_make_future()]. If `"worker"`,
 #'   the target builds on a parallel worker. If `"main"`,
