+++ NAMESPACE
@@ -84,6 +84,8 @@ S3method(store_serialize_value,tar_torch)
 S3method(store_sync_file_meta,default)
 S3method(store_sync_file_meta,tar_aws_s3)
 S3method(store_sync_file_meta,tar_url)
+S3method(store_unload,default)
+S3method(store_unload,tar_aws_s3_file)
 S3method(store_unserialize_value,default)
 S3method(store_unserialize_value,tar_keras)
 S3method(store_unserialize_value,tar_torch)
