### 2020-10-18 a6e72b2
Implement garbage_collection target-level setting
