+++ R/class_prelocal.R
@@ -4,7 +4,6 @@ prelocal_new <- function(
   names = NULL,
   queue = NULL,
   reporter = NULL,
-  garbage_collection = NULL,
   scheduler = NULL
 ) {
   prelocal_class$new(
@@ -13,7 +12,6 @@ prelocal_new <- function(
     names = names,
     queue = queue,
     reporter = reporter,
-    garbage_collection = garbage_collection,
     scheduler = scheduler
   )
 }
@@ -31,7 +29,6 @@ prelocal_class <- R6::R6Class(
       names = NULL,
       queue = NULL,
       reporter = NULL,
-      garbage_collection = NULL,
       scheduler = NULL
     ) {
       super$initialize(
@@ -39,8 +36,7 @@ prelocal_class <- R6::R6Class(
         meta = meta,
         names = names,
         queue = queue,
-        reporter = reporter,
-        garbage_collection = garbage_collection
+        reporter = reporter
       )
       self$scheduler <- scheduler
     },
