+++ R/class_clustermq.R
@@ -4,7 +4,6 @@ clustermq_init <- function(
   names = NULL,
   queue = "parallel",
   reporter = "verbose",
-  garbage_collection = FALSE,
   workers = 1L,
   log_worker = FALSE
 ) {
@@ -14,7 +13,6 @@ clustermq_init <- function(
     names = names,
     queue = queue,
     reporter = reporter,
-    garbage_collection = garbage_collection,
     workers = as.integer(workers),
     log_worker = log_worker
   )
@@ -26,7 +24,6 @@ clustermq_new <- function(
   names = NULL,
   queue = NULL,
   reporter = NULL,
-  garbage_collection = NULL,
   workers = NULL,
   crew = NULL,
   log_worker = NULL
@@ -37,7 +34,6 @@ clustermq_new <- function(
     names = names,
     queue = queue,
     reporter = reporter,
-    garbage_collection = garbage_collection,
     workers = workers,
     crew = crew,
     log_worker = log_worker
@@ -60,7 +56,6 @@ clustermq_class <- R6::R6Class(
       names = NULL,
       queue = NULL,
       reporter = NULL,
-      garbage_collection = NULL,
       workers = NULL,
       crew = NULL,
       log_worker = NULL
@@ -70,8 +65,7 @@ clustermq_class <- R6::R6Class(
         meta = meta,
         names = names,
         queue = queue,
-        reporter = reporter,
-        garbage_collection = garbage_collection
+        reporter = reporter
       )
       self$workers <- workers
       self$crew <- crew
@@ -183,7 +177,6 @@ clustermq_class <- R6::R6Class(
         names = self$names,
         queue = self$queue,
         reporter = self$reporter,
-        garbage_collection = self$garbage_collection,
         scheduler = self$scheduler
       )
     },
@@ -210,7 +203,6 @@ clustermq_class <- R6::R6Class(
     },
     validate = function() {
       super$validate()
-      assert_lgl(self$garbage_collection)
       assert_int(self$workers)
     }
   )
