### 2020-10-18 8004553
Exempt 2 lines and explain why
