+++ tests/hpc/test-clustermq.R
@@ -3,8 +3,8 @@ tar_test("clustermq iteration loop can wait and shut down workers", {
   skip_if_not_installed("clustermq")
   old <- getOption("clustermq.scheduler")
   options(clustermq.scheduler = "multicore")
-  x <- tar_target_raw("x", quote(Sys.sleep(2)))
-  y <- tar_target_raw("y", quote(list(x, a = "x")))
+  x <- tar_target_raw("x", quote(Sys.sleep(2)), garbage_collection = TRUE)
+  y <- tar_target_raw("y", quote(list(x, a = "x")), garbage_collection = TRUE)
   pipeline <- tar_pipeline(list(x, y))
   out <- clustermq_init(pipeline, reporter = "silent")
   out$run()
