### 2020-10-18 2693676
Give garbage_collection to tar_target_raw()
