### 2020-10-18 #401 @wlandau
An aside: in this situation, you can read a lot into the package names. "drake" is an acronym: "data frames in R for Make". In the first few versions, `drake::make(parallelism = "Makefile")` would actually take a data frame (a `drake` plan) and turn it into a real Makefile. `drake` has long since outgrown this approach, and the "Makefile" backend is no longer supported. In other words, `drake` outgrew its own name.

The name "targets" represents a different way of thinking. Whereas `drake`'s data structures operate at the level of the entire pipeline, nearly all the structure and all the reasoning of `targets` happens on an individual target-by-target basis. In `targets`, target objects are first-class citizens, and they express almost all of the behavior through method dispatch and through the nested OOP objects that compose them. This reductionism really fits the situation, and it ensures the mental model aligns with the realities of implementation.
