### 2020-10-19 #401 @bpbond
>Odyssean refactoring adventures

🚣

That's a thoughtful and interesting [comment](https://github.com/ropensci/software-review/issues/401#issuecomment-711103473) @wlandau . Happy to help review if useful.
