+++ R/tar_target.R
@@ -140,7 +140,7 @@
 #'   * Template file wildcards for `future::future()` in [tar_make_future()].
 #'   * Template file wildcards `clustermq::workers()` in [tar_make_clustermq()].
 #'   * Custom target-level `future::plan()`, e.g.
-#'     `resources = list(plan = future::plan(future::multicore))`.
+#'     `resources = list(plan = future.callr::callr)`.
 #'   * Custom `curl` handle if `format = "url"`,
 #'     e.g. `resources = list(handle = curl::new_handle())`.
 #'   * Custom preset for `qs::qsave()` if `format = "qs"`, e.g.
