### 2020-10-19 f8f8e44
Add hpc tests for target-specific future plans
