### 2020-10-20 4c839f0
Only use .cleanup if the user sets the future plan in resources
