+++ R/class_future.R
@@ -74,10 +74,18 @@ future_class <- R6::R6Class(
       self$ensure_globals()
       globals <- self$globals
       globals$.targets_target_5048826d <- target
-      plan_old <- future::plan()
-      on.exit(future::plan(plan_old, .cleanup = FALSE))
-      plan_new <- target$settings$resources$plan %||% future::plan()
-      future::plan(plan_new, .cleanup = FALSE)
+      plan_new <- target$settings$resources$plan
+      if (!is.null(plan_new)) {
+        # Temporary solution to allow heterogeneous workers
+        # from different plans. Uses .cleanup = FALSE # nolint
+        # to avoid destroying psock clusters.
+        # .cleanup may not be supported long-term # nolint
+        # so we only attempt to change plans if the user
+        # sets the plan in the resources argument/option.
+        plan_old <- future::plan()
+        on.exit(future::plan(plan_old, .cleanup = FALSE))
+        future::plan(plan_new, .cleanup = FALSE)
+      }
       future <- future::future(
         expr = target_run_worker(.targets_target_5048826d),
         packages = "targets",
@@ -157,7 +165,11 @@ future_class <- R6::R6Class(
       self$end()
     },
     end = function() {
-      future::plan(future::sequential, .cleanup = TRUE)
+      # Cleans up psock clusters.
+      # Implicitly assumes .cleanup = TRUE is the default. # nolint
+      # Does not set .cleanup directly # nolint
+      # because this argument may not be supported long-term.
+      future::plan(future::sequential)
       super$end()
     },
     validate = function() {
