+++ tests/testthat/test-class_future.R
@@ -33,13 +33,27 @@ tar_test("some targets up to date, some not", {
   x <- tar_target_raw("x", quote(1L))
   y <- tar_target_raw("y", quote(x + 1L))
   pipeline <- tar_pipeline(list(x, y))
-  cmq <- future_init(pipeline)
-  cmq$run()
-  out <- names(cmq$scheduler$progress$built$envir)
+  future <- future_init(pipeline)
+  future$run()
+  out <- names(future$scheduler$progress$built$envir)
   expect_equal(out, "y")
   value <- target_read_value(pipeline_get_target(pipeline, "y"))
   expect_equal(value$object, 2L)
   future::plan(future::sequential)
+})
+
+tar_test("specialized plans", {
+  skip_if_not_installed("future")
+  resources <- list(plan = future::sequential)
+  x <- tar_target_raw("x", quote(1L), resources = resources)
+  y <- tar_target_raw("y", quote(x))
+  pipeline <- tar_pipeline(list(x, y))
+  future <- future_init(pipeline)
+  future$run()
+  out <- sort(names(future$scheduler$progress$built$envir))
+  expect_equal(out, sort(c("x", "y")))
+  value <- target_read_value(pipeline_get_target(pipeline, "y"))
+  expect_equal(value$object, 1L)
 })
 
 tar_test("future algo can skip targets", {
@@ -54,9 +68,9 @@ tar_test("future algo can skip targets", {
   x <- tar_target_raw("x", quote(1L))
   y <- tar_target_raw("y", quote(x))
   pipeline <- tar_pipeline(list(x, y))
-  cmq <- future_init(pipeline)
-  cmq$run()
-  out <- names(cmq$scheduler$progress$built$envir)
+  future <- future_init(pipeline)
+  future$run()
+  out <- names(future$scheduler$progress$built$envir)
   expect_equal(out, "x")
   value <- target_read_value(pipeline_get_target(pipeline, "x"))
   expect_equal(value$object, 1L)
@@ -79,8 +93,8 @@ tar_test("nontrivial globals", {
   }, envir = envir)
   x <- tar_target_raw("x", quote(f(1L)))
   pipeline <- tar_pipeline(list(x))
-  cmq <- future_init(pipeline)
-  cmq$run()
+  future <- future_init(pipeline)
+  future$run()
   expect_equal(tar_read(x), 3L)
   future::plan(future::sequential)
   tar_option_set(envir = old_envir)
