### 2020-10-21 62dc071
6% speedup: clear caches faster
