+++ tests/testthat/test-class_cache.R
@@ -77,4 +77,8 @@ tar_test("cache_clear_objects()", {
     out$targets$names,
     character(0)
   )
+  expect_identical(
+    parent.env(out$targets$envir),
+    out$imports$envir
+  )
 })
