### 2020-10-22 48c0461
Error if none of the requested targets are found
Fix #203
