+++ tests/testthat/test-tar_make.R
@@ -72,3 +72,11 @@ tar_test("tar_make() finds the correct environment", {
   out <- tar_read(y)
   expect_equal(out, 3L)
 })
+
+tar_test("cannot find any of the requested targets", {
+  tar_script(tar_pipeline(tar_target(x, 1)))
+  expect_error(
+    tar_make(callr_function = NULL, names = "y"),
+    calss = "condition_validate"
+  )
+})
