### 2020-10-27 #401 @maurolepore
RE  https://github.com/ropensci/software-review/issues/401#issuecomment-711086276
Thanks for the heads up. If a potential conflict of interest is unclearly defined in the [guidelines](https://devguide.ropensci.org/policies.html#coi), I'll discuss with other editors. For now you should not worry.
