+++ paper.bib
@@ -1,0 +1,89 @@
+@misc{
+  targets,
+  author = {William Michael Landau},
+  title = {{targets: dynamic function-oriented {M}ake-like declarative pipelines for {R}}},
+  year = {2020},
+  howpublished = {https://github.com/wlandau/targets},
+  DOI = {}
+}
+
+@article{
+  drake,
+  doi = {10.21105/joss.00550},
+  url = {https://doi.org/10.21105/joss.00550},
+  year = {2018},
+  publisher = {The Open Journal},
+  volume = {3},
+  number = {21},
+  pages = {550},
+  author = {William Michael Landau},
+  title = {The drake R package: a pipeline toolkit for reproducibility and high-performance computing},
+  journal = {Journal of Open Source Software}
+}
+
+@misc{
+  knitr,
+  author = {Yihui Xie},
+  year = 2017,
+  title = {{knitr: A General-Purpose Package for Dynamic Report Generation in R}},
+  note = {R package version 1.17},
+  howpublished = {https://CRAN.R-project.org/package=knitr}
+}
+
+@book{
+  Make,
+  title = {{GNU Make, Version 3.77}},
+  publisher = {{Free Software Foundation}},
+  author = {Richard Stallman},
+  isbn = {1882114809},
+  year = {1998}
+}
+
+@misc{
+  memoise,
+  title = {{memoise: Memoisation of Functions}},
+  author = {Hadley Wickham and Jim Hester and Kirill Müller and Daniel Cook},
+  year = {2017},
+  note = {R package version 1.1.0},
+  howpublished = {https://CRAN.R-project.org/package=memoise}
+}
+
+@misc{
+  pipelines,
+  title = {{Awesome Pipeline}},
+  author = {Paolo Di Tommaso},
+  year = {2020},
+  howpublished = {\url{https://github.com/pditommaso/awesome-pipeline}},
+  commit = {7cbf1316b0fbca0791db8d6bed9acc35eecafc8d}
+}
+
+@misc{
+  remake,
+  title = {{remake: Make-like build management}},
+  author = {Rich FitzJohn},
+  note = {R package version 0.3.0},
+  year = {2017},
+  howpublished = {\url{https://github.com/richfitz/remake}},
+  commit = {e29028b548950a3132ea2d045b7f67344ce22a6b}
+}
+
+@Article{
+    tidyverse,
+    title = {Welcome to the {tidyverse}},
+    author = {Hadley Wickham and Mara Averick and Jennifer Bryan and Winston Chang and Lucy D'Agostino McGowan and Romain François and Garrett Grolemund and Alex Hayes and Lionel Henry and Jim Hester and Max Kuhn and Thomas Lin Pedersen and Evan Miller and Stephan Milton Bache and Kirill Müller and Jeroen Ooms and David Robinson and Dana Paige Seidel and Vitalie Spinu and Kohske Takahashi and Davis Vaughan and Claus Wilke and Kara Woo and Hiroaki Yutani},
+    year = {2019},
+    journal = {Journal of Open Source Software},
+    volume = {4},
+    number = {43},
+    pages = {1686},
+    doi = {10.21105/joss.01686},
+  }
+
+@Manual{
+  metaflow,
+  title = {metaflow: Metaflow for R-Lang},
+  author = {Jason Ge and Savin Goyal},
+  note = {https://metaflow.org/,
+  https://docs.metaflow.org/,
+  https://github.com/Netflix/metaflow},
+}
