### 2020-10-27 #401 @wlandau
> Thanks @wlandau for your patience. I can now confirm this submission satisfies editor checks. I'll start seeking reviewers (thanks for your suggestions).

So glad to hear it!

> Please ensure the README files of each package in this submission has an rOpenSci review badge -- maybe via --rodev::use_review_badge(), rodev::use_review_badge(<issue_number>).

Now added.

Also, just an FYI: my employer just approved my JOSS manuscript for disclosure, so I added it to the `targets` repo (`paper.md` and `paper.bib`). I will update the original post.
