### 2020-10-27 #401 @wlandau
Thanks, @MilesMcBain. Your feedback has helped a lot already, particularly with [interactive debugging support](https://wlandau.github.io/targets-manual/practice.html#interactive-debugging), and I look forward to more.

You also suggested this may be a good time to coordinate with RStudio, and which immediately brought more potential reviewers to mind:

* @jennybc
* @rich-iannone
* @gaborcsardi
* @slopp

And I originally meant to include @HenrikBengtsson and @mschubert.
