### 2020-11-03 29a3a94
Add tar_seed(). Make tar_name()/tar_path() debug-friendly.
Also opt into testthat 3e
