+++ R/tar_seed.R
@@ -1,0 +1,23 @@
+#' @title Get the random number generator seed of the target currently running.
+#' @export
+#' @description Get the random number generator seed
+#'   of the target currently running.
+#' @return Integer of length 1. If invoked inside a `targets` pipeline,
+#'   the return value is the seed of the target currently running,
+#'   which is a deterministic function of the target name. Otherwise,
+#'   the return value is `default`.
+#' @param default Integer, value to return if `tar_seed()`
+#'   is called on its own outside a `targets` pipeline.
+#'   Having a default lets users run things without [tar_make()],
+#'   which helps peel back layers of code and troubleshoot bugs.
+#' @examples
+#' tar_seed()
+#' tar_seed(default = 123L)
+#' tar_target(returns_seed, tar_seed(default = 123)) # Does not return 123.
+tar_seed <- function(default = 1L) {
+  trn(
+    exists(x = "seed", envir = envir_run, inherits = FALSE),
+    get(x = "seed", envir = envir_run),
+    as.integer(default)
+  )
+}
