+++ tests/testthat/test-class_torch.R
@@ -23,6 +23,22 @@ tar_test("torch format with in-memory serialization", {
   expect_equal(torch::as_array(out), c(1, 2))
 })
 
+tar_test("torch in-memory serialization of deps", {
+  skip_on_cran()
+  skip_on_os("windows")
+  skip_if_not_installed("clustermq")
+  skip_if_not_installed("torch")
+  tar_script({
+    tar_option_set(packages = "torch", retrieval = "main")
+    options(clustermq.scheduler = "multicore")
+    tar_pipeline(
+      tar_target(tensor, torch_zeros(10), format = "torch"),
+      tar_target(array, as.array(tensor))
+    )
+  })
+  tar_make_clustermq()
+})
+
 tar_test("validate torch format", {
   skip_if_not_installed("torch")
   x <- target_init(name = "a", expr = quote(f()), format = "torch")
