### 2020-11-04 60c11fd
Serialize at the object level
