+++ NAMESPACE
@@ -79,16 +79,15 @@ S3method(store_read_path,tar_rds)
 S3method(store_read_path,tar_store_file)
 S3method(store_read_path,tar_torch)
 S3method(store_read_path,tar_url)
-S3method(store_serialize_value,tar_keras)
-S3method(store_serialize_value,tar_torch)
+S3method(store_serialize_object,tar_keras)
+S3method(store_serialize_object,tar_torch)
 S3method(store_sync_file_meta,default)
 S3method(store_sync_file_meta,tar_aws_s3)
 S3method(store_sync_file_meta,tar_url)
 S3method(store_unload,default)
 S3method(store_unload,tar_aws_s3_file)
-S3method(store_unserialize_value,default)
-S3method(store_unserialize_value,tar_keras)
-S3method(store_unserialize_value,tar_torch)
+S3method(store_unserialize_object,tar_keras)
+S3method(store_unserialize_object,tar_torch)
 S3method(store_upload_object,tar_aws_s3)
 S3method(store_validate_packages,default)
 S3method(store_validate_packages,tar_aws_s3)
@@ -204,6 +203,7 @@ export(rstudio_addin_tar_read)
 export(rstudio_addin_tar_target)
 export(rstudio_addin_tar_visnetwork)
 export(starts_with)
+export(store_unserialize_value.default)
 export(tar_bind)
 export(tar_cancel)
 export(tar_cue)
