### 2020-11-04 afaf0fc
Shallow copy targets to make subpipeline
