### 2020-11-04 d46afa1
Remove stashing and restoring
