### 2020-11-04 350ea84
Create a subpipeline when deployment is "worker"
