### 2020-11-04 33c1b25
Use require() with custom safeguards
