+++ R/class_command.R
@@ -32,8 +32,10 @@ command_new <- function(
   environment()
 }
 
+# require() is faster than library() # nolint
+# but we should still fail early and loudly when needed.
 command_load_packages <- function(command) {
-  suppressPackageStartupMessages(
+  out <- suppressPackageStartupMessages(
     lapply(
       command$packages,
       require,
@@ -42,6 +44,9 @@ command_load_packages <- function(command) {
       character.only = TRUE
     )
   )
+  out <- unlist(out)
+  packages <- paste(command$packages[!out], collapse = ", ")
+  assert_true(all(out), paste("packages not found:", packages))
 }
 
 command_produce_build <- function(command, envir) {
