### 2020-11-04 d2f82ed
Set deployment to main in tar_make() for all targets
