+++ R/tar_make.R
@@ -60,7 +60,9 @@ tar_make <- function(
 
 tar_make_inner <- function(pipeline, names_quosure, reporter) {
   pipeline_validate_lite(pipeline)
-  names <- eval_tidyselect(names_quosure, pipeline_get_names(pipeline))
+  names <- pipeline_get_names(pipeline)
+  map(names, ~target_make_deploy_main(pipeline_get_target(pipeline, .x)))
+  names <- eval_tidyselect(names_quosure, names)
   local_init(
     pipeline = pipeline,
     names = names,
@@ -68,6 +70,10 @@ tar_make_inner <- function(pipeline, names_quosure, reporter) {
     reporter = reporter
   )$run()
   invisible()
+}
+
+target_make_deploy_main <- function(target) {
+  target$settings$deployment <- "main"
 }
 
 tar_make_reporters <- function() {
