### 2020-11-04 ae463c0
Reinitialize the value instance of each subpipeline target if retrieval is "worker"
