### 2020-11-04 231db1c
Sketch (un)serialization of targets in the subpipeline
