+++ R/class_target.R
@@ -345,6 +345,14 @@ target_sync_file_meta <- function(target, meta) {
 target_sync_file_meta.default <- function(target, meta) {
 }
 
+target_serialize_value <- function(target) {
+  store_serialize_value(target$store, target)
+}
+
+target_unserialize_value <- function(target) {
+  store_unserialize_value(target$store, target)
+}
+
 target_validate <- function(target) {
   UseMethod("target_validate")
 }
