### 2020-11-04 40c280d
Do not set deployment to main in tar_make()
