### 2020-11-04 8043448
Revert tar_make_inner()
