### 2020-11-04 7c10391
Rm superfluous code
