### 2020-11-04 472af56
Update where cache is cleared
