+++ NEWS.md
@@ -13,6 +13,7 @@
 * Stop sending target return values over the network when `storage` is `"remote"` (#182, @mattwarkentin).
 * Shorten lengths of warnings and error messages to 128 characters (#186, @gorkang).
 * Restrict in-memory metadata to avoid incorrectly recycling deleted targets (#191).
+* Serialize unexportable dependencies before sending them to workers. Transport data through `target$subpipeline` rather than `target$cache` to make that happen (#209, @mattwarkentin).
 
 ## New features
 
