### 2020-11-04 a287458
Cover lines and fix lint
