+++ README.Rmd
@@ -53,6 +53,10 @@ install_github("wlandau/targets")
 * [Machine learning with Keras](https://github.com/wlandau/targets-keras).
 * [Validating a Stan model](https://github.com/wlandau/targets-stan).
 
+## Recorded talks
+
+* [R/Pharma 2020 (9:24)](https://www.youtube.com/watch?v=GRqKJBaC5g4&list=PLMtxz1fUYA5C0YflXsR8EEAQXfjntlV1H&index=6)
+
 ## Courses
 
 * [Official half-day interactive tutorial](https://github.com/wlandau/targets-tutorial).
