### 2020-11-05 60dbbad
Evaluate commands directly in `targets$cache$targets$envir`
Improves convenience in interactive debugging (`ls()` just works now.) This is reasonably safe now that the cache is populated at the last minute and cleared as soon as possible (#209, #210).
