### 2020-11-05 a920b2f
Add LA R user group talk
