+++ README.Rmd
@@ -70,3 +70,9 @@ install_github("wlandau/targets")
 ## Participation
 
 Development is a community effort, and we welcome discussion and contribution. By participating in this project, you agree to abide by the [code of conduct](https://github.com/wlandau/targets/blob/main/CODE_OF_CONDUCT.md) and the [contributing guide](https://github.com/wlandau/targets/blob/main/CONTRIBUTING.md).
+
+## Citation
+
+```{r}
+citation("targets")
+```
