### 2020-11-10 23053c5
Sketch first steps
