### 2020-11-11 4f72544
About to implement pattern_produce_grid
