+++ NAMESPACE
@@ -9,6 +9,7 @@ S3method(pipeline_validate_lite,default)
 S3method(pipeline_validate_lite,tar_pipeline)
 S3method(print,tar_cross)
 S3method(print,tar_map)
+S3method(print,tar_pattern)
 S3method(print,tar_pipeline)
 S3method(print,tar_stem)
 S3method(store_assert_format,default)
@@ -134,6 +135,7 @@ S3method(target_get_parent,tar_stem)
 S3method(target_get_type,tar_branch)
 S3method(target_get_type,tar_cross)
 S3method(target_get_type,tar_map)
+S3method(target_get_type,tar_pattern)
 S3method(target_get_type,tar_stem)
 S3method(target_get_type_cli,default)
 S3method(target_get_type_cli,tar_stem)
@@ -149,7 +151,7 @@ S3method(target_patternview_meta,tar_branch)
 S3method(target_prepare,default)
 S3method(target_prepare,tar_builder)
 S3method(target_produce_junction,tar_cross)
-S3method(target_produce_junction,tar_map)
+S3method(target_produce_junction,tar_pattern)
 S3method(target_produce_junction,tar_stem)
 S3method(target_produce_record,tar_branch)
 S3method(target_produce_record,tar_pattern)

[…]
