+++ NAMESPACE
@@ -7,8 +7,6 @@ S3method(pipeline_validate,default)
 S3method(pipeline_validate,tar_pipeline)
 S3method(pipeline_validate_lite,default)
 S3method(pipeline_validate_lite,tar_pipeline)
-S3method(print,tar_cross)
-S3method(print,tar_map)
 S3method(print,tar_pattern)
 S3method(print,tar_pipeline)
 S3method(print,tar_stem)
@@ -133,8 +131,6 @@ S3method(target_get_parent,tar_bud)
 S3method(target_get_parent,tar_pattern)
 S3method(target_get_parent,tar_stem)
 S3method(target_get_type,tar_branch)
-S3method(target_get_type,tar_cross)
-S3method(target_get_type,tar_map)
 S3method(target_get_type,tar_pattern)
 S3method(target_get_type,tar_stem)
 S3method(target_get_type_cli,default)
@@ -150,7 +146,6 @@ S3method(target_patternview_meta,default)
 S3method(target_patternview_meta,tar_branch)
 S3method(target_prepare,default)
 S3method(target_prepare,tar_builder)
-S3method(target_produce_junction,tar_cross)
 S3method(target_produce_junction,tar_pattern)
 S3method(target_produce_junction,tar_stem)
 S3method(target_produce_record,tar_branch)
