### 2020-11-11 5fa5b1f
Validate patterns
