### 2020-11-11 fa16064
Update some tests
