+++ tests/testthat/test-class_pattern.R
@@ -1,18 +1,20 @@
 tar_test("print pipeline", {
-  expect_true(is.character(utils::capture.output(print(pipeline_init()))))
+  x <- tar_target(x, 1)
+  y <- pipeline_init(list(x))
+  expect_true(is.character(utils::capture.output(print(y))))
 })
 
-tar_test("map pattern initializes correctly", {
+tar_test("pattern initializes correctly", {
   x <- target_init("x", expr = quote(1 + 1), pattern = quote(map(a, b)))
-  expect_true(inherits(x, "tar_map"))
+  expect_true(inherits(x, "tar_pattern"))
 })
 
-tar_test("target_get_parent(map)", {
+tar_test("target_get_parent(pattern)", {
   x <- target_init("x", expr = quote(1 + 1), pattern = quote(map(a, b)))
   expect_equal(target_get_parent(x), "x")
 })
 
-tar_test("map$patternview", {
+tar_test("pattern$patternview", {
   x <- target_init("x", expr = quote(1 + 1), pattern = quote(map(a, b)))
   expect_silent(patternview_validate(x$patternview))
 })
@@ -350,17 +352,6 @@ tar_test("branches with different names use different seeds", {
   expect_false(out[1] == out[2])
 })
 
-tar_test("map validate", {
-  x <- target_init("x", expr = quote(1 + 1), pattern = quote(map(a, b)))
-  expect_silent(target_validate(x))
-})
-
-tar_test("map validate with bad junction", {
-  x <- target_init("x", expr = quote(1 + 1), pattern = quote(map(a, b)))
-  x$junction <- junction_new()
-  expect_error(target_validate(x), class = "condition_validate")
-})
-
 tar_test("map over a stem that was not mapped over last time", {
   pipeline <- pipeline_init(
     list(
@@ -422,7 +413,7 @@ tar_test("map over empty stem", {
   expect_error(local$run(), class = "condition_pattern")
 })
 
-tar_test("map$produce_record() of a successful map", {
+tar_test("pattern$produce_record() of a successful map", {
   stem <- target_init("x", quote(sample.int(4)))
   target <- target_init("y", quote(x), pattern = quote(map(x)))
   pipeline <- pipeline_init(list(stem, target))
@@ -433,7 +424,7 @@ tar_test("map$produce_record() of a successful map", {
   expect_silent(record_validate(record))
   expect_equal(record$name, "y")
   expect_equal(record$parent, NA_character_)
-  expect_equal(record$type, "map")
+  expect_equal(record$type, "pattern")
   expect_equal(nchar(record$command), 16L)
   expect_equal(record$depend, NA_character_)
   expect_equal(record$path, NA_character_)
@@ -473,14 +464,14 @@ tar_test("inconformable mapping variables", {
   )
   expect_error(
     local_init(pipeline)$run(),
-    class = "condition_pattern"
+    class = "condition_validate"
   )
 })
 
-tar_test("map print", {
-  x <- tar_target(x, 1, map(x, y, z))
+tar_test("pattern print", {
+  x <- tar_target(w, 1, map(x, y, z))
   out <- utils::capture.output(print(x))
-  expect_true(any(grepl("map", out)))
+  expect_true(any(grepl("pattern", out)))
 })
 
 tar_test("must branch over stems and patterns", {
@@ -546,4 +537,183 @@ tar_test("prohibit branching over stem files", {
   )
   algo <- local_init(pipeline)
   expect_error(algo$run(), class = "condition_validate")
+})
+
[…]
