### 2020-11-11 dad729a
Update some tests
