### 2020-11-11 4b6028d
Safeguard patterns
