### 2020-11-11 ff66a57
Bump version and update news
