### 2020-11-11 53813da
Remove obsolete test
