+++ tests/testthat/test-class_target.R
@@ -116,14 +116,6 @@ tar_test("error to validate a circular target", {
   )
 })
 
-tar_test("error to validate a circular target", {
-  x <- target_init(name = "abc", expr = quote(abc))
-  expect_error(
-    target_validate(x),
-    class = "condition_validate"
-  )
-})
-
 tar_test("targets resist self-referantiality", {
   x <- target_init(name = "identity", expr = quote(identity("i")))
   pipeline <- pipeline_init(list(x))
