### 2020-11-13 01ae6f6
Sketch new workspace infrastructure
