### 2020-11-13 0818b02
Make workspaces lighter
