+++ R/tar_workspace.R
@@ -5,46 +5,26 @@
 #'   Remove workspace files with [tar_undebug()] when you are done debugging.
 #' @details If you set `error = "save"` in [tar_option_set()]
 #'   or [tar_target()], then if that target throws an error
-#'   in [tar_make()], it will save its workspace to a compressed file
-#'   in `_targets/workspaces/`. The workspace includes the
-#'   target's required packages, the
-#'   global objects at the time [tar_make()] was called, the
-#'   dependency targets, the random number generator seed
-#'   (assigned to `.targets$seed` and set with `set.seed()`)
-#'   and the traceback of the error (assigned to `.targets$traceback`).
-#'   `tar_workspace()` loads the packages, populates the environment
-#'   with the objects, and sets the seed to the seed of the target.
-#'   Workspace files can be large sometimes, so it is good practice to
-#'   remove them with [tar_undebug()] when you are done debugging.
-#'
-#'   Although useful, this behavior does not perfectly replicate
-#'   what [tar_make()] does to set up the runtime environment
-#'   for a target. [tar_make()] creates a formal hierarchy of
-#'   environments that inherit from one another in order to
-#'   organize data and contain side effects.
-#'
-#'   In addition, sometimes, you may not be able to debug with
-#'   `error = "save"`. Workspace files are saved with `qs::qsave()`,
-#'   so if one of your dependencies cannot be properly serialized this way
-#'   (e.g. Keras models you store with `format = "keras"`)
-#'   then the workspace cannot be saved properly. If this happens to you,
-#'   either avoid non-exportable objects or use interactive debugging.
-#'   (See the `debug` argument of [tar_option_set()].)
+#'   in [tar_make()], it will save its workspace to an RDS file
+#'   in `_targets/workspaces/`. The workspace is a compact reference
+#'   that allows `tar_workspace()` to load the target's dependencies
+#'   and random number generator seed as long as the data objects
+#'   are still in the data store (usually files in `_targets/objects/`).
 #' @return the function returns no value, but it does load
 #'   the target's required packages, as well as multiple objects
 #'   into the environment (`envir` argument) in order to replicate the
 #'   workspace where the error happened. These objects include
-#'   the global objects at the time [tar_make()] was called, the
-#'   dependency targets, the random number generator seed
-#'   (assigned to `.targets$seed` and set with `set.seed()`)
-#'   and the traceback of the error (assigned to `.targets$traceback`).
+#'   the global objects at the time [tar_make()] was called and the
+#'   dependency targets. The random number generator seed for the
+#'   target is also assigned with `set.seed()`
 #' @param name Symbol, name of the target whose workspace to read.
 #' @param envir Environment in which to put the objects.
 #' @param packages Logical, whether to load the required packages
-#'   of the target. Uses `require()` with `lib.loc` equal to
-#'   the `library` setting of the target
-#'   (from `tar_target(library = "PATH_TO_LIB")` or
-#'   `tar_option_set(library = "PATH_TO_LIB")`).
+#'   of the target.
+#' @param source Logical, whether to run `_targets.R` to load user-defined
+#'   global object dependencies into `envir`. If `TRUE`, then `envir`
+#'   should either be the global environment or inherit from the
+#'   global environment.
 #' @examples
 #' \dontrun{
 #' tmp <- sample(1)
@@ -62,24 +42,24 @@
 #' exists("x") # Should be TRUE.
 #' print(x) # "loaded"
 #' tail(.Random.seed) # Should be different.
-#' tail(.targets$traceback, 1)
 #' }
[…]

[…]
