### 2020-11-13 bb3fe3c
Add a "workspace" option to tar_option_set()
