+++ man/tar_workspace.Rd
@@ -30,11 +30,11 @@ target is also assigned with \code{set.seed()}
 }
 \description{
 Load the packages, workspace, and random number generator seed
-of an errored target attempted with \code{error = "save"}.
+of an errored target attempted with \code{error = "workspace"}.
 Remove workspace files with \code{\link[=tar_undebug]{tar_undebug()}} when you are done debugging.
 }
 \details{
-If you set \code{error = "save"} in \code{\link[=tar_option_set]{tar_option_set()}}
+If you set \code{error = "workspace"} in \code{\link[=tar_option_set]{tar_option_set()}}
 or \code{\link[=tar_target]{tar_target()}}, then if that target throws an error
 in \code{\link[=tar_make]{tar_make()}}, it will save its workspace to an RDS file
 in \verb{_targets/workspaces/}. The workspace is a compact reference
@@ -46,7 +46,7 @@ are still in the data store (usually files in \verb{_targets/objects/}).
 \dontrun{
 tmp <- sample(1)
 tar_script({
-  tar_option_set(error = "save") # Required for saving workspaces.
+  tar_option_set(error = "workspace") # Required for saving workspaces.
   tar_pipeline(
     tar_target(x, "loaded"),
     tar_target(y, stop(x))
