+++ R/tar_workspaces_destroy.R
@@ -14,10 +14,10 @@
 #' })
 #' try(tar_make())
 #' file.exists("_targets/workspaces/y") # TRUE
-#' tar_undebug()
+#' tar_workspaces_destroy()
 #' file.exists("_targets/workspaces/y") # FALSE
 #' }
-tar_undebug <- function() {
+tar_workspaces_destroy <- function() {
   unlink(path_scratch_dir(), recursive = TRUE)
   unlink(path_workspaces_dir(), recursive = TRUE)
   invisible()
