+++ man/tar_workspaces_destroy.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tar_undebug.R
-\name{tar_undebug}
-\alias{tar_undebug}
+% Please edit documentation in R/tar_workspaces_destroy.R
+\name{tar_workspaces_destroy}
+\alias{tar_workspaces_destroy}
 \title{Remove debugging files.}
 \usage{
-tar_undebug()
+tar_workspaces_destroy()
 }
 \value{
 Nothing.
@@ -24,7 +24,7 @@ tar_script({
 })
 try(tar_make())
 file.exists("_targets/workspaces/y") # TRUE
-tar_undebug()
+tar_workspaces_destroy()
 file.exists("_targets/workspaces/y") # FALSE
 }
 }
