+++ NAMESPACE
@@ -244,10 +244,11 @@ export(tar_seed)
 export(tar_sitrep)
 export(tar_target)
 export(tar_target_raw)
-export(tar_undebug)
+export(tar_traceback)
 export(tar_validate)
 export(tar_visnetwork)
 export(tar_workspace)
+export(tar_workspaces_destroy)
 export(target_run_worker)
 importFrom(R6,R6Class)
 importFrom(callr,r)
