### 2020-11-13 50ba233
Implement tar_objects()
