+++ tests/testthat/test-tar_objects.R
@@ -1,0 +1,12 @@
+tar_test("tar_objects()", {
+  expect_equal(tar_objects(), character(0))
+  tar_script({
+    tar_option_set(error = "save") # Required for saving workspaces.
+    tar_pipeline(
+      tar_target(x, "value"),
+      tar_target(y, x)
+    )
+  })
+  tar_make(callr_function = NULL)
+  expect_equal(tar_objects(), sort(c("x", "y")))
+})
