### 2020-11-13 f73b0ef
Work on file helpers
