+++ R/tar_objects.R
@@ -6,17 +6,14 @@
 #' if (FALSE) {
 #' tar_dir({
 #' tar_script({
-#'   tar_option_set(error = "save") # Required for saving workspaces.
-#'   tar_pipeline(
-#'     tar_target(x, "value"),
-#'     tar_target(y, x)
-#'   )
+#'   tar_option_set(workspace = "x")
+#'   tar_pipeline(tar_target(x, "value"))
 #' })
 #' tar_make()
-#' tar_objects()
+#' tar_workspaces()
 #' })
 #' }
-tar_objects <- function() {
+tar_workspaces <- function() {
   trn(
     dir.exists(path_default_dir()),
     sort(list.files(path_default_dir(), all.files = TRUE, no.. = TRUE)),
