+++ tests/testthat/test-tar_workspaces.R
@@ -1,9 +1,9 @@
-tar_test("tar_objects()", {
-  expect_equal(tar_objects(), character(0))
+tar_test("tar_workspaces()", {
+  expect_equal(tar_workspaces(), character(0))
   tar_script({
-    tar_option_set(workspace = "x")
-    tar_pipeline(tar_target(x, "value"))
+    tar_option_set(workspace = c("x", "y"))
+    tar_pipeline(tar_target(x, "value"), tar_target(y, "value"))
   })
   tar_make(callr_function = NULL)
-  expect_equal(tar_objects(), sort(c("x", "y")))
+  expect_equal(tar_workspaces(), sort(c("x", "y")))
 })
