+++ tests/testthat/test-tar_destroy.R
@@ -2,7 +2,7 @@ tar_test("tar_destroy('all')", {
   pipeline <- pipeline_init(list(target_init("x", quote(0))))
   local_init(pipeline)$run()
   expect_true(file.exists("_targets"))
-  tar_destroy(what = "all")
+  tar_destroy(destroy = "all")
   expect_false(file.exists("_targets"))
 })
 
@@ -10,7 +10,7 @@ tar_test("tar_destroy('meta')", {
   pipeline <- pipeline_init(list(target_init("x", quote(0))))
   local_init(pipeline)$run()
   expect_true(file.exists("_targets"))
-  tar_destroy(what = "meta")
+  tar_destroy(destroy = "meta")
   expect_true(file.exists("_targets"))
   expect_false(file.exists(file.path("_targets", "meta", "meta")))
 })
@@ -19,7 +19,7 @@ tar_test("tar_destroy('progress')", {
   pipeline <- pipeline_init(list(target_init("x", quote(0))))
   local_init(pipeline)$run()
   expect_true(file.exists("_targets"))
-  tar_destroy(what = "progress")
+  tar_destroy(destroy = "progress")
   expect_true(file.exists("_targets"))
   expect_false(file.exists(file.path("_targets", "meta", "progress")))
 })
@@ -28,7 +28,7 @@ tar_test("tar_destroy('objects')", {
   pipeline <- pipeline_init(list(target_init("x", quote(0))))
   local_init(pipeline)$run()
   expect_true(file.exists("_targets"))
-  tar_destroy(what = "objects")
+  tar_destroy(destroy = "objects")
   expect_true(file.exists("_targets"))
   expect_false(file.exists(file.path("_targets", "objects")))
 })
@@ -37,7 +37,7 @@ tar_test("tar_destroy('scratch')", {
   path <- file.path("_targets", "scratch")
   dir.create(path, recursive = TRUE)
   expect_true(file.exists(path))
-  tar_destroy(what = "scratch")
+  tar_destroy(destroy = "scratch")
   expect_false(file.exists(path))
 })
 
@@ -45,6 +45,6 @@ tar_test("tar_destroy('workspaces')", {
   path <- file.path("_targets", "workspaces")
   dir.create(path, recursive = TRUE)
   expect_true(file.exists(path))
-  tar_destroy(what = "workspaces")
+  tar_destroy(destroy = "workspaces")
   expect_false(file.exists(path))
 })
