+++ R/class_aws_s3.R
@@ -11,7 +11,7 @@ store_produce_path.tar_aws_s3 <- function(store, name, object) {
   assert_chr(bucket, "S3 bucket name must be character.")
   assert_scalar(bucket, "invalid S3 bucket name.")
   assert_nzchar(bucket, "invalid S3 bucket name.")
-  prefix <- store$resources$prefix %||% path_default_dir()
+  prefix <- store$resources$prefix %||% path_objects_dir()
   assert_nonempty(prefix, "S3 object prefix cannot be empty.")
   assert_chr(prefix, "invalid S3 prefix.")
   assert_scalar(prefix, "invalid S3 prefix.")
