### 2020-11-13 427888e
Remove some dontrun blocks
