+++ R/tar_edit.R
@@ -3,7 +3,7 @@
 #' @description Looks for _targets.R in the current working directory.
 #'   Requires the `usethis` package.
 #' @examples
-#' \dontrun{
+#' if (FALSE) { # Interactive only.
 #' tar_dir(
 #' # tar_edit()
 #' )
