+++ R/tar_validate.R
@@ -15,7 +15,7 @@
 #'   reproducible work.
 #' @param callr_arguments A list of arguments to `callr_function`.
 #' @examples
-#' \dontrun{
+#' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
 #' tar_dir({
 #' tar_script(tar_pipeline(tar_target(x, 1 + 1)))
 #' tar_validate()
