+++ R/tar_load.R
@@ -10,7 +10,7 @@
 #'   check/build all the targets (default). Otherwise, you can supply
 #'   symbols, a character vector, or `tidyselect` helpers like [starts_with()].
 #' @examples
-#' if (FALSE) { # Reduce check time for examples.
+#' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
 #' tar_dir({
 #' tar_script(
 #'   tar_pipeline(
@@ -46,7 +46,7 @@ tar_load <- function(
 #'   for any targets that are patterns.
 #' @param envir Environment to put the loaded targets.
 #' @examples
-#' if (FALSE) { # Reduce check time for examples.
+#' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
 #' tar_dir({
 #' tar_script(
 #'   tar_pipeline(
