+++ R/tar_objects.R
@@ -3,7 +3,7 @@
 #' @description List targets currently saved to `_targets/objects/`.
 #'   Does not include dynamic files or cloud storage.
 #' @examples
-#' if (FALSE) {
+#' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
 #' tar_dir({
 #' tar_script({
 #'   tar_option_set(workspace = "x")
