### 2020-11-13 7ed386b
Put dynamic branching methods in an R6 class
