+++ R/class_pattern.R
@@ -322,29 +322,7 @@ pattern_name_branches <- function(parent, niblings) {
 }
 
 pattern_produce_grid <- function(pattern, niblings) {
-  out <- eval(pattern, envir = niblings, enclos = pattern_functions)
+  out <- eval(pattern, envir = niblings, enclos = dynamic_methods$self)
   rownames(out) <- NULL
   out
 }
-
-pattern_functions <- new.env(parent = environment())
-evalq({
-  map <- function(...) {
-    args <- list(...)
-    assert_scalar(
-      unique(map_int(args, nrow)),
-      paste("unequal lengths of vars in", deparse_safe(sys.call()))
-    )
-    do.call(cbind, args)
-  }
-  cross_iteration <- function(x, y) {
-    n_x <- nrow(x)
-    n_y <- nrow(y)
-    index_x <- rep(seq_len(n_x), each = n_y)
-    index_y <- rep(seq_len(n_y), times = n_x)
-    cbind(x[index_x,, drop = FALSE], y[index_y,, drop = FALSE]) # nolint
-  }
-  cross <- function(...) {
-    Reduce(cross_iteration, list(...))
-  }
-}, envir = pattern_functions)
