### 2020-11-13 762e6ff
Add documentation entries for dynamic branching
