+++ man/dynamic-cross.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/tar_dynamic_cross.R
+\name{dynamic-cross}
+\alias{dynamic-cross}
+\alias{cross}
+\title{Dynamic cross}
+\description{
+Create dynamic branches for all combinations of
+the selected dependencies.
+}
+\details{
+\code{pattern = cross(x, y)} creates one dynamic branch for
+each combination of values of \code{x} and \code{y}.
+}
+\examples{
+tar_pipeline(
+  tar_target(x, seq_len(2)),
+  tar_target(y, head(letters, 2)),
+  tar_target(dynamic, c(x, y), pattern = cross(x, y)) # 4 branches
+)
+}
+\seealso{
+Other Dynamic branching: 
+\code{\link{dynamic-map}}
+}
+\concept{Dynamic branching}
