+++ _pkgdown.yml
@@ -92,5 +92,6 @@ reference:
   - '`tar_path`'
   - '`tar_seed`'
 - title: Dynamic branching
+  contents:
   - '`dynamic-map`'
   - '`dynamic-cross`'
