+++ _pkgdown.yml
@@ -83,7 +83,6 @@ reference:
   - '`tar_destroy`'
   - '`tar_invalidate`'
   - '`tar_prune`'
-  - '`tar_workspaces_destroy`'
 - title: Helpers
   contents:
   - '`tar_cancel`'
