+++ R/tar_objects.R
@@ -10,7 +10,7 @@
 #'   tar_pipeline(tar_target(x, "value"))
 #' })
 #' tar_make()
-#' tar_workspaces()
+#' tar_objects()
 #' })
 #' }
 tar_objects <- function() {
