### 2020-11-13 8d6f7d0
Implement user-side helpers tar_dynamic_map() and tar_dynamic_cross()
