### 2020-11-13 d973c01
Programmatically get dynamic method names
