### 2020-11-13 f9ad4fe
Fix docstrings
