### 2020-11-13 f6b5aba
Sketch new patterns and eval with seed
