### 2020-11-13 f21b944
Implement and test tar_dynamic_head()
