+++ NAMESPACE
@@ -216,6 +216,7 @@ export(tar_deps_raw)
 export(tar_destroy)
 export(tar_dir)
 export(tar_dynamic_cross)
+export(tar_dynamic_head)
 export(tar_dynamic_map)
 export(tar_edit)
 export(tar_glimpse)
