### 2020-11-13 b9f0c85
Implement and test tar_dynamic_tail()
