+++ R/class_dynamic.R
@@ -33,14 +33,14 @@ dynamic_class <- R6::R6Class(
       omit_rownames(Reduce(private$cross_iteration, list(...)))
     },
     head = function(x, n = 1L) {
-      utils::head(x = x, n = n)
+      omit_rownames(utils::head(x = x, n = n))
     },
     tail = function(x, n = 1L) {
-      utils::tail(x = x, n = n)
+      omit_rownames(utils::tail(x = x, n = n))
     },
     sample = function(x, n = 1L) {
       index <- sample.int(n = nrow(x), size = n, replace = FALSE)
-      x[index,, drop = FALSE] # nolint
+      omit_rownames(x[index,, drop = FALSE]) # nolint
     }
   )
 )
