+++ man/tar_dynamic_cross.Rd
@@ -44,6 +44,7 @@ tar_dynamic_cross(z, tar_dynamic_map(x, y))
 \seealso{
 Other Dynamic branching: 
 \code{\link{tar_dynamic_head}()},
-\code{\link{tar_dynamic_map}()}
+\code{\link{tar_dynamic_map}()},
+\code{\link{tar_dynamic_tail}()}
 }
 \concept{Dynamic branching}
