+++ man/tar_dynamic_tail.Rd
@@ -1,0 +1,49 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/tar_dynamic_tail.R
+\name{tar_dynamic_tail}
+\alias{tar_dynamic_tail}
+\alias{tail}
+\title{Dynamic tail}
+\usage{
+tar_dynamic_tail(x, n = 1L)
+}
+\arguments{
+\item{...}{data frames with columns to represent targets and
+elements to represent branches or buds. Column names must
+be unique across all the data frames.}
+}
+\description{
+Branch over the first few elements of a pattern.
+}
+\details{
+In \code{\link[=tar_target]{tar_target()}}, \code{pattern = tail(x, n = 5)} makes
+branches out of the first 5 elements of \code{x}. \code{tail()} can be
+composed with other kinds of patterns: for example,
+\code{pattern = tail(map(x, y), n = 5)}.
+To test and experiment with this behavior outside
+the pipeline, use the \code{tar_dynamic_tail()} function.
+}
+\examples{
+# To use dynamic tail for real in a pipeline,
+# call tail() in a target's pattern.
+# The following code goes at the bottom of _targets.R.
+tar_pipeline(
+  tar_target(x, seq_len(26)),
+  tar_target(dynamic, c(x, y), pattern = tail(x, n = 2)) # 2 branches
+)
+# To explore dynamic tail() outside a pipeline,
+# use the tar_dynamic_tail() helper function.
+x <- data.frame(x = seq_len(26))
+tar_dynamic_tail(x, n = 2)
+# You can experiment with more complicated patterns.
+# The following is a representation of pattern = tail(map(x, y), n = 2).
+y <- data.frame(y = letters)
+tar_dynamic_tail(tar_dynamic_map(x, y), n = 2)
+}
+\seealso{
+Other Dynamic branching: 
+\code{\link{tar_dynamic_cross}()},
+\code{\link{tar_dynamic_head}()},
+\code{\link{tar_dynamic_map}()}
+}
+\concept{Dynamic branching}
