### 2020-11-13 a9311dc
Implement and test tar_dynamic_sample
