+++ NAMESPACE
@@ -218,6 +218,7 @@ export(tar_dir)
 export(tar_dynamic_cross)
 export(tar_dynamic_head)
 export(tar_dynamic_map)
+export(tar_dynamic_sample)
 export(tar_dynamic_tail)
 export(tar_edit)
 export(tar_glimpse)

[…]
