### 2020-11-13 da26b75
Test new patterns in pipelines
