+++ tests/testthat/test-tar_dynamic_sample.R
@@ -20,3 +20,26 @@ tar_test("tar_dynamic_sample() with many vars", {
   expect_true(all(out$y %in% letters))
   expect_true(all(out$z %in% LETTERS))
 })
+
+tar_test("sample pattern in target", {
+  tar_script({
+    tar_pipeline(
+      tar_target(x, seq_len(26)),
+      tar_target(dynamic, x, pattern = sample(x, n = 2))
+    )
+  })
+  set.seed(1)
+  tar_make(callr_function = NULL)
+  out <- tar_read(dynamic)
+  tar_destroy()
+  set.seed(2)
+  tar_make(callr_function = NULL)
+  out2 <- tar_read(dynamic)
+  out <- tar_dynamic_sample(x, n = 3)
+  out2 <- tar_dynamic_sample(x, n = 3)
+  expect_equal(out, out2)
+  expect_true(is.data.frame(out))
+  expect_equal(dim(out), c(3L, 1L))
+  expect_equal(colnames(out), "x")
+  expect_true(all(out$x %in% seq_len(26)))
+})
