+++ R/tar_dynamic_head.R
@@ -19,7 +19,7 @@
 #' # The following code goes at the bottom of _targets.R.
 #' tar_pipeline(
 #'   tar_target(x, seq_len(26)),
-#'   tar_target(dynamic, c(x, y), pattern = head(x, n = 2)) # 2 branches
+#'   tar_target(dynamic, x + 1, pattern = head(x, n = 2)) # 2 branches
 #' )
 #' # To explore dynamic head() outside a pipeline,
 #' # use the tar_dynamic_head() helper function.
