+++ R/class_dynamic.R
@@ -38,9 +38,12 @@ dynamic_class <- R6::R6Class(
     tail = function(x, n = 1L) {
       omit_rownames(utils::tail(x = x, n = n))
     },
+    slice = function(x, index = 1L) {
+      omit_rownames(x[index,, drop = FALSE]) # nolint
+    },
     sample = function(x, n = 1L) {
       index <- sample.int(n = nrow(x), size = n, replace = FALSE)
-      omit_rownames(x[index,, drop = FALSE]) # nolint
+      self$slice(x = x, index = index)
     }
   )
 )
