### 2020-11-13 bd26be9
Minor documentation updates for tar_dynamic_*
