+++ R/tar_dynamic_cross.R
@@ -2,7 +2,7 @@
 #' @export
 #' @aliases cross
 #' @family Dynamic branching
-#' @description Iterate over one or more targets in sequence.
+#' @description Branch over all combinations of elements in a pattern.
 #' @details In [tar_target()], `pattern = cross(x, y)`
 #'   creates one dynamic branch
 #'   for each combination of elements of `x` and `y`.
