### 2020-11-14 b6c2995
Rethink pattern emulation
