+++ R/tar_pattern.R
@@ -1,0 +1,98 @@
+#' @title Emulate dynamic branching.
+#' @export
+#' @aliases map cross head tail sample
+#' @description Emulate the dynamic branching process outside a pipeline.
+#'   `tar_pattern()` can help you understand which branches
+#'   will be created when you set the `pattern` argument of [tar_target()].
+#' @details Dynamic branching is a way to programmatically
+#'   create multiple new targets based on the values of other targets,
+#'   all while the pipeline is running. Use the `pattern` argument of
+#'   [tar_target()] to get started. `pattern` accepts a function call
+#'   composed of target names and any of the following patterns:
+#'   * `map()`: create one branch for each corresponding element of
+#'     its arguments. `map(x, y)` creates one branch for the first
+#'     elements of `x` and `y`, another branch for the second
+#'     elements of `x` and `y`, and so on.
+#'   * `cross()`: create one branch for each combination
+#'     of elements of the arguments.
+#'   * `head()`: restrict branching to the first few elements.
+#'   * `tail()`: restrict branching to the last few elements.
+#'   * `sample()`: restrict branching to a random subset of elements.
+#' @return A `tibble` showing the kinds of dynamic branches that
+#'   [tar_target()] would create in a real pipeline with the given `pattern`.
+#'   Each row is a dynamic branch, each column is a dependency target,
+#'   and each element is the name of an upstream bud or branch that the
+#'   downstream branch depends on. Buds are pieces of unbranching targets
+#'   ("stems") and branches are pieces of patterns.
+#' @param pattern Function call with the pattern specification.
+#' @param ... Named character vectors to represent upstream targets.
+#'   Each name is the name of a whole stem or pattern,
+#'   and each vector element is a character with the name of an
+#'   upstream bud or branch. Buds are pieces of unbranching targets
+#'   ("stems") and branches are pieces of patterns. Names must be unique.
+#' @param seed Integer of length 1, random number generator seed to
+#'   emulate the pattern reproducibly. (The `sample()` pattern is random).
+#'   In a real pipeline, the seed is automatically generated
+#'   from the target name in deterministic fashion.
+#' @example
+#' # To use dynamic map for real in a pipeline,
+#' # call map() in a target's pattern.
+#' # The following code goes at the bottom of _targets.R.
+#' tar_pipeline(
+#'   tar_target(x, seq_len(2)),
+#'   tar_target(y, head(letters, 2)),
+#'   tar_target(dynamic, c(x, y), pattern = map(x, y)) # 2 branches
+#' )
+#' Likewise for more complicated patterns.
+#' tar_pipeline(
+#'   tar_target(x, seq_len(2)),
+#'   tar_target(y, head(letters, 2)),
+#'   tar_target(z, head(LETTERS, 2)),
+#'   tar_target(dynamic, c(x, y, z), pattern = cross(z, map(x, y))) #4 branches
+#' )
+#' But you can emulate dynamic branching without running a pipeline
+#' in order to understand the patterns you are creating.
+#' First, you have to make up some names for pieces of targets
+#' ("buds" and branches). The following are reasonably realistic branch names
+#' that `targets` might generate during [tar_make()], but they need not be
+#' totally realistic for emulation purposes.
+#' x <- c("x_550d7456", "x_a20cadbf", "x_eeb00f1b")
+#' y <- c("x_42f35290", "x_f95ad1c7")
+#' z <- c("x_67188e74", "x_5512ec97")
+#' # Then, call `tar_pattern()` to emulate the pattern.
[…]
