### 2020-11-14 f506a22
tar_pattern() needs work
