### 2020-11-14 92f578a
Adjust docs and fix lint
