### 2020-11-14 425402a
Accept lengths, not branch names, in tar_pattern()
