+++ tests/testthat/test-tar_pattern.R
@@ -1,118 +1,94 @@
 tar_test("tar_pattern() input edge cases", {
-  x <- data_frame(x = as.character(seq_len(2)))
-  y <- head(letters, 2)
-  z <- tail(LETTERS, 2)
   expect_error(tar_pattern(2), class = "condition_validate")
-  expect_error(tar_pattern(map(x), y), class = "condition_validate")
-  expect_error(tar_pattern(map(x), y, z = z), class = "condition_validate")
+  expect_error(tar_pattern(map(x), x = "x"), class = "condition_validate")
+  expect_error(tar_pattern(map(x), 3, x = 1), class = "condition_validate")
 })
 
 tar_test("tar_pattern() map unequal length inputs", {
-  x <- head(letters, 3)
-  y <- tail(LETTERS, 2)
   expect_error(
-    tar_pattern(map(x, y), x = x, y = y),
+    tar_pattern(map(x, y), x = 2, y = 3),
     class = "condition_validate"
   )
 })
 
 tar_test("tar_pattern() map", {
-  x <- as.character(seq_len(2))
-  y <- head(letters, 2)
   exp <- data_frame(
-    y = head(letters, 2),
-    x = as.character(seq_len(2))
+    y = paste("y", seq_len(2), sep = "_"),
+    x = paste("x", seq_len(2), sep = "_")
   )
-  out <- tar_pattern(map(y, x), x = x, y = y)
+  out <- tar_pattern(map(y, x), x = 2, y = 2)
   expect_equiv(out, exp)
 })
 
 tar_test("tar_pattern() cross", {
-  x <- as.character(seq_len(2))
-  y <- head(letters, 2)
   exp <- data_frame(
-    y = rep(head(letters, 2), each = 2),
-    x = rep(as.character(seq_len(2)), times = 2)
+    y = rep(paste("y", seq_len(2), sep = "_"), each = 2),
+    x = rep(paste("x", seq_len(2), sep = "_"), times = 2)
   )
-  out <- tar_pattern(cross(y, x), x = x, y = y)
+  out <- tar_pattern(cross(y, x), x = 2, y = 2)
   expect_equiv(out, exp)
 })
 
 tar_test("tar_pattern() composed cross", {
-  x <- as.character(seq_len(2))
-  y <- head(letters, 2)
-  z <- tail(LETTERS, 2)
   exp <- data_frame(
-    y = rep(head(letters, 2), each = 2),
-    z = rep(tail(LETTERS, 2), each = 2),
-    x = rep(as.character(seq_len(2)), times = 2)
+    y = rep(paste("y", seq_len(2), sep = "_"), each = 2),
+    z = rep(paste("z", seq_len(2), sep = "_"), each = 2),
+    x = rep(paste("x", seq_len(2), sep = "_"), times = 2)
   )
-  out <- tar_pattern(cross(map(y, z), x), x = x, y = y, z = z)
+  out <- tar_pattern(cross(map(y, z), x), x = 2, y = 2, z = 2)
   expect_equiv(out, exp)
 })
 
 tar_test("tar_pattern() head with one var", {
-  x <- as.character(seq_len(26))
-  out <- tar_pattern(head(x, n = 3), x = x)
-  exp <- data_frame(x = as.character(seq_len(3)))
+  out <- tar_pattern(head(x, n = 3), x = 26)
+  exp <- data_frame(x = paste("x", seq_len(3), sep = "_"))
   expect_equiv(out, exp)
 })
 
 tar_test("tar_pattern() with many vars", {
-  x <- as.character(seq_len(26))
-  y <- letters
-  z <- LETTERS
[…]
