### 2020-11-15 9ca879e
Move cleanup to on.exit()
