### 2020-11-15 ae72655
Add tar_envir() in order to support tar_render()
