+++ man/tar_envir.Rd
@@ -1,0 +1,36 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/tar_envir.R
+\name{tar_envir}
+\alias{tar_envir}
+\title{For developers only: get the environment of the current target.}
+\usage{
+tar_envir(default = parent.frame())
+}
+\arguments{
+\item{default}{Environment, value to return if \code{tar_envir()}
+is called on its own outside a \code{targets} pipeline.
+Having a default lets users run things without \code{\link[=tar_make]{tar_make()}},
+which helps peel back layers of code and troubleshoot bugs.}
+}
+\value{
+If called from a running target, \code{tar_envir()} returns
+the environment where the target runs its command.
+If called outside a pipeline, the return value is
+whatever the user supplies to \code{default}
+(which defaults to \code{parent.frame()}).
+}
+\description{
+For developers only: get the environment where a
+target runs its command. Inherits from \code{tar_option_get("envir")}.
+}
+\details{
+Users should not call \code{tar_envir()} directly because accidental
+modifications to \code{parent.env(tar_envir())} could break the pipeline.
+\code{tar_envir()} only exists in order to support third-party interface
+packages such as \code{tarchetypes}.
+}
+\examples{
+tar_envir()
+tar_envir(default = parent.frame())
+tar_target(name, tar_envir(default = parent.frame()))
+}
