+++ NAMESPACE
@@ -216,6 +216,7 @@ export(tar_deps_raw)
 export(tar_destroy)
 export(tar_dir)
 export(tar_edit)
+export(tar_envir)
 export(tar_glimpse)
 export(tar_group)
 export(tar_invalidate)
