### 2020-11-17 8a9604b
Add more safeguards to prevent infinite recursion in _targets.R
