### 2020-11-18 5400681
Add generic store_get_packages() and methods
