+++ R/class_torch.R
@@ -39,6 +39,6 @@ store_unserialize_object.tar_torch <- function(store, object) {
 }
 
 #' @export
-store_validate_packages.tar_torch <- function(store) {
-  assert_package("torch")
+store_get_packages.tar_torch <- function(store) {
+  "torch"
 }
