### 2020-11-18 3aea490
Align AWS subclass names to store subclass names
