+++ R/class_aws_fst_tbl.R
@@ -9,7 +9,7 @@ aws_fst_tbl_new <- function(file = NULL, resources = NULL) {
     environment(),
     c(
       "tar_aws_fst_tbl",
-      "tar_aws_s3",
+      "tar_aws",
       "tar_fst_tbl",
       "tar_fst",
       "tar_store"
