+++ R/class_aws_rds.R
@@ -7,7 +7,7 @@ aws_rds_new <- function(file = NULL, resources = NULL) {
   force(resources)
   enclass(
     environment(),
-    c("tar_aws_s3_rds", "tar_aws_s3", "tar_rds", "tar_store")
+    c("tar_aws_rds", "tar_aws", "tar_rds", "tar_store")
   )
 }
 
