+++ R/class_aws.R
@@ -5,7 +5,7 @@
 # external contributors from the open source community.
 # nocov start
 #' @export
-store_produce_path.tar_aws_s3 <- function(store, name, object) {
+store_produce_path.tar_aws <- function(store, name, object) {
   bucket <- store$resources$bucket
   assert_nonempty(bucket, "S3 bucket name cannot be empty.")
   assert_chr(bucket, "S3 bucket name must be character.")
@@ -21,23 +21,23 @@ store_produce_path.tar_aws_s3 <- function(store, name, object) {
 }
 
 #' @export
-store_produce_stage.tar_aws_s3 <- function(store, name, object) {
+store_produce_stage.tar_aws <- function(store, name, object) {
   tempfile(pattern = name)
 }
 
-store_aws_s3_bucket <- function(path) {
+store_aws_bucket <- function(path) {
   path[1]
 }
 
-store_aws_s3_key <- function(path) {
+store_aws_key <- function(path) {
   path[2]
 }
 
 #' @export
-store_read_object.tar_aws_s3 <- function(store) {
+store_read_object.tar_aws <- function(store) {
   path <- store$file$path
-  bucket <- store_aws_s3_bucket(path)
-  key <- store_aws_s3_key(path)
+  bucket <- store_aws_bucket(path)
+  key <- store_aws_key(path)
   tmp <- tempfile()
   on.exit(unlink(tmp))
   aws.s3::save_object(object = key, bucket = bucket, file = tmp)
@@ -45,16 +45,16 @@ store_read_object.tar_aws_s3 <- function(store) {
 }
 
 #' @export
-store_write_object.tar_aws_s3 <- function(store, object) {
+store_write_object.tar_aws <- function(store, object) {
   stage <- store$file$stage
   dir_create(dirname(stage))
   store_write_path(store, store_coerce_object(store, object), stage)
 }
 
 #' @export
-store_upload_object.tar_aws_s3 <- function(store) {
-  key <- store_aws_s3_key(store$file$path)
-  bucket <- store_aws_s3_bucket(store$file$path)
+store_upload_object.tar_aws <- function(store) {
+  key <- store_aws_key(store$file$path)
+  bucket <- store_aws_bucket(store$file$path)
   hash <- store$file$hash
   aws.s3::put_object(
     file = store$file$stage,
@@ -65,17 +65,17 @@ store_upload_object.tar_aws_s3 <- function(store) {
   )
 }
 
-store_aws_s3_exists <- function(key, bucket) {
+store_aws_exists <- function(key, bucket) {
   suppressWarnings(aws.s3::object_exists(object = key, bucket = bucket))
 }
 
-store_aws_s3_hash <- function(key, bucket) {
+store_aws_hash <- function(key, bucket) {
   head <- aws.s3::head_object(object = key, bucket = bucket)
   hash_worker <- attr(head, "x-amz-meta-targets-hash")
 }
 
 #' @export
-store_late_hash.tar_aws_s3 <- function(store) {
+store_late_hash.tar_aws <- function(store) {
   file <- file_init(path = store$file$stage)
   file_update_hash(file)
   store$file$hash <- file$hash
@@ -84,27 +84,27 @@ store_late_hash.tar_aws_s3 <- function(store) {
 }
 
 #' @export
-store_has_correct_hash.tar_aws_s3 <- function(store) {
-  bucket <- store_aws_s3_bucket(store$file$path)
-  key <- store_aws_s3_key(store$file$path)
+store_has_correct_hash.tar_aws <- function(store) {
+  bucket <- store_aws_bucket(store$file$path)
+  key <- store_aws_key(store$file$path)
   trn(
-    store_aws_s3_exists(key, bucket),
[…]
