+++ NAMESPACE
@@ -40,11 +40,11 @@ S3method(store_early_hash,default)
 S3method(store_early_hash,tar_store_file)
 S3method(store_early_hash,tar_url)
 S3method(store_ensure_correct_hash,default)
-S3method(store_ensure_correct_hash,tar_aws_s3)
+S3method(store_ensure_correct_hash,tar_aws)
 S3method(store_ensure_correct_hash,tar_store_file)
 S3method(store_ensure_correct_hash,tar_url)
 S3method(store_get_packages,default)
-S3method(store_get_packages,tar_aws_s3)
+S3method(store_get_packages,tar_aws)
 S3method(store_get_packages,tar_fst)
 S3method(store_get_packages,tar_fst_dt)
 S3method(store_get_packages,tar_fst_tbl)
@@ -53,10 +53,10 @@ S3method(store_get_packages,tar_qs)
 S3method(store_get_packages,tar_torch)
 S3method(store_get_packages,tar_url)
 S3method(store_has_correct_hash,default)
-S3method(store_has_correct_hash,tar_aws_s3)
+S3method(store_has_correct_hash,tar_aws)
 S3method(store_has_correct_hash,tar_url)
 S3method(store_late_hash,default)
-S3method(store_late_hash,tar_aws_s3)
+S3method(store_late_hash,tar_aws)
 S3method(store_late_hash,tar_store_file)
 S3method(store_late_hash,tar_url)
 S3method(store_new,default)
@@ -70,14 +70,14 @@ S3method(store_new,rds)
 S3method(store_new,torch)
 S3method(store_new,url)
 S3method(store_produce_path,default)
-S3method(store_produce_path,tar_aws_s3)
+S3method(store_produce_path,tar_aws)
 S3method(store_produce_path,tar_store_file)
 S3method(store_produce_path,tar_url)
 S3method(store_produce_stage,default)
-S3method(store_produce_stage,tar_aws_s3)
-S3method(store_produce_stage,tar_aws_s3_file)
-S3method(store_read_object,tar_aws_s3)
-S3method(store_read_object,tar_aws_s3_file)
+S3method(store_produce_stage,tar_aws)
+S3method(store_produce_stage,tar_aws_file)
+S3method(store_read_object,tar_aws)
+S3method(store_read_object,tar_aws_file)
 S3method(store_read_path,tar_fst)
 S3method(store_read_path,tar_fst_dt)
 S3method(store_read_path,tar_fst_tbl)
@@ -93,18 +93,18 @@ S3method(store_serialize_object,tar_torch)
 S3method(store_serialize_value,default)
 S3method(store_serialize_value,tar_unexportable)
 S3method(store_sync_file_meta,default)
-S3method(store_sync_file_meta,tar_aws_s3)
+S3method(store_sync_file_meta,tar_aws)
 S3method(store_sync_file_meta,tar_url)
 S3method(store_unload,default)
-S3method(store_unload,tar_aws_s3_file)
+S3method(store_unload,tar_aws_file)
 S3method(store_unserialize_object,default)
 S3method(store_unserialize_object,tar_keras)
 S3method(store_unserialize_object,tar_torch)
 S3method(store_unserialize_value,default)
 S3method(store_unserialize_value,tar_unexportable)
-S3method(store_upload_object,tar_aws_s3)
+S3method(store_upload_object,tar_aws)
 S3method(store_write_object,default)
-S3method(store_write_object,tar_aws_s3)
+S3method(store_write_object,tar_aws)
 S3method(store_write_object,tar_store_file)
 S3method(store_write_object,tar_url)
 S3method(store_write_path,tar_fst)
