### 2020-11-18 4d9b7f3
Add methods to get package reqs more easily
