+++ tests/interactive/test-tar_renv.R
@@ -1,25 +1,0 @@
-tar_test("Run tar_renv() interactively", {
-  expect_false(file.exists("_packages.R"))
-  pkgs1 <- c("tibble", "qs")
-  pkgs2 <- c("tibble")
-  lines1 <- c(
-  "# Generated by targets::tar_env(): do not edit by hand",
-  paste0("library(", pkgs1, ")")
-  )
-  lines2 <- c(
-  "# Generated by targets::tar_env(): do not edit by hand",
-  paste0("library(", pkgs2, ")")
-  )
-  tar_renv(packages = pkgs1)
-  expect_true(file.exists("_packages.R"))
-  expect_equal(readLines("_packages.R"), lines1)
-  old_tar_ask <- Sys.getenv("TAR_ASK")
-  Sys.setenv(TAR_ASK = "true")
-  tar_renv(packages = pkgs2) # See a menu and choose 2.
-  expect_equal(readLines("_packages.R"), lines1)
-  tar_renv(packages = pkgs2) # See a menu and choose 0.
-  expect_equal(readLines("_packages.R"), lines1)
-  tar_renv(packages = pkgs2) # See a menu and choose 1.
-  expect_equal(readLines("_packages.R"), lines2)
-  Sys.setenv(TAR_ASK = old_tar_ask)
-})
