+++ R/tar_renv.R
@@ -5,17 +5,18 @@
 #'   (by default, named `_packages.R` in the root project directory).
 #'   Each package is written to a separate line
 #'   as a standard [library()] call (e.g. `library(package)`) so
-#'   `renv` can identify them automatically. This is only
-#'   necessary if the user invokes [`tar_option_set()`] or [`tar_target()`] to
-#'   declare packages rather than the usual `library()` or namespaced call
+#'   `renv` can identify them automatically.
+#'
+#'   This function tries to identify package dependencies that would be
+#'   hidden to `renv` because they are declared using the `targets` API.
+#'   Thus, this is only necessary if using [`tar_option_set()`] or
+#'   [`tar_target()`] to use specialized storage formats or manage packages,
+#'   rather than the usual `library()` or namespaced call
 #'   (i.e. `package::function()`). See `Details` for more information.
 #' @details This function gets called for its side-effect, which writes
-#'   package dependencies provided to the `packages` argument to
-#'   a script. If you do not explicitly set the `packages` argument,
-#'   `tar_renv()` will only detect the packages that
-#'   `tar_option_get("packages")` returns at that point in the code.
-#'   The generated file should __not__ be edited by hand and will be
-#'   overwritten each time `tar_renv()` runs non-interactively.
+#'   package dependencies to a script. The generated file should __not__ be
+#'   edited by hand and will be overwritten each time `tar_renv()` runs
+#'   non-interactively.
 #'
 #'   With the script written by `tar_renv()`, `renv` is able to crawl the
 #'   file to identify package dependencies (with `renv::dependencies()`).
@@ -25,35 +26,55 @@
 #'   library. See <https://rstudio.github.io/renv/index.html> for
 #'   more information.
 #' @return Nothing, invisibly.
-#' @inheritParams tar_option_set
+#' @inheritParams tar_prune
+#' @param extras Character vector of additional packages to declare as
+#'   project dependencies.
 #' @param path Character of length 1, path to the script file to
 #'   populate with `library()` calls.
-#' @param ask Logical, whether to ask before writing if `_packages.R`
-#'   already exists. If `NULL`, defaults to `Sys.getenv("TAR_ASK")`.
-#'   (Set to `"true"` or `"false"` with `Sys.setenv()`).
-#'   If `ask` and the `TAR_ASK` environment variable are both
-#'   indeterminate, defaults to `interactive()`.
 #' @examples
 #' tar_dir({
-#'   tar_option_set(packages = c("tibble", "qs"))
+#'   tar_script({
+#'     tar_option_set(packages = c("tibble", "qs"))
+#'     tar_pipeline()
+#'   })
 #'   tar_renv()
 #'   readLines("_packages.R")
 #' })
 tar_renv <- function(
-  packages = targets::tar_option_get("packages"),
+  extras = c("targets", "clustermq", "future", "rstudioapi", "visNetwork"),
   path = "_packages.R",
-  ask = NULL
+  callr_function = callr::r,
+  callr_arguments = list()
 ) {
-  assert_chr(path)
-  assert_scalar(path)
-  if (!tar_should_overwrite(ask, path)) {
-    # Covered in tests/interactive/test-tar_renv.R. # nolint
-    return(invisible()) #nocov
-  }
+  assert_target_script()
+  assert_chr(extras, "extras must be a character vector")
+  assert_chr(path, "path must have type character")
+  assert_scalar(path, "path must have length 1")
+  assert_callr_function(callr_function)
+  assert_list(callr_arguments, "callr_arguments must be a list.")
+  callr_outer(
+    targets_function = tar_renv_inner,
[…]
