+++ tests/testthat/test-tar_renv.R
@@ -1,13 +1,16 @@
 tar_test("tar_renv() works", {
   expect_false(file.exists("_packages.R"))
-  tar_option_set(packages = c("tibble", "qs"))
-  tar_renv(ask = FALSE)
+  tar_script({
+    tar_option_set(packages = c("tibble", "qs")); tar_pipeline()
+    })
+  tar_renv()
   expect_true(file.exists("_packages.R"))
   expect_equal(
     readLines("_packages.R"),
     c(
-      "# Generated by targets::tar_env(): do not edit by hand",
-      paste0("library(", c("tibble", "qs"), ")")
+      "# Generated by targets::tar_renv(): do not edit by hand",
+      paste0("library(", c("clustermq", "future", "qs", "rstudioapi",
+                           "tibble", "visNetwork"), ")")
       )
     )
 })
@@ -15,14 +18,17 @@ tar_test("tar_renv() works", {
 tar_test("tar_renv() works with custom path", {
   path <- tempfile()
   expect_false(file.exists(path))
-  tar_option_set(packages = c("tibble", "qs"))
-  tar_renv(ask = FALSE, path = path)
+  tar_script({
+    tar_option_set(packages = c("tibble", "qs")); tar_pipeline()
+    })
+  tar_renv(path = path)
   expect_true(file.exists(path))
   expect_equal(
     readLines(path),
     c(
-      "# Generated by targets::tar_env(): do not edit by hand",
-      paste0("library(", c("tibble", "qs"), ")")
+      "# Generated by targets::tar_renv(): do not edit by hand",
+      paste0("library(", c("clustermq", "future", "qs", "rstudioapi",
+                           "tibble", "visNetwork"), ")")
     )
   )
 })
