### 2020-11-21 #401 @maurolepore
I'm excited to announce that @limnoliver and @tjmahr will be reviewing this submission. Please notice this submission includes two closely related packages: "targets" and "tarchetypes".

There is a third package, "targets.manual", that provides documentation; your feedback on "targets.manual" is welcome but a full review is not required.

The due date for your reviews is 2020-12-19. This is four weeks from today -- a bit longer than usual considering this double-submission and COVID.

Please let me know if you have any issues, questions or comments. I look forward to working with you all.
