### 2020-11-23 e54950a
Return full patterns from tar_manifest()
