### 2020-11-23 f2890f3
Remove . from nzchar error msg
