+++ tests/testthat/test-tar_renv.R
@@ -1,28 +1,151 @@
 tar_test("tar_renv() works", {
   expect_false(file.exists("_packages.R"))
-  tar_option_set(packages = c("tibble", "qs"))
-  tar_renv(ask = FALSE)
+  tar_script({
+    tar_option_set(packages = c("tibble", "qs")); tar_pipeline()
+  })
+  tar_renv(callr_function = NULL)
   expect_true(file.exists("_packages.R"))
   expect_equal(
     readLines("_packages.R"),
     c(
-      "# Generated by targets::tar_env(): do not edit by hand",
-      paste0("library(", c("tibble", "qs"), ")")
-      )
+      "# Generated by targets::tar_renv(): do not edit by hand",
+      paste0("library(", c("clustermq", "future", "qs", "rstudioapi",
+                           "tibble", "visNetwork"), ")")
     )
+  )
 })
 
 tar_test("tar_renv() works with custom path", {
   path <- tempfile()
   expect_false(file.exists(path))
-  tar_option_set(packages = c("tibble", "qs"))
-  tar_renv(ask = FALSE, path = path)
+  tar_script({
+    tar_option_set(packages = c("tibble", "qs")); tar_pipeline()
+  })
+  tar_renv(path = path, callr_function = NULL)
   expect_true(file.exists(path))
   expect_equal(
     readLines(path),
     c(
-      "# Generated by targets::tar_env(): do not edit by hand",
-      paste0("library(", c("tibble", "qs"), ")")
+      "# Generated by targets::tar_renv(): do not edit by hand",
+      paste0("library(", c("clustermq", "future", "qs", "rstudioapi",
+                           "tibble", "visNetwork"), ")")
+    )
+  )
+})
+
+tar_test("tar_renv() packages set in tar_target()", {
+  tar_script({
+    tar_pipeline(
+      tar_target(x, "foo", packages = "tibble")
+    )
+  })
+  tar_renv()
+  pkgs_when_opt_not_set <-
+    c(
+      "base",
+      "clustermq",
+      "datasets",
+      "future",
+      "grDevices",
+      "graphics",
+      "methods",
+      "rstudioapi",
+      "stats",
+      "targets",
+      "tibble",
+      "utils",
+      "visNetwork"
+    )
+  readLines("_packages.R")
+  expect_true(file.exists("_packages.R"))
+  expect_equal(
+    readLines("_packages.R"),
+    c(
+      "# Generated by targets::tar_renv(): do not edit by hand",
+      paste0("library(", pkgs_when_opt_not_set, ")")
+    )
+  )
+})
+
+tar_test("tar_renv() formats set in tar_target()", {
+  tar_script({
+    tar_pipeline(
+      tar_target(x, "foo", format = "qs")
+    )
+  })
+  tar_renv()
+  pkgs_when_pkg_not_set <-
+    c(
+      "base",
+      "clustermq",
+      "datasets",
+      "future",
+      "grDevices",
+      "graphics",
+      "methods",
+      "qs",
+      "rstudioapi",
+      "stats",
+      "targets",
+      "utils",
+      "visNetwork"
+    )
+  readLines("_packages.R")
+  expect_true(file.exists("_packages.R"))
+  expect_equal(
+    readLines("_packages.R"),
+    c(
+      "# Generated by targets::tar_renv(): do not edit by hand",
+      paste0("library(", pkgs_when_pkg_not_set, ")")
+    )
+  )
+})
+
+tar_test("tar_renv() packages set in tar_option_set()", {
+  tar_script({
+    tar_option_set(
+      packages = "tibble",
+      format = "qs"
+    )
+    tar_pipeline()
+  })
+  tar_renv(callr_function = NULL)
[…]
