+++ tests/testthat/test-tar_renv.R
@@ -5,15 +5,14 @@ tar_test("tar_renv() works", {
     tar_pipeline()
   })
   tar_renv()
-  pkgs <-
-    c(
-      "clustermq",
-      "future",
-      "qs",
-      "rstudioapi",
-      "tibble",
-      "visNetwork"
-    )
+  pkgs <- c(
+    "clustermq",
+    "future",
+    "qs",
+    "rstudioapi",
+    "tibble",
+    "visNetwork"
+  )
   expect_true(file.exists("_packages.R"))
   expect_equal(
     readLines("_packages.R"),
@@ -32,15 +31,14 @@ tar_test("tar_renv() works with custom path", {
     tar_pipeline()
   })
   tar_renv(path = path, callr_function = NULL)
-  pkgs <-
-    c(
-      "clustermq",
-      "future",
-      "qs",
-      "rstudioapi",
-      "tibble",
-      "visNetwork"
-    )
+  pkgs <- c(
+    "clustermq",
+    "future",
+    "qs",
+    "rstudioapi",
+    "tibble",
+    "visNetwork"
+  )
   expect_true(file.exists(path))
   expect_equal(
     readLines(path),
@@ -59,14 +57,13 @@ tar_test("tar_renv() packages set in tar_target()", {
     )
   })
   tar_renv(callr_function = NULL)
-  pkgs <-
-    c(
-      "clustermq",
-      "future",
-      "rstudioapi",
-      "tibble",
-      "visNetwork"
-    )
+  pkgs <- c(
+    "clustermq",
+    "future",
+    "rstudioapi",
+    "tibble",
+    "visNetwork"
+  )
   expect_true(file.exists("_packages.R"))
   expect_equal(
     readLines("_packages.R"),
@@ -85,14 +82,13 @@ tar_test("tar_renv() formats set in tar_target()", {
     )
   })
   tar_renv(callr_function = NULL)
-  pkgs <-
-    c(
-      "clustermq",
-      "future",
-      "qs",
-      "rstudioapi",
-      "visNetwork"
-    )
+  pkgs <- c(
+    "clustermq",
+    "future",
+    "qs",
+    "rstudioapi",
+    "visNetwork"
+  )
   expect_true(file.exists("_packages.R"))
   expect_equal(
     readLines("_packages.R"),
@@ -112,15 +108,14 @@ tar_test("tar_renv() packages set in tar_option_set()", {
     tar_pipeline()
   })
   tar_renv(callr_function = NULL)
-  pkgs <-
-    c(
-      "clustermq",
-      "future",
-      "qs",
-      "rstudioapi",
-      "tibble",
-      "visNetwork"
-    )
+  pkgs <- c(
+    "clustermq",
+    "future",
+    "qs",
+    "rstudioapi",
+    "tibble",
+    "visNetwork"
+  )
   expect_true(file.exists("_packages.R"))
   expect_equal(
     readLines("_packages.R"),
@@ -135,7 +130,7 @@ tar_test("tar_renv() non-default extra packages", {
   tar_script({
     tar_option_set(packages = character())
     tar_pipeline()
-    })
+  })
   tar_renv(extras = "tibble", callr_function = NULL)
   pkgs <- "tibble"
   expect_true(file.exists("_packages.R"))
@@ -145,5 +140,16 @@ tar_test("tar_renv() non-default extra packages", {
       "# Generated by targets::tar_renv(): do not edit by hand",
       paste0("library(", pkgs, ")")
     )
+  )
+})
+
+tar_test("tar_env() cannot go inside _targets.R", {
+  tar_script({
+    tar_renv()
+    tar_pipeline(tar_target(x, 1))
+  })
+  expect_error(
+    tar_validate(callr_function = NULL),
+    class = "condition_validate"
   )
 })
