+++ R/utils_assert.R
@@ -231,6 +231,7 @@ assert_target_script <- function() {
     "network",
     "outdated",
     "prune",
+    "renv",
     "sitrep",
     "validate",
     "visnetwork"
