### 2020-11-23 bceabc9
Repair target_get_packages() for patterns
