+++ R/class_target.R
@@ -369,9 +369,10 @@ target_unserialize_value <- function(target) {
 }
 
 target_get_packages <- function(target) {
-  command <- target$command$packages
-  store <- store_get_packages(target$store)
-  sort(unique(c(command, store)))
+  packages_command <- target$command$packages
+  store <- target$store %||% settings_produce_store(target$settings)
+  packages_store <- store_get_packages(store)
+  sort(unique(c(packages_command, packages_store)))
 }
 
 target_validate <- function(target) {
