+++ tests/testthat/test-tar_renv.R
@@ -99,6 +99,32 @@ tar_test("tar_renv() formats set in tar_target()", {
   )
 })
 
+tar_test("tar_renv() formats set in pattern targets", {
+  tar_script({
+    tar_option_set(packages = character())
+    tar_pipeline(
+      tar_target(y, "foo", packages = character(0)),
+      tar_target(x, y, format = "qs", pattern = map(y))
+    )
+  })
+  tar_renv(callr_function = NULL)
+  pkgs <- c(
+    "clustermq",
+    "future",
+    "qs",
+    "rstudioapi",
+    "visNetwork"
+  )
+  expect_true(file.exists("_packages.R"))
+  expect_equal(
+    readLines("_packages.R"),
+    c(
+      "# Generated by targets::tar_renv(): do not edit by hand",
+      paste0("library(", pkgs, ")")
+    )
+  )
+})
+
 tar_test("tar_renv() packages set in tar_option_set()", {
   tar_script({
     tar_option_set(
