### 2020-11-23 89be7e5
Pedantically use S3 for target_get_packages()
