+++ R/class_builder.R
@@ -155,6 +155,13 @@ target_sync_file_meta.tar_builder <- function(target, meta) {
 }
 
 #' @export
+target_get_packages.tar_builder <- function(target) {
+  packages_command <- target$command$packages
+  packages_store <- store_get_packages(target$store)
+  sort(unique(c(packages_command, packages_store)))
+}
+
+#' @export
 target_validate.tar_builder <- function(target) {
   NextMethod()
   if (!is.null(target$store)) {
