+++ NAMESPACE
@@ -126,6 +126,8 @@ S3method(target_ensure_buds,tar_stem)
 S3method(target_get_children,default)
 S3method(target_get_children,tar_pattern)
 S3method(target_get_children,tar_stem)
+S3method(target_get_packages,default)
+S3method(target_get_packages,tar_builder)
 S3method(target_get_parent,tar_branch)
 S3method(target_get_parent,tar_bud)
 S3method(target_get_parent,tar_pattern)
