### 2020-11-25 9575b1b
Update README.md
There seems to be many dependencies, including `tarchetypes` and we want to user experience to be as smooth as possible I guess? Will still need to install `biglm` after this function.
