### 2020-11-30 #401 @wlandau
I have a new package on GitHub called [`stantargets`](https://github.com/wlandau/stantargets), an extension to [`targets`](https://github.com/wlandau/targets) for Stan-powered Bayesian data analysis. [`stantargets`](https://github.com/wlandau/stantargets) makes it super easy to set up useful pipelines without having to write many functions or think about branching. It supports all the features of [`cmdstanr`](https://github.com/stan-dev/cmdstanr) (MCMC, variational Bayes, optimization), and it supports both single-fit workflows and multi-rep simulation studies ([vignettes here](https://wlandau.github.io/stantargets/articles/index.html)).

I mention [`stantargets`](https://github.com/wlandau/stantargets) here to liven up the discussion and to give you an overall sense of where I am taking [`targets`](https://github.com/wlandau/targets). I intend [`stantargets`](https://github.com/wlandau/stantargets) to set a precedent for other methodology-specific workflow frameworks built on top of [`targets`](https://github.com/wlandau/targets), e.g. possibly for machine learning with [`keras`](https://github.com/rstudio/keras) and [`torch`](https://github.com/rstudio/torch).
