### 2020-12-01 ceb5240
Sketch new imports class
