### 2020-12-01 befa458
Finalize imports class tests
