+++ R/class_active.R
@@ -24,7 +24,7 @@ active_class <- R6::R6Class(
     ensure_meta = function() {
       self$meta$validate()
       self$meta$database$preprocess(write = TRUE)
-      self$meta$record_imports(self$pipeline$envir, self$pipeline)
+      self$meta$record_imports(self$pipeline$imports, self$pipeline)
       self$meta$restrict_records(self$pipeline)
     },
     produce_exports = function(envir) {
