+++ DESCRIPTION
@@ -57,6 +57,7 @@ Suggests:
   keras (>= 2.2.5.0),
   knitr (>= 1.30),
   rmarkdown (>= 2.4),
+  pkgload (>= 1.1.0),
   qs (>= 0.23.2),
   rstudioapi (>= 0.11),
   testthat (>= 3.0.0),
