+++ R/class_meta.R
@@ -59,7 +59,6 @@ meta_class <- R6::R6Class(
     hash_dep = function(name, target, pipeline) {
       exists <- self$exists_record(name) &&
         pipeline_exists_object(pipeline, name)
-      )
       trn(
         exists,
         self$get_record(name)$data,
