+++ R/class_active.R
@@ -50,7 +50,7 @@ active_class <- R6::R6Class(
     process_target = function(name) {
       target <- pipeline_get_target(self$pipeline, name)
       target_debug(target)
-      target_update_depend(target, meta)
+      target_update_depend(target, self$pipeline, self$meta)
       trn(
         target_should_run(target, self$meta),
         self$run_target(name),
