### 2020-12-02 4607a89
Update signatures in tests
