### 2020-12-05 a966ade
Link to nyhackr talk
