+++ README.md
@@ -80,8 +80,8 @@ install_github("wlandau/targets")
 
   - [R/Pharma 2020
     (9:24)](https://www.youtube.com/watch?v=GRqKJBaC5g4&list=PLMtxz1fUYA5C0YflXsR8EEAQXfjntlV1H&index=6)
-  - [New York Open Statistical Programming Meetup,
-    December 2020](https://youtu.be/Gqn7Xn4d5NI)
+  - [New York Open Statistical Programming Meetup, December 2020
+    (1:54:28)](https://youtu.be/Gqn7Xn4d5NI)
   - [LA R Users Meetup, October 2020
     (1:14:40)](https://www.youtube.com/watch?v=Qq25BUxpJu4)
 
