### 2020-12-12 68f71f8
Set deployment to "main" in tar_make() (but not tar_make_clustermq() etc.)
