### 2020-12-14 #401 @wlandau
To give you more of a sense of where I am going with this, I just announced the targetopia, an ecosystem of packages to democratize Make-like pipelines in R. I hope to add more packages and cover more kinds of data science in 2021.

* Packages: https://wlandau.github.io/targetopia.html
* Blog: https://wlandau.github.io/posts/2020-12-14-targetopia/
