+++ NAMESPACE
@@ -251,6 +251,7 @@ export(tar_seed)
 export(tar_sitrep)
 export(tar_target)
 export(tar_target_raw)
+export(tar_test)
 export(tar_traceback)
 export(tar_validate)
 export(tar_visnetwork)
@@ -305,5 +306,6 @@ importFrom(vctrs,vec_rbind)
 importFrom(vctrs,vec_size)
 importFrom(vctrs,vec_slice)
 importFrom(withr,local_dir)
+importFrom(withr,local_envvar)
 importFrom(withr,with_dir)
 importFrom(withr,with_seed)
