+++ _pkgdown.yml
@@ -98,3 +98,7 @@ reference:
   contents:
   - '`tar_cancel`'
   - '`tar_group`'
+- title: Targetopia
+  contents:
+  - '`tar_dir`'
+  - '`tar_test`'
