### 2020-12-16 #401 @wlandau
Maybe targetopia is slightly out of scope for this review, but I would at least like to mention this: I envision it to cover more fields of data science than I have the expertise to cover. (I am starting with Bayesian statistics because that is where I am strongest.) So if anyone reading this would like to contribute a targetopia package, that would be amazing. I just added some tips at the bottom of <https://wlandau.github.io/targetopia.html>, and I would be eager to get in touch.
