+++ inst/paper.bib
@@ -87,3 +87,25 @@
   https://docs.metaflow.org/,
   https://github.com/Netflix/metaflow},
 }
+
+@Manual{
+clustermq,
+  title = {clustermq: Evaluate Function Calls on HPC Schedulers (LSF, SGE, SLURM,
+PBS/Torque)},
+  author = {Michael Schubert},
+  year = {2020},
+  note = {R package version 0.8.95.1},
+  url = {https://CRAN.R-project.org/package=clustermq},
+}
+
+@Misc{
+  future,
+  author = {Henrik Bengtsson},
+  title = {A Unifying Framework for Parallel and Distributed Processing in R using Futures},
+  year = {2020},
+  month = {aug},
+  eprint = {2008.00553},
+  archiveprefix = {arXiv},
+  primaryclass = {cs.DC},
+  url = {https://arxiv.org/abs/2008.00553},
+}
