### 2020-12-17 17d6a56
Choose more colorblind-friendly colors
and show "waiting" instead of "undefined"
if outdated = FALSE. Add a non-gray color for
"waiting" and test with colorblind filter.
