+++ tests/testthat/test-class_glimpse.R
@@ -27,7 +27,7 @@ tar_test("vertices and edges", {
   exp_imports <- data_frame(
     name = c("g", "f", "h", "i"),
     type = c("function", "function", "object", "object"),
-    status = rep("undefined", 4L),
+    status = rep("waiting", 4L),
     seconds = NA_real_,
     bytes = NA_real_,
     branches = NA_real_
@@ -35,7 +35,7 @@ tar_test("vertices and edges", {
   exp_targets <- data_frame(
     name = c("x", "y", "z"),
     type = rep("stem", 3L),
-    status = rep("undefined", 3L),
+    status = rep("waiting", 3L),
     seconds = NA_real_,
     bytes = NA_real_,
     branches = NA_real_
