### 2020-12-17 deec0c4
Update semi-automated test
