+++ man/tar_watch_ui.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/tar_watch.R
+\name{tar_watch_ui}
+\alias{tar_watch_ui}
+\title{Shiny module UI for tar_watch()}
+\usage{
+tar_watch_ui(id, label = "tar_watch")
+}
+\arguments{
+\item{id}{An ID string that corresponds with the ID used to call the module's
+UI function.}
+
+\item{label}{Label for the module.}
+}
+\description{
+Use \code{tar_watch_ui()} and \code{\link[=tar_watch_server]{tar_watch_server()}}
+to include \code{\link[=tar_watch]{tar_watch()}} as a Shiny module in an app.
+}
+\examples{
+str(tar_watch_ui("my_id"))
+}
