+++ man/tar_watch_server.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/tar_watch.R
+\name{tar_watch_server}
+\alias{tar_watch_server}
+\title{Shiny module server for tar_watch()}
+\usage{
+tar_watch_server(id, height = "600px")
+}
+\arguments{
+\item{id}{An ID string that corresponds with the ID used to call the module's
+UI function.}
+
+\item{height}{Character of length 1,
+height of the \code{visNetwork} widget.}
+}
+\description{
+Use \code{\link[=tar_watch_ui]{tar_watch_ui()}} and \code{tar_watch_server()}
+to include \code{\link[=tar_watch]{tar_watch()}} as a Shiny module in an app.
+}
+\examples{
+# tar_watch_server("my_id") # Only call inside an app.
+}
