### 2020-12-17 9f20035
Add inputs for refreshing the app
