### 2020-12-17 7d33655
Add inputs for refresh and interval
