### 2020-12-17 831ac16
Use invalidateLater()
