### 2020-12-17 91b45a9
Implement tar_watch()
