+++ R/tar_watch.R
@@ -1,35 +1,114 @@
 # nocov start # Tested in tests/interactive/test-tar_watch.R
 #' @title Shiny app to watch the dependency graph.
 #' @export
-#' @description Calls [tar_visnetwork()] and offers a GUI
-#'   to select settings.
+#' @description Calls [tar_visnetwork()] every few seconds.
+#' @details The controls of the app are in the left panel.
+#'   The `seconds` control is the number of seconds between
+#'   refreshes of the graph, and the other settings match
+#'   the arguments of [`tar_visnetwork()`].
+#' @return A Shiny app.
+#' @inheritParams tar_watch_ui
 #' @examples
 #' if (FALSE) { # Only run interactively.
 #' tar_dir({
-#'   tar_script()
-#'   tar_watch()
+#' tar_script({
+#'   sleep_run <- function(...) {
+#'     Sys.sleep(10)
+#'   }
+#'   tar_pipeline(
+#'     tar_target(settings, sleep_run()),
+#'     tar_target(data1, sleep_run(settings)),
+#'     tar_target(data2, sleep_run(settings)),
+#'     tar_target(data3, sleep_run(settings)),
+#'     tar_target(model1, sleep_run(data1)),
+#'     tar_target(model2, sleep_run(data2)),
+#'     tar_target(model3, sleep_run(data3)),
+#'     tar_target(figure1, sleep_run(model1)),
+#'     tar_target(figure2, sleep_run(model2)),
+#'     tar_target(figure3, sleep_run(model3)),
+#'     tar_target(conclusions, sleep_run(c(figure1, figure2, figure3)))
+#'   )
+#' })
+#' px <- tar_make(callr_function = callr::r_bg)
+#' tar_watch(seconds = 10, outdated = FALSE, targets_only = TRUE)
 #' })
 #' }
-tar_watch <- function() {
+tar_watch <- function(
+  seconds = 5,
+  seconds_min = 1,
+  seconds_max = 100,
+  seconds_step = 1,
+  targets_only = FALSE,
+  outdated = TRUE,
+  label = NULL,
+  level_separation = 150,
+  height = "700px"
+) {
   assert_package("bs4Dash")
   assert_package("shiny")
   assert_package("shinycssloaders")
   assert_package("visNetwork")
   assert_target_script()
-  shiny::shinyApp(tar_watch_app_ui(), tar_watch_app_server)
+  assert_dbl(seconds, "seconds must be numeric.")
+  assert_dbl(seconds_min, "seconds_min must be numeric.")
+  assert_dbl(seconds_max, "seconds_max must be numeric.")
+  assert_dbl(seconds_step, "seconds_step must be numeric.")
+  assert_scalar(seconds, "seconds must have length 1.")
+  assert_scalar(seconds_min, "seconds_min must have length 1.")
+  assert_scalar(seconds_max, "seconds_max must have length 1.")
+  assert_scalar(seconds_step, "seconds_step must have length 1.")
+  seconds_min <- min(seconds_min, seconds)
+  seconds_max <- max(seconds_max, seconds)
+  seconds_step <- min(seconds_step, seconds_max)
+  ui <- tar_watch_app_ui(
+    seconds = seconds,
+    seconds_min = seconds_min,
+    seconds_max = seconds_max,
+    seconds_step = seconds_step,
+    targets_only = targets_only,
+    outdated = outdated,
+    label_tar_visnetwork = label,
+    level_separation = level_separation,
+    height = height
+  )
+  server <- function(input, output, session) {
+    tar_watch_server("tar_watch_id", height = height)
+  }
+  shiny::shinyApp(ui = ui, server = server)
 }
 
[…]
