### 2020-12-18 0364463
Allow tar_watch() to run the app in the background.
