+++ R/tar_watch.R
@@ -1,13 +1,23 @@
 # nocov start # Tested in tests/interactive/test-tar_watch.R
 #' @title Shiny app to watch the dependency graph.
 #' @export
-#' @description Calls [tar_visnetwork()] every few seconds.
+#' @description Launches a background process with a Shiny app
+#'   that calls [tar_visnetwork()] every few seconds.
+#'   To embed this app in other apps, use the Shiny module
+#'   in [tar_watch_ui()] and [tar_watch_server()].
 #' @details The controls of the app are in the left panel.
 #'   The `seconds` control is the number of seconds between
 #'   refreshes of the graph, and the other settings match
 #'   the arguments of [`tar_visnetwork()`].
-#' @return A Shiny app.
+#' @return A handle to `callr::r_bg()` background process running the app.
 #' @inheritParams tar_watch_ui
+#' @param background Logical, whether to run the app in a background process
+#'   so you can still use the R console while the app is running.
+#' @param host Character of length 1, IPv4 address to listen on.
+#'   Ignored if `background` is `FALSE`.
+#' @param port Positive integer of length 1, TCP port to listen on.
+#'   Ignored if `background` is `FALSE`.
+#' @param ... Named arguments to `callr::r_bg()`.
 #' @examples
 #' if (FALSE) { # Only run interactively.
 #' tar_dir({
@@ -29,8 +39,10 @@
 #'     tar_target(conclusions, sleep_run(c(figure1, figure2, figure3)))
 #'   )
 #' })
-#' px <- tar_make(callr_function = callr::r_bg)
+#' # Launch the app in a background process.
 #' tar_watch(seconds = 10, outdated = FALSE, targets_only = TRUE)
+#' # Run the pipeline.
+#' tar_make()
 #' })
 #' }
 tar_watch <- function(
@@ -42,7 +54,11 @@ tar_watch <- function(
   outdated = TRUE,
   label = NULL,
   level_separation = 150,
-  height = "700px"
+  height = "700px",
+  background = TRUE,
+  host = getOption("shiny.host", "127.0.0.1"),
+  port = getOption("shiny.port", targets::tar_watch_port()),
+  ...
 ) {
   assert_package("bs4Dash")
   assert_package("shiny")
@@ -60,6 +76,91 @@ tar_watch <- function(
   seconds_min <- min(seconds_min, seconds)
   seconds_max <- max(seconds_max, seconds)
   seconds_step <- min(seconds_step, seconds_max)
+  args <- list(
+    dir = getwd(),
+    seconds = seconds,
+    seconds_min = seconds_min,
+    seconds_max = seconds_max,
+    seconds_step = seconds_step,
+    targets_only = targets_only,
+    outdated = outdated,
+    label = label,
+    level_separation = level_separation,
+    height = height
+  )
+  call <- as.call(c(quote(targets::tar_watch_app), args))
+  if (!background) {
+    return(eval(call))
+  }
+  text <- deparse_safe(call, collapse = " ")
+  dir <- tempfile()
+  dir_create(dir)
+  app <- file.path(dir, "app.R")
+  writeLines(text, app)
+  args <- list(
+    appDir = dir,
+    host = host,
+    port = port,
+    launch.browser = TRUE,
+    quiet = FALSE
+  )
+  args <- list(dir = dir, port = port, host = host)
+  px <- callr::r_bg(func = tar_watch_process, args = args, ...)
+  utils::browseURL(paste0("http://", host, ":", port))
+  px
+}
+
+tar_watch_process <- function(dir, port, host) {
+  on.exit(unlink(dir, recursive = TRUE))
+  shiny::runApp(
+    appDir = dir,
+    host = host,
+    port = port,
+    launch.browser = FALSE,
[…]
