+++ R/tar_package.R
@@ -24,7 +24,7 @@
 #' @importFrom tibble as_tibble
 #' @importFrom tidyselect all_of any_of contains ends_with everything
 #'   last_col matches num_range one_of starts_with
-#' @importFrom utils globalVariables head menu stack tail
+#' @importFrom utils browseURL globalVariables head menu stack tail
 #' @importFrom vctrs vec_c vec_rbind vec_size vec_slice
 #' @importFrom withr local_dir local_envvar with_dir with_seed
 NULL
