### 2020-12-18 8023142
Make tar_watch() ping until the app is ready
