+++ R/tar_watch.R
@@ -11,13 +11,16 @@
 #'   the arguments of [`tar_visnetwork()`].
 #' @return A handle to `callr::r_bg()` background process running the app.
 #' @inheritParams tar_watch_ui
+#' @param label Label argument to [tar_visnetwork()].
 #' @param background Logical, whether to run the app in a background process
 #'   so you can still use the R console while the app is running.
 #' @param host Character of length 1, IPv4 address to listen on.
 #'   Ignored if `background` is `FALSE`.
 #' @param port Positive integer of length 1, TCP port to listen on.
 #'   Ignored if `background` is `FALSE`.
-#' @param ... Named arguments to `callr::r_bg()`.
+#' @param spinner show a command line spinner while `tar_watch()` is
+#'   waiting for the background app to initialize. Only applies to
+#'   `background = FALSE`.
 #' @examples
 #' if (FALSE) { # Only run interactively.
 #' tar_dir({
@@ -58,13 +61,11 @@ tar_watch <- function(
   background = TRUE,
   host = getOption("shiny.host", "127.0.0.1"),
   port = getOption("shiny.port", targets::tar_watch_port()),
-  ...
+  spinner = TRUE
 ) {
-  assert_package("bs4Dash")
-  assert_package("shiny")
-  assert_package("shinycssloaders")
-  assert_package("visNetwork")
-  assert_target_script()
+  pkgs <- c("bs4Dash", "pingr", "shiny", "shinycssloaders", "visNetwork")
+  msg <- paste("tar_watch() requires packages", paste(pkgs, collapse = ", "))
+  map(pkgs, ~assert_package(.x, msg = msg))
   assert_dbl(seconds, "seconds must be numeric.")
   assert_dbl(seconds_min, "seconds_min must be numeric.")
   assert_dbl(seconds_max, "seconds_max must be numeric.")
@@ -77,7 +78,6 @@ tar_watch <- function(
   seconds_max <- max(seconds_max, seconds)
   seconds_step <- min(seconds_step, seconds_max)
   args <- list(
-    dir = getwd(),
     seconds = seconds,
     seconds_min = seconds_min,
     seconds_max = seconds_max,
@@ -86,39 +86,30 @@ tar_watch <- function(
     outdated = outdated,
     label = label,
     level_separation = level_separation,
-    height = height
+    height = height,
+    host = host,
+    port = port
   )
-  call <- as.call(c(quote(targets::tar_watch_app), args))
   if (!background) {
-    return(eval(call))
+    return(do.call(tar_watch_app, args))
   }
-  text <- deparse_safe(call, collapse = " ")
-  dir <- tempfile()
-  dir_create(dir)
-  app <- file.path(dir, "app.R")
-  writeLines(text, app)
-  args <- list(
-    appDir = dir,
-    host = host,
-    port = port,
-    launch.browser = TRUE,
-    quiet = FALSE
+  spin <- cli::make_spinner()
+  trn(spinner, spin$spin(), NULL)
+  px <- callr::r_bg(
+    func = tar_watch_app,
+    args = args,
+    stdout = "|",
+    stderr = "|",
+    supervise = TRUE
   )
-  args <- list(dir = dir, port = port, host = host)
-  px <- callr::r_bg(func = tar_watch_process, args = args, ...)
-  utils::browseURL(paste0("http://", host, ":", port))
-  px
-}
-
-tar_watch_process <- function(dir, port, host) {
-  on.exit(unlink(dir, recursive = TRUE))
-  shiny::runApp(
-    appDir = dir,
-    host = host,
-    port = port,
-    launch.browser = FALSE,
-    quiet = FALSE
-  )
+  trn(spinner, spin$spin(), NULL)
[…]
