+++ man/tar_watch_app_ui.Rd
@@ -1,20 +1,19 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/tar_watch.R
-\name{tar_watch_app}
-\alias{tar_watch_app}
-\title{Shiny app to watch the dependency graph (process version).}
+\name{tar_watch_app_ui}
+\alias{tar_watch_app_ui}
+\title{Create the full \code{\link[=tar_watch]{tar_watch()}} app UI.}
 \usage{
-tar_watch_app(
-  dir = NULL,
-  seconds = 5,
-  seconds_min = 1,
-  seconds_max = 100,
-  seconds_step = 1,
-  targets_only = FALSE,
-  outdated = TRUE,
-  label = NULL,
-  level_separation = 150,
-  height = "700px"
+tar_watch_app_ui(
+  seconds,
+  seconds_min,
+  seconds_max,
+  seconds_step,
+  targets_only,
+  outdated,
+  label,
+  level_separation,
+  height
 )
 }
 \arguments{
@@ -41,7 +40,7 @@ takes a lot of time for large pipelines with lots of branches,
 and setting \code{outdated} to \code{FALSE} is a nice way to speed up the graph
 if you only want to see dependency relationships and build progress.}
 
-\item{label}{Label for the module.}
+\item{label}{Label argument to \code{\link[=tar_visnetwork]{tar_visnetwork()}}.}
 
 \item{level_separation}{Numeric of length 1,
 \code{levelSeparation} argument of \code{visNetwork::visHierarchicalLayout()}.
@@ -55,11 +54,11 @@ defaults to \code{150}.}
 height of the \code{visNetwork} widget.}
 }
 \value{
-A Shiny app.
+A Shiny UI.
 }
 \description{
-User should directly call \code{\link[=tar_watch]{tar_watch()}} instead
-of \code{tar_watch_app()}. The latter is an internal utility
-that blocks the main process.
+Only exported for infrastructure purposes.
+Not a user-side function. Users should instead
+call \code{\link[=tar_watch]{tar_watch()}} directly.
 }
 \keyword{internal}
