+++ NAMESPACE
@@ -256,7 +256,7 @@ export(tar_traceback)
 export(tar_validate)
 export(tar_visnetwork)
 export(tar_watch)
-export(tar_watch_app)
+export(tar_watch_app_ui)
 export(tar_watch_port)
 export(tar_watch_server)
 export(tar_watch_ui)
@@ -266,6 +266,7 @@ export(target_run_worker)
 importFrom(R6,R6Class)
 importFrom(callr,r)
 importFrom(cli,col_green)
+importFrom(cli,make_spinner)
 importFrom(cli,symbol)
 importFrom(codetools,findGlobals)
 importFrom(data.table,data.table)

[…]
