### 2020-12-18 4dd66ba
Finish cli spinner
