### 2020-12-18 c963b1a
Improve tar_watch() internals
