+++ man/tar_random_port.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tar_watch.R
-\name{tar_watch_port}
-\alias{tar_watch_port}
-\title{Random port for \code{\link[=tar_watch]{tar_watch()}}}
+% Please edit documentation in R/utils_url.R
+\name{tar_random_port}
+\alias{tar_random_port}
+\title{Random TCP port}
 \usage{
-tar_watch_port(lower = 49152L, upper = 65355L)
+tar_random_port(lower = 49152L, upper = 65355L)
 }
 \arguments{
 \item{lower}{Integer of length 1, lowest possible port.}
@@ -15,9 +15,10 @@ tar_watch_port(lower = 49152L, upper = 65355L)
 A random port not likely to be used by another process.
 }
 \description{
-Required for \code{\link[=tar_watch]{tar_watch()}}. Not a user-side function.
+Not a user-side function. Exported for infrastructure
+purposes only.
 }
 \examples{
-tar_watch_port()
+tar_random_port()
 }
 \keyword{internal}
