+++ NAMESPACE
@@ -243,6 +243,7 @@ export(tar_pattern)
 export(tar_pipeline)
 export(tar_progress)
 export(tar_prune)
+export(tar_random_port)
 export(tar_read)
 export(tar_read_raw)
 export(tar_renv)
@@ -257,7 +258,6 @@ export(tar_validate)
 export(tar_visnetwork)
 export(tar_watch)
 export(tar_watch_app_ui)
-export(tar_watch_port)
 export(tar_watch_server)
 export(tar_watch_ui)
 export(tar_workspace)
