### 2020-12-19 #401 @wlandau
I think academic peer review in general is already generous and altruistic, not to mention daunting. The holidays, COVID, and the volume of this submission compound all that so much. So I am extremely grateful that you signed up, and I will do whatever I can to recognize your efforts. December 29 is small as delays go, not a big deal on my end.
