### 2020-12-19 ba146f0
Add RStudio addin for tar_watch()
