+++ vignettes/need.Rmd
@@ -210,4 +210,4 @@ By contrast, the internal data structures in `targets` are dynamic by design, wh
 
 In [`drake`](https://github.com/ropensci/drake), pipelines are defined with the `drake_plan()` function. `drake_plan()` supports an elaborate domain specific language that diffuses user-supplied R expressions. This makes it convenient to assign commands to targets in the vast majority of cases, but it also obstructs custom metaprogramming by users ([example here](https://github.com/ropensci/drake/issues/1251)). Granted, it is possible to completely circumvent `drake_plan()` and create the whole data frame from scratch, but this is hardly ideal and seldom done in practice.
 
-The `targets` package tries to make customization easier. Relative to `drake`, `targets` takes a decentralized approach to setting up pipelines, moving as much custom configuration as possible to the target level rather than the whole pipeline level. In addition, the `tar_target_raw()` function avoids non-standard evaluation while mirroring `tar_target()` in all other respects. All this makes it much easier to create [custom metaprogrammed pipelines](https://wlandau.github.io/targets-manual/static.html#metaprogram) and [target archetypes](https://github.com/wlandau/tarchetypes) while avoiding an elaborate domain specific language for [static branching](https://books.ropensci.org/drake/static.html), which was extremely difficult to understand and error prone in [`drake`](https://github.com/ropensci/drake).
+The `targets` package tries to make customization easier. Relative to `drake`, `targets` takes a decentralized approach to setting up pipelines, moving as much custom configuration as possible to the target level rather than the whole pipeline level. In addition, the `tar_target_raw()` function avoids non-standard evaluation while mirroring `tar_target()` in all other respects. All this makes it much easier to create [custom metaprogrammed pipelines](https://wlandau.github.io/targets-manual/static.html#metaprogram) and [target archetypes](https://github.com/wlandau/tarchetypes) while avoiding an elaborate domain specific language for [static branching](https://books.ropensci.org/drake/static.html), which was extremely difficult to understand and error prone in [`drake`](https://github.com/ropensci/drake). The [targetopia](https://wlandau.github.io/targetopia.html) is an emerging ecosystem of workflow frameworks that take full advantage of this customization and democratize reproducible pipelines.
