### 2020-12-30 #401 @tjmahr
I expect to find time in the first week of the new year. I created an issue and Will resolved my biggest hang up with the package so far.
