### 2021-01-03 eb607fd
Accept lists of target objects
