+++ R/class_pipeline.R
@@ -28,6 +28,7 @@ pipeline_new <- function(
 
 pipeline_targets_init <- function(targets) {
   targets <- targets %||% list()
+  assert_target_list(targets)
   names <- map_chr(targets, ~.x$settings$name)
   assert_unique_targets(names)
   names(targets) <- names
@@ -286,6 +287,22 @@ pipeline_validate_lite.tar_pipeline <- function(pipeline) {
 #' @keywords internal
 pipeline_validate_lite.default <- function(pipeline) {
   throw_validate("not a tar_pipeline() object. _targets.R must end with one.")
+}
+
+as_pipeline <- function(x) {
+  UseMethod("as_pipeline")
+}
+
+#' @export
+#' @keywords internal
+as_pipeline.tar_pipeline <- function(x) {
+  x
+}
+
+#' @export
+#' @keywords internal
+as_pipeline.default <- function(x) {
+  pipeline_init(unlist(x, recursive = TRUE))
 }
 
 #' @export
