+++ R/utils_callr.R
@@ -22,7 +22,7 @@ callr_outer <- function(
 }
 
 callr_inner <- function(targets_script, targets_function, targets_arguments) {
-  targets_arguments$pipeline <- source(targets_script)$value
+  targets_arguments$pipeline <- as_pipeline(source(targets_script)$value)
   do.call(targets_function, targets_arguments)
 }
 
