+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(as_pipeline,default)
+S3method(as_pipeline,tar_pipeline)
 S3method(hash_object,"function")
 S3method(hash_object,character)
 S3method(hash_object,default)
