+++ NAMESPACE
@@ -196,6 +196,7 @@ S3method(value_validate,tar_group)
 S3method(value_validate,tar_value)
 export(all_of)
 export(any_of)
+export(as_pipeline)
 export(contains)
 export(ends_with)
 export(everything)
