+++ R/tar_group.R
@@ -35,7 +35,7 @@
 #' tar_dir({
 #' tar_script({
 #' library(dplyr)
-#' tar_pipeline(
+#' list(
 #'   tar_target(
 #'     data,
 #'     data.frame(
