+++ R/tar_traceback.R
@@ -19,7 +19,7 @@
 #' tmp <- sample(1)
 #' tar_script({
 #'   tar_option_set(error = "workspace")
-#'   tar_pipeline(
+#'   list(
 #'     tar_target(x, "loaded"),
 #'     tar_target(y, stop(x))
 #'   )
