+++ R/tar_validate.R
@@ -17,7 +17,7 @@
 #' @examples
 #' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
 #' tar_dir({
-#' tar_script(tar_pipeline(tar_target(x, 1 + 1)))
+#' tar_script(list(tar_target(x, 1 + 1)))
 #' tar_validate()
 #' })
 #' }
