+++ R/tar_prune.R
@@ -14,7 +14,7 @@
 #' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
 #' tar_dir({
 #' tar_script(
-#'   tar_pipeline(
+#'   list(
 #'     tar_target(y1, 1 + 1),
 #'     tar_target(y2, 1 + 1),
 #'     tar_target(z, y1 + y2)
@@ -22,7 +22,7 @@
 #' )
 #' tar_make()
 #' # Remove some targets from the pipeline.
-#' tar_script(tar_pipeline(tar_target(y1, 1 + 1)))
+#' tar_script(list(tar_target(y1, 1 + 1)))
 #' # Keep only the remaining targets in the data store.
 #' tar_prune()
 #' })
