+++ R/tar_sitrep.R
@@ -65,7 +65,7 @@
 #' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
 #' tar_dir({
 #' tar_script(
-#'   tar_pipeline(
+#'   list(
 #'     tar_target(x, seq_len(2)),
 #'     tar_target(y, 2 * x, pattern = map(x))
 #'   )
