+++ tests/testthat/test-class_future.R
@@ -1,5 +1,5 @@
 tar_test("future$workers", {
-  out <- future_init(tar_pipeline(), workers = 3L)
+  out <- future_init(list(), workers = 3L)
   expect_equal(out$workers, 3L)
 })
 
@@ -8,7 +8,7 @@ tar_test("workerless deployment works", {
   x <- tar_target_raw("x", quote(1L), deployment = "main")
   y <- tar_target_raw("y", quote(x), deployment = "main")
   z <- tar_target_raw("z", quote(x + 1L), deployment = "main")
-  pipeline <- tar_pipeline(list(x, y, z))
+  pipeline <- pipeline_init(list(x, y, z))
   future_init(pipeline)$run()
   expect_equal(target_read_value(x)$object, 1L)
   expect_equal(target_read_value(y)$object, 1L)
@@ -16,7 +16,7 @@ tar_test("workerless deployment works", {
   x <- tar_target_raw("x", quote(1L), deployment = "main")
   y <- tar_target_raw("y", quote(x), deployment = "main")
   z <- tar_target_raw("z", quote(x + 1L), deployment = "main")
-  pipeline <- tar_pipeline(list(x, y, z))
+  pipeline <- pipeline_init(list(x, y, z))
   out <- future_init(pipeline)
   out$run()
   built <- names(out$scheduler$progress$built$envir)
@@ -27,12 +27,12 @@ tar_test("some targets up to date, some not", {
   skip_if_not_installed("future")
   x <- tar_target_raw("x", quote(1L))
   y <- tar_target_raw("y", quote(x))
-  pipeline <- tar_pipeline(list(x, y))
+  pipeline <- pipeline_init(list(x, y))
   local <- local_init(pipeline)
   local$run()
   x <- tar_target_raw("x", quote(1L))
   y <- tar_target_raw("y", quote(x + 1L))
-  pipeline <- tar_pipeline(list(x, y))
+  pipeline <- pipeline_init(list(x, y))
   future <- future_init(pipeline)
   future$run()
   out <- names(future$scheduler$progress$built$envir)
@@ -47,7 +47,7 @@ tar_test("specialized plans", {
   resources <- list(plan = future::sequential)
   x <- tar_target_raw("x", quote(1L), resources = resources)
   y <- tar_target_raw("y", quote(x))
-  pipeline <- tar_pipeline(list(x, y))
+  pipeline <- pipeline_init(list(x, y))
   future <- future_init(pipeline)
   future$run()
   out <- sort(names(future$scheduler$progress$built$envir))
@@ -61,13 +61,13 @@ tar_test("future algo can skip targets", {
   skip_if_not_installed("future")
   x <- tar_target_raw("x", quote(1L))
   y <- tar_target_raw("y", quote(x))
-  pipeline <- tar_pipeline(list(x, y))
+  pipeline <- pipeline_init(list(x, y))
   local <- local_init(pipeline)
   local$run()
   unlink(file.path("_targets", "objects", "x"))
   x <- tar_target_raw("x", quote(1L))
   y <- tar_target_raw("y", quote(x))
-  pipeline <- tar_pipeline(list(x, y))
+  pipeline <- pipeline_init(list(x, y))
   future <- future_init(pipeline)
   future$run()
   out <- names(future$scheduler$progress$built$envir)
@@ -92,7 +92,7 @@ tar_test("nontrivial globals", {
     }
   }, envir = envir)
   x <- tar_target_raw("x", quote(f(1L)))
-  pipeline <- tar_pipeline(list(x))
+  pipeline <- pipeline_init(list(x))
   future <- future_init(pipeline)
   future$run()
   expect_equal(tar_read(x), 3L)
[…]
