+++ tests/testthat/test-tar_traceback.R
@@ -1,7 +1,7 @@
 tar_test("tar_traceback()", {
   tar_script({
     tar_option_set(error = "workspace")
-    tar_pipeline(tar_target(y, stop("3c47b24bd4a7ad8e5ce70f05eefe7c9c")))
+    list(tar_target(y, stop("3c47b24bd4a7ad8e5ce70f05eefe7c9c")))
   })
   try(tar_make(callr_function = NULL), silent = TRUE)
   out <- tar_traceback(y)
@@ -11,7 +11,7 @@ tar_test("tar_traceback()", {
 tar_test("tar_traceback()", {
   tar_script({
     tar_option_set(workspace = "z")
-    tar_pipeline(tar_target(z, 0))
+    list(tar_target(z, 0))
   })
   tar_make(callr_function = NULL)
   out <- tar_traceback(z)
