### 2021-01-03 8ce473f
Fix more transient bugs
