+++ tests/testthat/test-tar_bind.R
@@ -24,3 +24,12 @@ tar_test("tar_bind() errors if names are duplicated", {
   })
   expect_error(tar_make(callr_function = NULL), class = "condition_validate")
 })
+
+tar_test("valid pipeline with patterns only (#245)", {
+  expect_silent(out1 <- pipeline_init(tar_target(x, y, pattern = map(y))))
+  expect_silent(out2 <- pipeline_init(tar_target(z, w, pattern = map(w))))
+  expect_silent(out3 <- tar_bind(out1, out2))
+  expect_silent(pipeline_validate(out1))
+  expect_silent(pipeline_validate(out2))
+  expect_silent(pipeline_validate(out3))
+})
