### 2021-01-03 47291b9
Decouple tar_pipeline() and tar_bind() from rest of pkg
