+++ tests/testthat/test-utils_imports.R
@@ -111,11 +111,13 @@ tar_test("exclude special objects from imports", {
     tar_option_set(envir = envir)
     envir$regular_object <- "object"
     envir$target_object <- tar_target(x, regular_object)
-    envir$pipeline_object <- tar_pipeline(
-      envir$target_object,
-      tar_target(y, x)
+    envir$pipeline_object <- pipeline_init(
+      list(
+        envir$target_object,
+        tar_target(y, x)
+      )
     )
-    envir$pipeline_object
+    list(envir$target_object)
   })
   out <- tar_network(callr_function = NULL, targets_only = FALSE)
   expect_true(length(out$vertices$name) > 0L)
