+++ tests/performance/test-maps.R
@@ -3,7 +3,7 @@ library(proffer)
 # Should barely see target_load_deps() in the profiling data.
 # Use debug(store_read_path.tar_rds) to be sure readRDS() is not called. # nolint
 tar_script({
-  tar_pipeline(
+  list(
     tar_target(x, seq_len(1e3)),
     tar_target(y, x, pattern = map(x))
   )
@@ -17,7 +17,7 @@ px <- pprof(tar_outdated(callr_function = NULL))
 # Use debug(store_read_path.tar_rds) to be sure readRDS() is called. # nolint
 tar_destroy()
 tar_script({
-  tar_pipeline(
+  list(
     tar_target(x, seq_len(1e3), retrieval = "worker"),
     tar_target(y, x, pattern = map(x), retrieval = "worker")
   )
