+++ inst/_targets.R
@@ -11,12 +11,8 @@ summ <- function(dataset) {
 # Set target-specific options such as packages.
 tar_option_set(packages = "dplyr")
 
-# Define targets
-targets <- list(
+# End this file with a list of target objects.
+list(
   tar_target(data, data.frame(x = sample.int(100), y = sample.int(100))),
   tar_target(summary, summ(data)) # Call your custom functions as needed.
 )
-
-# End with a call to tar_pipeline() to wrangle the targets together.
-# This target script must return a pipeline object.
-tar_pipeline(targets)
