### 2021-01-03 1db5bc6
Add special condition for deprecations
