+++ R/tar_pipeline.R
@@ -11,13 +11,13 @@
 #' list( # You no longer need tar_pipeline() here.
 #'   tar_target(data_file, "data.csv", format = "file")
 #'   list( # Target lists can be arbitrarily nested.
-#'     tar_target(data_object, read.csv(data_file))
+#'     tar_target(data_object, read.csv(data_file)),
 #'     tar_target(analysis, analyze(data_object))
 #'   )
 #' )
 tar_pipeline <- function(...) {
   warn_deprecate(
-    "tar_pipelines() is deprecated. ",
+    "tar_pipeline() is deprecated. ",
     "Simply end your _targets.R file ",
     "with a list of tar_target() objects (arbitrarily nested)."
   )
