+++ man/tar_pipeline.Rd
@@ -23,7 +23,7 @@ library(targets)
 list( # You no longer need tar_pipeline() here.
   tar_target(data_file, "data.csv", format = "file")
   list( # Target lists can be arbitrarily nested.
-    tar_target(data_object, read.csv(data_file))
+    tar_target(data_object, read.csv(data_file)),
     tar_target(analysis, analyze(data_object))
   )
 )
