### 2021-01-03 a14debe
Fix more checks
