+++ R/tar_pipeline.R
@@ -9,7 +9,7 @@
 #' # In _targets.R:
 #' library(targets)
 #' list( # You no longer need tar_pipeline() here.
-#'   tar_target(data_file, "data.csv", format = "file")
+#'   tar_target(data_file, "data.csv", format = "file"),
 #'   list( # Target lists can be arbitrarily nested.
 #'     tar_target(data_object, read.csv(data_file)),
 #'     tar_target(analysis, analyze(data_object))
@@ -17,7 +17,7 @@
 #' )
 tar_pipeline <- function(...) {
   warn_deprecate(
-    "tar_pipeline() is deprecated. ",
+    "tar_pipeline() is deprecated in targets version >= 0.0.0.9004. ",
     "Simply end your _targets.R file ",
     "with a list of tar_target() objects (arbitrarily nested)."
   )
