+++ R/tar_bind.R
@@ -9,7 +9,7 @@
 #' # In _targets.R:
 #' library(targets)
 #' list( # You no longer need tar_pipeline() here.
-#'   tar_target(data_file, "data.csv", format = "file")
+#'   tar_target(data_file, "data.csv", format = "file"),
 #'   list( # Target lists can be arbitrarily nested.
 #'     tar_target(data_object, read.csv(data_file)),
 #'     tar_target(analysis, analyze(data_object))
@@ -17,8 +17,8 @@
 #' )
 tar_bind <- function(...) {
   warn_deprecate(
-    "tar_bind() and tar_pipeline() are deprecated. ",
-    "Simply end your _targets.R file ",
+    "tar_bind() and tar_pipeline() are deprecated ",
+    "in targets version >= 0.0.0.9004. Simply end your _targets.R file ",
     "with a list of tar_target() objects (arbitrarily nested)."
   )
   x <- unlist(list(...), recursive = TRUE)
