### 2021-01-03 f67c531
Update some tests
