+++ tests/hpc/test-future.R
@@ -33,8 +33,10 @@ tar_test("custom future plans through resources", {
     )
   })
   # The following should launch 4 targets running simultaneously.
-  # Terminate early if necessary.
   tar_make_future(workers = 4)
+  # After those 4 targets launch, terminate early and show progress.
+  # x should be built, and y and its 4 branches should be listed as running.
+  tar_progress()
 })
 
 test_that("packages are actually loaded", {
