+++ tests/hpc/test-retrieval.R
@@ -17,6 +17,7 @@ pipeline <- pipeline_init(
 cmq <- clustermq_init(pipeline)
 debug(cmq$run_worker)
 cmq$run()
+# The target object should be several MB.
 # pryr::object_size(target) # Run in the debugger. # nolint
 tar_destroy()
 tar_option_set(retrieval = "worker")
@@ -34,5 +35,6 @@ pipeline <- pipeline_init(
 cmq <- clustermq_init(pipeline)
 debug(cmq$run_worker)
 cmq$run()
+# The target object should be a lot smaller.
 # pryr::object_size(target) # Run in the debugger. # nolint
 tar_destroy()
