### 2021-01-04 #401 @limnoliver
Sorry for the delay! The review took a big longer than expected, but I really enjoyed doing it. Kudos to you @wlandau, for creating such a useful tool. For context, I'm coming at this review with experience using pipelines in R (using `remake` and a wrapper package to handle errors, shared caching, and what you're calling branching), but am almost completely naïve to `drake`.

Comments on documentation:
1. In the `Contributing` document, you reference a three-paragraph format, but don't give an example in section `Issues`.

2. I read a lot of your justification for why documentation is in certain places, and this all made sense to me (and overall I thought your documentation was accessible and thorough enough). However, I got tripped up in a few places, so here are some suggestions based on my experience. I don't think the `Statement of Need` vignette belongs in the package. A lot of this info is duplicated in the user manual, and it feels like a better fit in the manual because there is a lot of narrative relative to code/demonstrating functions within `targets`. Overall, I don't think the three vignettes in `targets` help someone use the functions within the package. "The composition of target objects" totally caught me off guard because I had read the entire manual and never saw mention of these target classes. While overall this vignette helps me to understand the package, it didn't help me use it. In the "orchestration and branching model", I found it hard to wrap my head around how or when I would interact with object classes without concrete examples preceding it. I think this is important info, but if I had to pick three vignettes for the package itself, I would maybe stick this in the manual and opt for more of a how-to vignette (or two or three). I think the `targets` package itself would benefit from a "Getting Started" vignette. While this info is elsewhere (through the tutorial or minimal example), I found the vignettes for the package itself to not be a good starting point for a new user like myself. There are hours and hours of documentation across your ecosystem (which is great!), but how can someone get started in under an hour? This should be the role of the vignettes (in my opinion!).

3. In the "skipping up-to-date targets" section of the "Data and metadata management" vignette, I think it's supposed to say "targets checks these rules in the order given _below_"

4. In section 2.2 of "The orchestration and branching model", in the example pipeline, it's unclear how "all analyses use the entirety of data1" because I don't see that target used in any other targets?

5. In `targets-manual`, chapter 7 introduces a pipeline in the opening section, but the code chunk does not contain a pipeline, only “tar_visnetwork()”. I think the pipeline that is supposed to appear is below in section 7.1.2?

6. It would be nice to have a section on "cleaning up", and best practices around when to use some of the cleaner functions. I came across old targets that no longer existed when I was experimenting, and I wasn't sure what the best approach was to cleaning up the pipeline.

7. In the manual in section `4.1 Internal Files`, it says the progress file is `_targets/meta_progress`, but in my experience this was `_targets/meta/progress`.

8. Some of the examples in the help files are more helpful than others. I appreciated examples like in `?tar_make` that actually can run a complete example. In cases like ?tar_cancel, because it's not a complete example (and `is_monday()` does not exist) we don't get to really see it in action, and it took me slightly longer to think about what how I should use it in my pipeline, and what the consequences of cancelling a target would be (I was pleasantly surprised to see that the pipeline skips it as if it is up-to-date...this is a really handy function!).

Comments: Packages passed `devtools::check()`, `devtools::test()`, `goodpractice::gp()`. `covr::package_coverage()` indicated 96% coverage, and I believe the author explained manual tests in the convo above. For `tarchetypes` there were 14 warnings about `testthat`:
```
Warning (test-tar_render_raw.R:44:1): (code run outside of `test_that()`)
The `code` argument to `test_that()` must be a braced expression to get accurate file-line information for failures.
    1. testthat::test_that(...) test-tar_render_raw.R:44:0
```
Despite the power and complexity of the package, overall I thought the user interface was great. It was easy to look at the list of functions, guess what they did, and get to a point of understanding what they did.

Estimated hours spent reviewing: 10+ (I would call it 10 critically thinking about the package/documentation, but spent much more time teaching myself `targets` in anticipation of using it.)

**testing package functions**
I wanted to test the debug functions. I built upon your example, and came across something I wasn't expecting. I often want to expose important choices/options in the function arguments so anyone else using the pipeline doesn't have to dig into the functions to find something. In my example, I expected the value of y to also be included in the workspace environment, but it wasn't:

```{r, eval = FALSE}
options(tidyverse.quiet = TRUE)
library(targets)
library(tidyverse)
options(crayon.enabled = FALSE)
tar_option_set(error = "workspace")
f <- function(x, y) {
  stopifnot(x < 4)
  print(y)
}
tar_pipeline(
  tar_target(x, seq_len(4)),
  tar_target(y, f(x, 'hello!'), pattern = map(x))
)
```
```{r, eval = FALSE}
# in console
tar_make()
#* run target x
#* run branch y_80b9a5d2
# "hello!"
#* run branch y_54f0a599
# "hello!"
#* run branch y_be82812d
# "hello!"
#* run branch y_dca2bb53
#x error branch y_dca2bb53
#* record workspace y_dca2bb53
```

```{r, eval=FALSE}
tar_workspace(y_dca2bb53)
print(x)
# [1] 4
print(y)
# Error in print(y) : object 'y' not found
```

Going back to inspect the manual, I figured out why this is happening (targets doesn't recognize y as a dependency, y is not in the global environment). I expected the workspace to function similar to browser(), so this threw me off when it wasn't a fully ready interactive environment. Maybe explicitly include more arguments in your top-level functions in the arguments to demonstrate the limitations there, and solutions/tradeoffs? In a bigger pipeline, it would be cumbersome to declare those as targets themselves, and I think the other option is to set those variables in the _targets.R script. Could you talk about how you handle those limitations? (Now after reviewing `tarchetypes`, I also realize the targets can get simpler with `tar_plan`.)

To learn the package and test the core functionality, I tried out a common use-case for us. Currently, we use remake and an
internal wrapper package ([example here](https://github.com/USGS-R/national-flow-observations)) that provides shared cache, error handling, and task splitting (similar to branching) services on top of remake. The ease of creating this example made me really, really excited about targets!

```{r use_case, eval=FALSE}
library(targets)
library(dataRetrieval)
library(MESS)

options(tidyverse.quiet = TRUE, crayon.enabled = FALSE)

# use case:
# want to retrieve data from the National Water Information System (NWIS)
# problem for big calls is that service times out or errors if
# too many records returned or too many sites called at once.
# Solution is to figure out what sites are available and how much data
# each site has, and download in batches of sites. For a national pull, this can
# take a day or two, so a good use case for pipelining.
query_nwis <- function(states, parameters) {
  sites <- dataRetrieval::whatNWISdata(stateCd = states, parameterCd = parameters, service = 'dv')
  return(sites)
}

group_sites <- function(sites, max_results, max_sites) {
  out <- sites %>%
    filter(stat_cd %in% "00003") %>%
    select(site_no, count_nu) %>%
    distinct() %>%
    arrange(desc(count_nu)) %>%
    mutate(task_num = MESS::cumsumbinning(x = count_nu,
                                          threshold = max_results,
                                          maxgroupsize = max_sites)) %>%
    group_by(task_num) %>%
    targets::tar_group()
}
pull_data <- function(sites, parameters) {
  dat <- readNWISdv(siteNumbers = sites$site_no, parameterCd = parameters)
  return(dat)
}

tar_pipeline(
  tar_target(states, c('RI', 'DE')), # NWIS accepts two-letter state abbreviations
  tar_target(parameters, '00010'), # parameter is USGS code for temperature
  tar_target(sites, query_nwis(states, parameters), map(states)),
  tar_target(site_batches, group_sites(sites, max_results = 50000, max_sites = 50), iteration = 'group'),
  tar_target(dat, pull_data(site_batches, parameters), pattern = map(site_batches))
)
```
One surprise digging into my example above was that `tar_path(dat)` returns `_targets/objects/dat`, despite that file not existing, and despite the path in the metadata being `NA`. Should this return `NA`, or the children names, or the filepath with a wildcard for clarity? I do appreciate that you can `tar_read(dat)` or `tar_load(dat)` with helpers behind the scenes.

A common use-case I have is once we get to doing site-level things (plotting, modeling), I'll notice a problem and want to
see the original data that was pulled from the web service. Or, the combined target is huge, and I just want to look at the (much smaller) branch that containes a particular site.  How do I know which branch file site "01115120" is in, for example? Because in the code above, my batches are groups of sites, I want a dataframe of site IDs with corresponding filenames of which branches they belong to. I was expecting that crosswalk information to exist in `_targets/meta/meta`. The manual spends a lot of time showing how you can check the pattern created, what order the branches are created, and tar_pattern gets close. I wonder if a function like tar_pattern could accept the name of a target with `type == pattern`, and it produces a dataframe that shows the existing filenames and corresponding values of the variables that were mapped over? Admittedly, it wasn't too terrible to come up with what I wanted (see below), so at a minimum maybe include an example somewhere in the manual when discussing patterns and branching for those that want a post hoc crosswalk?

```{r crosswalk, eval = FALSE}
file_names <- filter(tar_meta(), name %in% 'dat') %>%
  pull(children) %>% unlist()

crosswalk <- tar_read(site_batches) %>%
  left_join(tibble(file_names,
                   tar_group = seq(length(file_names))))

# check if my assumptions were true!
crosswalk$file_names[1]
# [1] "dat_abef3008"

actual <- tar_read("dat_abef3008") %>%
  select(site_no) %>% distinct() %>% pull(site_no)
predicted <- filter(crosswalk, file_names %in% "dat_abef3008") %>%
  pull(site_no)

all.equal(sort(actual), sort(predicted))
# [1] TRUE
```
