### 2021-01-05 0ebefec
Remove mention of 3-paragraph issue format
https://github.com/ropensci/software-review/issues/401
