### 2021-01-05 f1938c6
Streamline design documents
