+++ vignettes/need.Rmd
@@ -1,213 +1,0 @@
----
-title: "Statement of need"
-output:
-  rmarkdown::html_vignette:
-    toc: true
-    number_sections: true
-vignette: >
-  %\VignetteIndexEntry{Statement of need}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r setup, include = FALSE}
-library(dplyr)
-library(targets)
-```
-
-# Role
-
-[Make](https://www.gnu.org/software/make/)-like [pipeline toolkits](https://github.com/pditommaso/awesome-pipeline) enhance the reproducibility, efficiency, and hygiene of large data science projects. Pipelines orchestrate tasks such as data munging, model fitting, summaries, and R Markdown reports, while skipping tasks that are already up to date with the underlying code and upstream data. With pipelines, we save time, and we have evidence to inform which results we can trust. `targets` is such a pipeline tool, and its major benefits are as follows.
-
-Benefit | Description
----|---|---
-Efficiency | Orchestrates the pipeline, running computational tasks in parallel when possible and skipping tasks that are already up to date.
-Reproducibility | Detects when all targets are already up to date, which provides tangible evidence that results match the underlying code and data.
-Maintainability | Nudges users toward a clean function-oriented programming style, and it abstracts output data as R objects.
-Portability | Saves only one file per target in the data store and gives files informative names to allow users to more easily understand their data and collaborate on projects.
-Domain specificity | Designed to work entirely within the R language.
-
-Myriad workflow tools already exist, but most of them have different goals and behaviors. Different tools suit different situations.
-
-Tool | Issues and differences
----|---
-[pditmmaso/awesome-pipeline](https://github.com/pditommaso/awesome-pipeline) | Most tools listed are not for R.
-[Metaflow](https://github.com/netflix/metaflow) | Resembles [Airflow](https://github.com/apache/airflow/) more than [GNU Make](https://www.gnu.org/software/make/). Features include versioned runs and seamless cloud integration, but not automated decision-making about which steps to skip.
-[GNU Make](https://www.gnu.org/software/make/) | Not designed for a function-oriented style in R.
-[`remake`](https://github.com/richfitz/remake) | No longer maintained.
-[`drake`](https://github.com/ropensci/drake) | Permanent systemic limitations on portability, reproducibility, parallel efficiency, dynamic branching, user-side guardrails, and metaprogramming.
-[`memoise`](https://github.com/r-lib/memoise) | Operates on individual functions, not on a pipeline as a whole.
-[`knitr`](https://yihui.name/knitr/), [`rmarkdown`](https://github.com/rstudio/rmarkdown), [`workflowr`](https://github.com/jdblischak/workflowr), and related | Designed for literate programming and communication. Not designed to scale in the same ways as pipeline tools, but useful for small components of pipelines.
-[Git](https://git-scm.com/) | Designed for version control, not pipeline management specifically, but recommended for use with pipelines. 
-[`renv`](https://github.com/rstudio/renv), [Docker](https://www.docker.com/), and related | Designed for package environment management and containerization. Not pipeline management specifically, but recommended for use with pipelines.
-
-# drake
-
-The [`drake`](https://github.com/ropensci/drake) package is an older and more established R-focused pipeline toolkit. It is has become a key piece of the R ecosystem, and development and support will continue. Future efforts will focus on the issues that bring the most concrete value to users: requested features, performance improvements for known inefficiencies, bug fixes, documentation, education, and prompt one-on-one help. Existing `drake`-powered projects can safely continue to use `drake`, and there is no urgency to migrate to `targets`.
-
[…]
