+++ vignettes/orchestration.Rmd
@@ -1,100 +1,18 @@
 ---
-title: "The orchestration and branching model"
+title: "Major internal classes"
 output:
   rmarkdown::html_vignette:
     toc: true
     number_sections: true
 vignette: >
-  %\VignetteIndexEntry{The orchestration and branching model}
+  %\VignetteIndexEntry{Major internal classes}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
 
 The `targets` package runs the correct targets in the correct order (orchestration) and creates new targets dynamically at runtime (branching). This vignette describes the underlying package structure and mental model that gives `targets` its flexibility and parallel efficiency.
 
-# Structure
 
-`targets` relies heavily on object-oriented programming, and the following subsections describe the hierarchy and composition of the objects relevant to orchestration and branching. Objects that appear once per workflow are instances of `R6` classes. To maximize performance, target-specific data structures are lower-tech classes: simple environments with formal constructors, helpers, and validators.
-
-## Algorithm class
-
-An algorithm in `targets` is an abstract class that represents how to iterate through the pipeline target by target. Different algorithms describe different kinds of deployment: for example, execution in the main process on the host machine versus parallel execution on a cluster. Every algorithm has a scheduler object and a pipeline object.
-
-## Pipeline class
-
-A pipeline is a wrapper around a collection of targets, and it is responsible for the initial reasoning about the topology of the pipeline before runtime. Pipelines express their reasoning by producing static graphs and scheduler objects early on. In addition, pipelines contain new buds and branches created dynamically during the pipeline.
-
-## Scheduler class
-
-Whereas pipelines are responsible for *static* topology, schedulers are responsible for *dynamic* topology. Schedulers know
-
-1. The upstream and downstream neighbors of each target.
-1. The progress of each target, e.g. queued, running, or finished.
-1. How many upstream dependencies need to be checked or built before a target is ready to run.
-
-To meet these responsibilities, the scheduler is composed of three smaller objects:
-
-1. A graph object.
-1. A progress object.
-1. A priority queue.
-
-## Graph class
-
-The graph class keeps track of the upstream and downstream neighbors of each target. The scheduler adds edges to the graph when new targets are created dynamically.
-
-The graph is implemented as two adjacency lists: one for upstream edges and another for downstream edges. For the purposes of powering a pipeline, we find this low-tech structure to be more efficient than `igraph` in our situation where we repeatedly query the graph and the number of nodes is small. (Transient `igraph` objects, however, are created for validation and visualization purposes.)
-
-## Progress class
-
-The progress class keeps track of the state of each target: queued, running, skipped, built, outdated, canceled, or errored. To accomplish this, the progress object maintains a counter object for each category.
-
-## Queue class
-
-The queue class is a priority queue, essentially a wrapper around a named integer vector of ranks. For `targets`' purposes, the rank of a target is the number of unmet dependencies so far, minus a per-target priority value in the interval `[0, 1)` to control the order in which targets are dequeued. As the pipeline progresses, the queue is checked and modified periodically as dependencies are met. The next target to build is the lowest rank target such that `-1L < rank <= 0L`. Branches are pushed to the queue when they are dynamically created.
-
-## Counter class
-
-A counter is an efficient abstraction for keeping track of target membership in a category. A counter stores the number of targets in the category and a hash table with the names of those targets. Counters are used to efficiently keep track of runtime progress (e.g. running, queued, or built) as well membership in the queue.
-
-## Target class
-
[…]
