+++ vignettes/data.Rmd
@@ -1,106 +1,0 @@
----
-title: "Data and metadata management"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Data and metadata management}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-# A design document for developers
-
-This article is for developers who are designing and implementing the internal architecture of [`targets`](https://github.com/wlandau/targets). For help using the package, please start with the [walkthrough chapter of the manual](https://wlandau.github.io/targets-manual/walkthrough.html). The [reference website](https://wlandau.github.io/targets/) links to other useful resources such as recorded talks and example projects.
-
-# About
-
-Like its predecessor, `drake`, the `targets` package
-
-1. Abstracts files as R objects.
-2. Records the real-time progress of targets as they are running.
-3. Records special metadata in order to skip targets that are already up to date.
-
-Unlike `drake`, which outsources data and metadata management to an external package, `targets` has an entirely custom internal data system. `targets` goes out of its way to reduce the number of files in storage, centralize the progress data and metadata, assign informative file names, and expose the file system to the user. This approach increases efficiency and portability, and it helps users understand and take control of their data.
-
-# File system
-
-When a `targets` pipeline runs, it creates a folder called `_targets` to store all the files it needs.
-
-```
-_targets/
-├── meta/
-├────── progress
-├────── meta
-├── objects/
-├────── target1 
-├────── target2
-├────── branching_target_c7bcb4bd
-├────── branching_target_285fb6a9
-├────── branching_target_874ca381
-└── scratch/ # Temporary files deleted at the end of tar_make().
-```
-
-The number of files equals the number of targets plus two, which makes projects easier to upload and share among collaborators than `drake`'s `.drake/` cache. (Files in `_targets/scratch/` do not count because they can all be safely deleted after `tar_make()`.) However, these files may still be too large and too numerous for code-specific version control systems like Git. For such projects, it may be more appropriate to share caches through external version-aware platforms such as Dropbox, Microsoft OneDrive, and Google Docs.
-
-# Targets
-
-With the exception of dynamic files, the return value of each target lives in its own file inside `_targets/objects/`. The file name is the name of the target, and there is no file extension. The metadata keeps track of the storage format that governs how to read and write the target's data. The default format is RDS, so if target `x` has no explicit format, then `readRDS("_targets/objects/x")` will read the data. (However, we state this just for the sake of understanding. The recommended way to read data is `tar_read()`, which takes the storage format into account.)
-
-# Progress
-
-The file `_targets/meta/progress` is a pipe-separated flat file with the name of each target and it's current runtime progress (running, built, cancelled, or errored). The information in this file helps users keep track of what the pipeline is doing at a given moment. `targets` periodically appends rows to `_targets/meta/progress` as the pipeline progresses, so duplicated names usually appear. For any target with duplicated rows in `_targets/meta/progress`, only the lowest row is valid.
-
-In most situations, the progress file can be safely excluded from version control. Functions like `tar_graph()` use progress information, but it is not essential to the reproducible end product.
-
-# Metadata
-
-`targets` uses special metadata to decide which targets are up to date and which need to run. The metadata file `_targets/meta/meta` is a flat file with one row for every target and every global object relevant to the pipeline. `targets` appends new rows to this file as the pipeline progresses. Unlike `drake`, the metadata is centralized and compatible with `data.table`, which makes it far faster to check which targets are up to date. In addition, the metadata system allows `targets` to check not only for up-to-date targets, but also up-to-date global objects, which makes it easier for the user to understand *why* a target is outdated.
-
[…]
