+++ vignettes/start.Rmd
@@ -1,0 +1,14 @@
+---
+title: "How to get started with targets"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{How to get started with targets}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+The easiest way to get started with [`targets`](https://github.com/wlandau/targets) is to watch one of the [recorded talks](https://wlandau.github.io/targets/#recorded-talks) linked from the [reference website](https://wlandau.github.io/targets/). Next, read the [walkthrough chapter](https://wlandau.github.io/targets-manual/walkthrough.html) of the [use manual](https://wlandau.github.io/targets-manual/), a brief tutorial of the most essential features of [`targets`](https://github.com/wlandau/targets) using an easy example. Then, skim the next chapters of the [user manual](https://wlandau.github.io/targets-manual/) just to get a sense of the overall best practices and capabilities. You may use these later chapters for reference later on.
+
+Next, it is helpful to get some hands-on practice using the package. The [`targets-tutorial`](https://github.com/wlandau/targets-tutorial) repository contains a short course with step-by-step hands-on guided practice to gradually build up some basic skills. Follow [these instructions](https://github.com/wlandau/targets-tutorial) to access the exercises in a free [RStudio Cloud](http://rstudio.cloud) workspace (no need to install anything locally).
+
+After that, you may be ready to try out [`targets`](https://github.com/wlandau/targets) in a fully-fledged example project. The [reference website](http://wlandau.github.io/targets/) links to [several example projects](https://wlandau.github.io/targets/#examples), each with its own free [RStudio Cloud](http://rstudio.cloud) workspace. These are great places to experience [`targets`](https://github.com/wlandau/targets) projects as they exist in the real world.
