+++ R/tar_name.R
@@ -11,7 +11,13 @@
 #' @examples
 #' tar_name()
 #' tar_name(default = "custom_target_name")
-#' tar_target(name, tar_name(default = "custom")) # Returns "name".
+#' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
+#' tar_dir({
+#' tar_script(tar_target(x, tar_name()))
+#' tar_make()
+#' tar_read(x)
+#' })
+#' }
 tar_name <- function(default = "target") {
   assert_chr(default)
   assert_scalar(default)
