+++ R/tar_make_clustermq.R
@@ -28,7 +28,7 @@
 #' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
 #' tar_dir({
 #' tar_script({
-#'   options(clustermq.scheduler = "multicore")
+#'   options(clustermq.scheduler = "multicore") # Does not work on Windows.
 #'   tar_option_set()
 #'   list(tar_target(x, 1 + 1))
 #' })
