+++ R/tar_option_get.R
@@ -19,6 +19,16 @@
 #' tar_target(x, 1)$settings$format
 #' tar_option_reset() # reset the format
 #' tar_target(x, 1)$settings$format
+#' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
+#' tar_dir({
+#' tar_script({
+#'   tar_option_set(cue = tar_cue(mode = "always")) # All targets always run.
+#'   list(tar_target(x, 1), tar_target(y, 2))
+#' })
+#' tar_make()
+#' tar_make()
+#' })
+#' }
 tar_option_get <- function(option) {
   option <- match.arg(option, choices = names(formals(tar_option_set)))
   tar_envir_options[[option]] %||% tar_option_default(option)
