### 2021-01-05 b7b9981
Enhance exmaples
