+++ R/tar_path.R
@@ -20,7 +20,13 @@
 #' @examples
 #' tar_path()
 #' tar_path(your_target)
-#' tar_target(returns_path, tar_path())
+#' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
+#' tar_dir({
+#' tar_script(tar_target(returns_path, tar_path()))
+#' tar_make()
+#' tar_read(returns_path)
+#' })
+#' }
 tar_path <- function(name = NULL, default = tempfile()) {
   name <- deparse_language(substitute(name))
   assert_chr(name %||% character(0), "name arg of tar_path() must be a symbol")
