+++ R/tar_target.R
@@ -179,6 +179,14 @@
 #' data <- tar_target(target_name, get_data(!!n_rows), tidy_eval = FALSE)
 #' print(data)
 #' tar_option_reset()
+#' # In a pipeline:
+#' if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
+#' tar_dir({
+#' tar_script(tar_target(x, 1 + 1))
+#' tar_make()
+#' tar_read(x)
+#' })
+#' }
 tar_target <- function(
   name,
   command,
