+++ man/tar_path.Rd
@@ -35,5 +35,11 @@ the return value is \code{default}.
 \examples{
 tar_path()
 tar_path(your_target)
-tar_target(returns_path, tar_path())
+if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
+tar_dir({
+tar_script(tar_target(returns_path, tar_path()))
+tar_make()
+tar_read(returns_path)
+})
+}
 }
