+++ man/tar_target_raw.Rd
@@ -238,4 +238,11 @@ of targets programmatically (metaprogramming, static branching).
   })
   print(target_list[[1]])
   print(target_list[[2]])
+if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
+tar_dir({
+tar_script(tar_target_raw("x", quote(1 + 1)))
+tar_make()
+tar_read(x)
+})
+}
 }
