+++ man/tar_target.Rd
@@ -231,4 +231,12 @@ print(data)
 data <- tar_target(target_name, get_data(!!n_rows), tidy_eval = FALSE)
 print(data)
 tar_option_reset()
+# In a pipeline:
+if (identical(Sys.getenv("TARGETS_LONG_EXAMPLES"), "true")) {
+tar_dir({
+tar_script(tar_target(x, 1 + 1))
+tar_make()
+tar_read(x)
+})
+}
 }
