### 2021-01-05 2d72e3c
Clarify role of tar_path()
