+++ tests/testthat/test-tar_path.R
@@ -1,5 +1,6 @@
 tar_test("tar_path() outside a pipeline with no arguments", {
   expect_true(is.character(tar_path()))
+  expect_true(is.na(tar_path()))
   expect_equal(length(tar_path()), 1L)
   expect_equal(tar_path(default = "x"), "x")
 })
