### 2021-01-05 b3eaf49
Store the seeds of patterns
