+++ NAMESPACE
@@ -214,6 +214,7 @@ export(rstudio_addin_tar_target)
 export(rstudio_addin_tar_visnetwork)
 export(starts_with)
 export(tar_bind)
+export(tar_branches)
 export(tar_cancel)
 export(tar_cue)
 export(tar_deduplicate)
