### 2021-01-05 7cd5309
Implement and test tar_branches()
