+++ vignettes/start.Rmd
@@ -7,8 +7,8 @@ vignette: >
   %\VignetteEncoding{UTF-8}
 ---
 
-The easiest way to get started with [`targets`](https://github.com/wlandau/targets) is to watch one of the [recorded talks](https://wlandau.github.io/targets/#recorded-talks) linked from the [reference website](https://wlandau.github.io/targets/). Next, read the [walkthrough chapter](https://wlandau.github.io/targets-manual/walkthrough.html) of the [use manual](https://wlandau.github.io/targets-manual/), a brief tutorial of the most essential features of [`targets`](https://github.com/wlandau/targets) using an easy example. Then, skim the next chapters of the [user manual](https://wlandau.github.io/targets-manual/) just to get a sense of the overall best practices and capabilities. You may use these later chapters for reference later on.
+The easiest way to get started with [`targets`](https://github.com/wlandau/targets) is to watch one of the [recorded talks](https://wlandau.github.io/targets/#recorded-talks) linked from the [reference website](https://wlandau.github.io/targets/). Next, read the [walkthrough chapter](https://wlandau.github.io/targets-manual/walkthrough.html) of the [use manual](https://wlandau.github.io/targets-manual/), a brief tutorial of the most essential features of [`targets`](https://github.com/wlandau/targets) using an easy example. Afterwards, it would help to skim the next chapters of the [user manual](https://wlandau.github.io/targets-manual/) just to get a sense of the overall best practices and capabilities. You will likely use these later chapters for reference later on.
 
-Next, it is helpful to get some hands-on practice using the package. The [targets-tutorial](https://github.com/wlandau/targets-tutorial) repository contains a short course with step-by-step hands-on guided practice to gradually build up some basic skills. Follow [these instructions](https://github.com/wlandau/targets-tutorial) to access the exercises in a free [RStudio Cloud](http://rstudio.cloud) workspace (no need to install anything locally). Do not worry about completely finishing the course. Even if you only get through the first three chapters, you will come away understanding of how the package works. When this tutorial is presented at external conferences, it almost never covers everything. 
+If [`targets`](https://github.com/wlandau/targets) looks useful, the next step is hands-on practice. The [targets-tutorial](https://github.com/wlandau/targets-tutorial) repository contains a short course with step-by-step guided practice. Follow [these instructions](https://github.com/wlandau/targets-tutorial) to access the exercises in a free [RStudio Cloud](http://rstudio.cloud) workspace (no need to install anything locally). Do not worry about completely finishing the course. The first three chapters cover most of the essentials.
 
-After that, you may be ready to try out [`targets`](https://github.com/wlandau/targets) in a fully-fledged example project. The [reference website](http://wlandau.github.io/targets/) links to [several example projects](https://wlandau.github.io/targets/#examples), each with its own free [RStudio Cloud](http://rstudio.cloud) workspace. These are great places to experience [`targets`](https://github.com/wlandau/targets) projects as they exist in the real world.
+After that, you may be ready to try out a more realistic example project. The [reference website](http://wlandau.github.io/targets/) links to [several options](https://wlandau.github.io/targets/#examples), each with its own free [RStudio Cloud](http://rstudio.cloud) workspace.
