### 2021-01-05 #401 @wlandau
# Reply to @limnoliver's review

Thank you for your thorough and insightful review. Your feedback is super impactful, and I incorporated everything I could into the packages and documentation.

> 1. In the Contributing document, you reference a three-paragraph format, but don't give an example in section Issues.

I actually ended up changing my mind about that. All mentions of that three-paragraph format should now be gone.

> 2. I don't think the Statement of Need vignette belongs in the package. A lot of this info is duplicated in the user manual, and it feels like a better fit in the manual because there is a lot of narrative relative to code/demonstrating functions within targets.

Now that you mention it, I completely agree. I moved most of the statement of need into a [new chapter on `drake`](https://wlandau.github.io/targets-manual/drake.html): the future direction of `drake`, how to transition to `targets`, and technical details about the advantages of `targets`. Together, I think that chapter and the intro chapter sufficiently motivate the package, and the shift deduplicates a lot of text. Really glad to have done this.

> Overall, I don't think the three vignettes in targets help someone use the functions within the package. "The composition of target objects" totally caught me off guard because I had read the entire manual and never saw mention of these target classes. While overall this vignette helps me to understand the package, it didn't help me use it. In the "orchestration and branching model", I found it hard to wrap my head around how or when I would interact with object classes without concrete examples preceding it. I think this is important info, but if I had to pick three vignettes for the package itself, I would maybe stick this in the manual and opt for more of a how-to vignette (or two or three). I think the targets package itself would benefit from a "Getting Started" vignette. While this info is elsewhere (through the tutorial or minimal example), I found the vignettes for the package itself to not be a good starting point for a new user like myself. There are hours and hours of documentation across your ecosystem (which is great!), but how can someone get started in under an hour? This should be the role of the vignettes (in my opinion!).

Sorry, it must have been frustrating to expect friendly user-side documentation in the vignettes and find esoteric developer-side documentation instead. I just moved all those developer vignettes to a new `bookdown` website at <https://wlandau.github.io/targets-design> and added a short new introductory vignette at `targets/vignettes/start.Rmd`. The vignette is pretty high-level, it describes the order in which to walk through resources and what to expect from each.

Beyond that, I am concerned about adding more vignettes to the `targets` package itself. In addition to the potential tech debt I described before, it seems like it might create some ambiguity with the manual. (Which vignettes belong with `targets`, which belong with `targets-manual`, and how will users know when to use which resource?). In addition, <https://github.com/wlandau/targets-tutorial> is its own supply of introductory vignettes, and these come with an RStudio Cloud workspace to help users try out the code. Lastly, `drake` users told me they sometimes struggled with the volume of reading material (and documentation in general) and I would like to be conscious of that this time around.

> 3. In the "skipping up-to-date targets" section of the "Data and metadata management" vignette, I think it's supposed to say "targets checks these rules in the order given below"

Should be fixed now in https://github.com/wlandau/targets-design/commit/50d428096af1e9d149be36718ae6c0e28d7c3409.

> 4. In section 2.2 of "The orchestration and branching model", in the example pipeline, it's unclear how "all analyses use the entirety of data1" because I don't see that target used in any other targets?

Yup, that's a typo in the pipeline code. I meant to use `data1` in the `analysis` target. Fixed in https://github.com/wlandau/targets-design/commit/52b74b7b092123988402fcf347a9fdf23144a5ac.

> 5. In targets-manual, chapter 7 introduces a pipeline in the opening section, but the code chunk does not contain a pipeline, only “tar_visnetwork()”. I think the pipeline that is supposed to appear is below in section 7.1.2?

Yes, the pipeline and graph were supposed to appear. Should be fixed now.

> 6. It would be nice to have a section on "cleaning up", and best practices around when to use some of the cleaner functions. I came across old targets that no longer existed when I was experimenting, and I wasn't sure what the best approach was to cleaning up the pipeline.

I just added a new section at the end of the chapter on best practices: https://github.com/wlandau/targets-manual/commit/7ae13345be107ff21c64e7f5e33b0a2e5aa8c9cb. It describes all the [cleaning functions](https://wlandau.github.io/targets/reference/index.html#section-clean) except [`tar_deduplicate()`](https://wlandau.github.io/targets/reference/tar_deduplicate.html). (`tar_make()` etc. deduplicate the metadata automatically, so most users should not need to invoke it manually.)

> 7. In the manual in section 4.1 Internal Files, it says the progress file is _targets/meta_progress, but in my experience this was _targets/meta/progress.

Definitely a typo, fixed in https://github.com/wlandau/targets-manual/commit/f481e60ded9c82c5592cc6775a35c1eb7873fae9.

> 8. Some of the examples in the help files are more helpful than others. I appreciated examples like in ?tar_make that actually can run a complete example. In cases like ?tar_cancel, because it's not a complete example (and is_monday() does not exist) we don't get to really see it in action, and it took me slightly longer to think about what how I should use it in my pipeline, and what the consequences of cancelling a target would be (I was pleasantly surprised to see that the pipeline skips it as if it is up-to-date...this is a really handy function!).

I went back and wrote runnable examples for the functions that lacked them:

* `tar_cancel()`
* `tar_envir()`
* `tar_name()`
* `tar_option_get()`
* `tar_option_reset()`
* `tar_option_get()`
* `tar_path()`
* `tar_seed()`
* `tar_target()`
* `tar_target_raw()`

> Comments: Packages passed devtools::check(), devtools::test(), goodpractice::gp(). covr::package_coverage() indicated 96% coverage, and I believe the author explained manual tests in the convo above.

That's odd. I meant to write explicit `#nocov` comments in all the places I gave up on automated coverage, and the [online coverage reports](https://codecov.io/gh/wlandau/targets) show 100%. What do you see when you run `devtools::test()`? Are any tests skipped because `Suggests:` packages are missing?

> For tarchetypes there were 14 warnings about testthat:

Yes, I reproduced that with `testthat` version 3.0.1. It was because I was writing tests that look like `test_that("", suppressMessages(...))`. I recently updated `tarchetypes` to begin the test code with braces, and the warnings should be gone as of https://github.com/wlandau/tarchetypes/commit/a461e1ff4ed243b547e8cb66f782e53a7d8c6503.

> Estimated hours spent reviewing: 10+ (I would call it 10 critically thinking about the package/documentation, but spent much more time teaching myself targets in anticipation of using it.)

Knowing what you know now, should I have opened a different submission for each of `targets`, `tarchetypes`, and the manual? Do `tarchetypes` and the manual each make sense to review in isolation? How much time would that have saved you? The editors tried to warn me about the workload I would be inflicting on reviewers, and if I judged poorly, maybe this could help future submissions.

> I wanted to test the debug functions. I built upon your example, and came across something I wasn't expecting. I often want to expose important choices/options in the function arguments so anyone else using the pipeline doesn't have to dig into the functions to find something. In my example, I expected the value of y to also be included in the workspace environment, but it wasn't: ... Going back to inspect the manual, I figured out why this is happening (targets doesn't recognize y as a dependency, y is not in the global environment). I expected the workspace to function similar to browser(), so this threw me off when it wasn't a fully ready interactive environment.

I updated the [debugging section](https://wlandau.github.io/targets-manual/practice.html#debugging) to try to clear up this ambiguity.

> Maybe explicitly include more arguments in your top-level functions in the arguments to demonstrate the limitations there, and solutions/tradeoffs?

Yeah, I now see that it might have been confusing to have target named `x` and function `f()` whose only argument was also named `x`. Now, `f()` has multiple arguments, none of which is named `x`.

> In a bigger pipeline, it would be cumbersome to declare those as targets themselves, and I think the other option is to set those variables in the _targets.R script. Could you talk about how you handle those limitations?

Yes, that's a big advantage of the [built-in interactive debugging support](https://wlandau.github.io/targets-manual/practice.html#interactive-debugging) in `targets`, as opposed to `tar_workspace()`. I added some more explanation in https://github.com/wlandau/targets-manual/commit/9f97ac72148837f8f000bcb5374b5a0bba04f334 to try to clear things up.

> To learn the package and test the core functionality, I tried out a common use-case for us. Currently, we use remake and an
internal wrapper package (example here) that provides shared cache, error handling, and task splitting (similar to branching) services on top of remake.

That's a really neat use case and a great example of `tar_group()` in action.

> One surprise digging into my example above was that tar_path(dat) returns _targets/objects/dat, despite that file not existing, and despite the path in the metadata being NA.

Looking back at the documentation, I realize now that the help file was not totally clear. My intention was to give `tar_path()` a much simpler niche: just identify the path without worrying if it exists. This gives advanced users some building blocks for custom cues, as in https://github.com/wlandau/targets/issues/131. Hopefully https://github.com/wlandau/targets/commit/2d72e3c06a0510992e9cb8596df24bca93bb93f7 clears that up.

> Should this return NA, or the children names, or the filepath with a wildcard for clarity?

I believe `tar_meta()` covers this.

```r
tar_meta(dat, path)$path[[1]]
#> [1] NA
tar_meta(dat, children)$children[[1]]
#> [1] "dat_ab76cb22" "dat_f2821206" "dat_b8dadf92"
```

> A common use-case I have is once we get to doing site-level things (plotting, modeling), I'll notice a problem and want to
see the original data that was pulled from the web service. Or, the combined target is huge, and I just want to look at the (much smaller) branch that containes a particular site. How do I know which branch file site "01115120" is in, for example? Because in the code above, my batches are groups of sites, I want a dataframe of site IDs with corresponding filenames of which branches they belong to. I was expecting that crosswalk information to exist in _targets/meta/meta. The manual spends a lot of time showing how you can check the pattern created, what order the branches are created, and tar_pattern gets close. I wonder if a function like tar_pattern could accept the name of a target with type == pattern, and it produces a dataframe that shows the existing filenames and corresponding values of the variables that were mapped over? Admittedly, it wasn't too terrible to come up with what I wanted (see below), so at a minimum maybe include an example somewhere in the manual when discussing patterns and branching for those that want a post hoc crosswalk?

That's a tough one because the metadata file does not actually store dependency relationships (to keep storage light) and there could be a mismatch between the old `_targets/meta/meta` and current `_targets.R`. So I implemented a new `tar_branches()` function that accepts a pattern definition and returns a data frame of branch names and their dependencies, all constructed from the metadata file alone.

```r
tar_branches(dat, map(site_batches))
#> # A tibble: 3 x 2
#>   dat          site_batches
#>   <chr>        <chr>
#> 1 dat_ab76cb22 site_batches_d249d09b
#> 2 dat_f2821206 site_batches_eaa4bb57
#> 3 dat_b8dadf92 site_batches_3a2d01e7
```

It's an improvement, but it's not perfect. If `site_batches` were a pattern, you could just call `readd(site_batches_eaa4bb57)` to check the site of branch `dat_f2821206`. But `site_batches` is a stem, so `site_batches_eaa4bb57` is a bud, not a branch. In other words, there is no actual value of `site_batches_eaa4bb57` in storage. (And I do not think it would be wise to reconstruct the actual data in each bud because it could get quite large.) We have a similar problem when we match `sites` and `states`.

```r
tar_branches(sites, map(states))
#> # A tibble: 2 x 2
#>   sites          states
#>   <chr>          <chr>
#> 1 sites_83036b8f states_ce340191
#> 2 sites_2b7c2325 states_14e6a3f8
```

So I think some use cases will require the user to proactively keep track of branch names (with `tar_name()`) and site names, and then pass that information forward to downstream branches. Fortunately, you are already working with data frames as much as possible, and I think it would be straightforward to manually append new columns. I added a new section to the manual to discuss this issue: <https://wlandau.github.io/targets-manual/dynamic.html#branch-provenance>.
