### 2021-01-06 #401 @wlandau
Update: after thinking more about your experience with the documentation, I decided to add a new ["How to get started section](https://wlandau.github.io/targets/#how-to-get-started) at the top of the README. Minutes 6-40 of the video, along with the walkthrough chapter of the manual, should help most people get started in under an hour. For continued learning, the first 3 chapters of the [short course](https://github.com/wlandau/targets-tutorial) have friendly hands-on practice, and then the example projects should get folks ready to start big projects of their own. I also rearranged the sections of the README to convey a more logical progression through the material.
