### 2021-01-06 #401 @wlandau
I also decided to delete `targets/vignettes/start.Rmd` because I feel it just wasn't working. It feels like anything I write there would be better off somewhere else: the README and reference website home page are more obvious and discoverable for the small stuff, and the two `bookdown` books are more natural homes for the large stuff. It's not that I disagree with vignettes in general. In fact, [`jagstargets`](https://wlandau.github.io/jagstargets/) and [`stantargets`](https://wlandau.github.io/stantargets/) each have two, and they fit quite well and serve crucial roles. But for a package system as large and sprawling as `targets`, it seems like the vignette system has both poor discoverability and little room to scale, and I find it hard to bring myself to put anything there.
