+++ README.Rmd
@@ -24,36 +24,35 @@ knitr::opts_chunk$set(
 
 The `targets` package is a [Make](https://www.gnu.org/software/make/)-like pipeline toolkit for Statistics and data science in R. With `targets`, you can maintain a reproducible workflow without repeating yourself. `targets` learns how your pipeline fits together, skips costly runtime for tasks that are already up to date, runs only the necessary computation, supports implicit parallel computing, abstracts files as R objects, and shows tangible evidence that the results match the underlying code and data. For deep dives into why and how to use `targets`, as well as help transitioning from [`drake`](https://github.com/ropensci/drake), please see the [user manual](https://wlandau.github.io/targets-manual).
 
+## How to get started
+
+If you are new to `targets`, here is the easiest way to begin.
+
+1. First, watch minutes 6 through 40 of the [New York Open Statistical Programming Meetup from December 2020](https://youtu.be/Gqn7Xn4d5NI).
+1. Next, read the [short walkthrough chapter](https://wlandau.github.io/targets-manual/walkthrough.html) of the [user manual](https://wlandau.github.io/targets-manual/).
+1. For some guided hand-on practice, sign up for a free [RStudio Cloud](https://rstudio.cloud) account and log into the [cloud workspace](https://rstudio.cloud/project/1699460) of the [official short course](https://github.com/wlandau/targets-tutorial/blob/main/README.md). The [R notebooks](https://github.com/wlandau/targets-tutorial) have interactive exercises to practice the machine learning workflow from [these slides](https://wlandau.github.io/targets-tutorial/). For new users, notebooks [`1-functions.Rmd`](https://github.com/wlandau/targets-tutorial/blob/main/1-functions.Rmd), [`2-pipelines.Rmd`](https://github.com/wlandau/targets-tutorial/blob/main/2-pipelines.Rmd), and [`3-changes.Rmd`](https://github.com/wlandau/targets-tutorial/blob/main/3-changes.Rmd) are more than enough.
+1. Try out one of the [example projects](https://wlandau.github.io/targets/index.html#example-projects) linked from the [reference website](https://wlandau.github.io/targets/index.html#example-projects). Each project has all the components of a real-world workflow, and each one links to a public [RStudio Cloud](https://rstudio.cloud) workspace where you can try out the code without needing to install anything locally.
+
 ## Installation
 
-Install the GitHub development version to access the latest features and patches.
+You can install the GitHub development version of `targets` to access the latest features and patches.
 
 ```{r, eval = FALSE}
 library(remotes)
 install_github("wlandau/targets")
 ```
 
-## Get started
+## Recorded talks
 
-The easiest way to quickly get started is to read the [walkthrough chapter of the user manual](https://wlandau.github.io/targets-manual/walkthrough.html). You might also watch the recorded talks linked below.
-
-## Help
-
-* Post to the [GitHub issue tracker](https://github.com/wlandau/targets/issues) to elicit help from the maintainer.
-* The [RStudio Community](https://community.rstudio.com/) forum is full of friendly enthusiasts of R and the tidyverse. Use the [`targets` tag](https://community.rstudio.com/tag/targets).
-* [Stack Overflow](https://stackoverflow.com/) broadcasts to the entire open source community. Use the [`targets-r-package` tag](https://stackoverflow.com/questions/tagged/targets-r-package).
+* [R/Pharma 2020 (9:24)](https://www.youtube.com/watch?v=GRqKJBaC5g4&list=PLMtxz1fUYA5C0YflXsR8EEAQXfjntlV1H&index=6)
[…]
