### 2021-01-06 #401 @maurolepore
@wlandau, I really like your new section https://wlandau.github.io/targets/#how-to-get-started. I did steps 1. and 2. and they did help me understand targets in about 1h. While I did that I took some notes. This is not strictly a review but I figured you might find some of this useful:

* It surprised me that the [README](https://github.com/wlandau/targets) file of targets did not show any example.

* The [chapter Walkthrough](https://wlandau.github.io/targets-manual/walkthrough.html) seems out of sync with the [source code](https://github.com/wlandau/targets-minimal). It seems that the book chapter hard-codes content rather than reference it from the source code. It seems that the source code changed and the manual became outdated. In general, code chunks of the manual that use `eval = FALSE` seem oudtated, including the directory tree of the targets-minimal repo, and the use of `list()` instead of `tar_pipeline()` in the file `_targets.R` (which causes `targets::tar_visnetwork()` to throw `Error : not a tar_pipeline() object. _targets.R must end with one.`).

* In the book I didn't find an "edit this file" icon. Is this intentional? (I use it a lot when I read bookdown books and want to help fix a typo).

* `tar_manifest()` and `tar_meta()` return a tibble but I see them wrapped in `as.data.frame()`. Is that because in that particular case you prefer the print method of vanilla dataframe?

* `targets::tar_visnetwork()` warns:
```
Warning message:
Strategy 'multiprocess' is deprecated in future (>= 1.20.0). Instead, explicitly specify either 'multisession' or 'multicore'. In the current R session, 'multiprocess' equals 'multicore'.
```
