### 2021-01-06 807f030
Always put concluding message at end of pipeline
