+++ R/class_timestamp.R
@@ -41,12 +41,8 @@ timestamp_class <- R6::R6Class(
       cli_workspace(target_get_name(target), time_stamp = TRUE)
     },
     report_end = function(progress = NULL) {
-      if (progress$uptodate()) {
-        cli_uptodate()
-      }
-      if (any(progress$warned$count > 0L)) {
-        cli_warned(progress$warned$count)
-      }
+      progress$cli_end(time_stamp = TRUE)
+      super$report_end(progress)
     }
   )
 )
