### 2021-01-06 1523367
Update news and shorten message
