### 2021-01-06 #401 @wlandau
Thanks for the feedback, @maurolepore.

> @wlandau, I really like your new section https://wlandau.github.io/targets/#how-to-get-started. I did steps 1. and 2. and they did help me understand targets in about 1h.

Glad to hear that helped.

> It surprised me that the README file of targets did not show any example.

Yes, I realize this unconventional, but I find it hard to find room because there are so many external resources that need to be visible. I feel the README is the most discoverable piece of documentation, so I am trying to keep it concise and avoid `drake`'s cache-all approach that buried important links under a lot of volume.

> The chapter Walkthrough seems out of sync with the source code. It seems that the book chapter hard-codes content rather than reference it from the source code. It seems that the source code changed and the manual became outdated. In general, code chunks of the manual that use eval = FALSE seem oudtated, including the directory tree of the targets-minimal repo,

I wanted to include more in https://github.com/wlandau/targets-minimal than I do at https://wlandau.github.io/targets-manual/walkthrough.html: more realistic workflow files, HPC, etc. So I just updated the manual to claim that the latter is an really abridged version of the former.

> and the use of list() instead of tar_pipeline() in the file _targets.R (which causes targets::tar_visnetwork() to throw Error : not a tar_pipeline() object. _targets.R must end with one.).

https://github.com/wlandau/targets/pull/254 is a recent update that deprecates `tar_pipeline()` in favor of `list()`. I feel this is a much friendlier interface.

> In the book I didn't find an "edit this file" icon. Is this intentional? (I use it a lot when I read bookdown books and want to help fix a typo).

I just forgot, thanks for the reminder. Just pushed an `_output.yml` file which should take care of that.

> tar_manifest() and tar_meta() return a tibble but I see them wrapped in as.data.frame(). Is that because in that particular case you prefer the print method of vanilla dataframe?

For `tar_manifest()` and `tar_progress()`, there is no real reason, and I just pushed a fix. For `tar_meta()` I used a plain data frame so users could see more columns.

> targets::tar_visnetwork() warns:

Yes, it is better to keep `future::plan()` commented out and to use the multisession plan instead of multiprocess. Should be fixed now in https://github.com/wlandau/targets-minimal and the cloud workspace.
