### 2021-01-06 #401 @wlandau
> I wanted to include more in https://github.com/wlandau/targets-minimal than I do at https://wlandau.github.io/targets-manual/walkthrough.html: more realistic workflow files, HPC, etc. So I just updated the manual to mention that the latter is an really abridged version of the former.

Just changed my mind on that. I simplified https://github.com/wlandau/targets-minimal so it is easier to digest and agrees with the manual.
