+++ DESCRIPTION
@@ -8,8 +8,8 @@ Description: The 'targets' package is a pipeline toolkit that brings together
   computation with optional parallel workers, abstracts files as
   R objects, and provides tangible evidence that the results match
   the underlying code and data. The methodology in this package
-  borrows from GNU 'Make' by Richard Stallman (2015, ISBN:978-9881443519)
-  and 'drake' by Will Landau (2018) <doi:10.21105/joss.00550>.
+  borrows from GNU 'Make' (2015, ISBN:978-9881443519)
+  and 'drake' (2018, <doi:10.21105/joss.00550>).
 Version: 0.0.0.9004
 License: MIT + file LICENSE
 URL: https://wlandau.github.io/targets/, https://github.com/wlandau/targets
