+++ man/targets-package.Rd
@@ -14,5 +14,5 @@ computations with optional parallel workers, abstracts files as
 R objects, and provides tangible evidence that the results match
 the underlying code and data. The methodology in this package
 borrows from GNU Make (2015, ISBN:978-9881443519)
-and 'drake' by Will Landau (2018, \url{doi:10.21105/joss.00550}).
+and \code{drake} by Will Landau (2018, \url{doi:10.21105/joss.00550}).
 }
