+++ inst/_targets.R
@@ -1,5 +1,9 @@
-# This is an example target script.
-# Read the tar_script() help file for details.
+# This is an example _targets.R file. Every
+# {targets} pipeline needs one.
+# Use tar_script() to create _targets.R and tar_edit()
+# to open it again for editing.
+# Then, run tar_make() to run the pipeline
+# and tar_read(summary) to view the results.
 
 # Define custom functions and other global objects.
 # This is where you write source(\"R/functions.R\")
