### 2021-01-07 #401 @tjmahr
I met the package author at an R-Unconference meetup in Chicago in 2019 (?). At the event, Will taught me how to use the drake package and I closed an issue for the package. Overall, I am a two-time contributor to drake.

  - No vignettes, but a full manual instead.

targets: All tests pass or skip. Package check succeeds. All examples run.
tarchetypes: All tests pass. Package check succeeds. All examples run. **There is a bad argument in `tar_map()`'s call to `tar_manifest()`.**

The documentation has this trick where it skips examples unless a system environment variable is set, instead of using `\dontrun` or `\donttest` directives. Would it be better to use these conventions?

I was able (I think) to run all the examples with the following code. It confirms that all help topics for user-facing functions have examples.

```r
library(magrittr)
library(purrr)
library(targets)
library(tarchetypes)
Sys.setenv(TARCHETYPES_LONG_EXAMPLES = "true")

shhh_example <- purrr::quietly(example)

d <- hsearch_db("targets")

results <- d$Base$Topic %>%
  lapply(
    shhh_example,
    character.only = TRUE,
    run.dontrun = TRUE
  )

results %>%
  map("warnings") %>%
  flatten_chr()
#> [1] "‘as_pipeline’ has a help file but no examples"
#> [2] "‘reexports’ has a help file but no examples"
#> [3] "‘rstudio_addin_tar_glimpse’ has a help file but no examples"
#> [4] "‘rstudio_addin_tar_load’ has a help file but no examples"
#> [5] "‘rstudio_addin_tar_make_bg’ has a help file but no examples"
#> [6] "‘rstudio_addin_tar_outdated’ has a help file but no examples"
#> [7] "‘rstudio_addin_tar_progress’ has a help file but no examples"
#> [8] "‘rstudio_addin_tar_read’ has a help file but no examples"
#> [9] "‘rstudio_addin_tar_target’ has a help file but no examples"
#>[10] "‘rstudio_addin_tar_visnetwork’ has a help file but no examples"
#>[11] "‘tar_watch_app_ui’ has a help file but no examples"
#>[12] "‘target_run_worker’ has a help file but no examples"
#>[13] "‘targets-package’ has a help file but no examples"

d2 <- hsearch_db("tarchetypes")

results2 <- d2$Base$Topic %>%
  lapply(
    shhh_example,
    character.only = TRUE,
    run.dontrun = TRUE
  )

results2 %>%
  map("warnings") %>%
  flatten_chr()
#> [1] "‘reexports’ has a help file but no examples"
#> [2] "‘tar_force_change’ has a help file but no examples"
#> [3] "‘tar_knit_run’ has a help file but no examples"
#> [4] "‘tar_render_rep_run’ has a help file but no examples"
#> [5] "‘tar_render_rep_run_params’ has a help file but no examples"
#> [6] "‘tar_render_run’ has a help file but no examples"
#> [7] "‘tar_rep_run’ has a help file but no examples"
#> [8] "‘tarchetypes-package’ has a help file but no examples"
```

---

First, a bit about the scope of my review. I am not able to confirm all of the functionality of this package, as I lack the experience and resources to test some of the high-performance or cloud-based features of targets. I used the package as local single-machine user. I also did not closely review the source code in the package as it is immense. My review, beyond the checkboxes above, focuses on the manual and using targets in two test cases.

### Notes on the manual

Whenever I use drake, I find myself with a browser window with tabs on the manual pacakges. My test-drive of targets was no different. Therefore, I closely read over the manual. Here are notes on each chapter.

The first two chapters of the manual are clear and brief. Chapter 1 introduce build pipelines and targets. Chapter 2 walks through a minimal example.

Chapter 3 describes best practices. Here the focus abruptly shifts from describing a style of programming (away from imperative, using functions instead of scripts) to describing how to use targets. I think there are two sets of best practices. One is a change in mindset and programming style (from a sequence of files that run R code to a sequence of function calls that run data through a pipeline), and the other is a set of tips for using targets. I recall that drake had a bunch of documentation (and even functions) for how to make this conceptual shift, but some more guidance here for users might be helpful.  Currently, the "clean, organized, modular, function-oriented style of programming" is mentioned but not much guidance is given about it. Just a basic sketch of how

```
01-data.R
02-preprocess.R
03-analysis.R
04-summaries.R
```

might be realized as

```
raw_data <- get_data(...)
data <- preprocess_data(raw_data, ...)
model <- fit_model(data, ...)
summary <- summarize_model(model, ...)
```

would be helpful.

The debugging section is really helpful and informative and I had missed these originally. It might make sense to promote debugging features to their own chapter.

Because Chapter 3 provides miscellaneous guidance about using targets, there should be a section on how to interface with other tools out of R. For example, how might I run a nodejs script to spellcheck or lint my markdown files produced by RMarkdown whenever they change? How might I perform some basic python task on a csv produced within the pipeline?

Chapter 4 talks about files. This section is well done. One thing I would impress upon users is that they might think of "data.csv" as a file. But it's just a string of characters. We have to use `tar_target(data, "data.csv", format = "file")` to tell targets that this string names a file that needs to be tracked.

Chapters 5 and Chapters 6 cover dynamic and static branching. I go back and forth on what I would like to see here. I like that each chapter starts with a summary of when to use each kind of branching, but the basic problem that branching is meant to solve--creating several related targets at once--is not fleshed out in either spot. This is just an observation and not a request.

In both my test projects, I went with static branching because I wanted the created targets to have meaningful names. Am I missing something about how to get good names with dynamic branches? The examples all seem to use trivial examples instead of fitting multiple models or loading different datasets.

Chapter 7 covers high performance computing. I am able to use `tar_make_future()` successfully. Does the default `worker` value (`1L`) mean that `tar_make_future()` by default is not parallel? In one of my projects, I had been using `furrr::future_map()` inside of a function to sweep through a large dataframe. Is there any guidance about having two layers of parallelism?

Chapter 8 covers cloud features. The details for Amazon S3 look detailed enough.

And additional thought, just about teaching or describing this package: Should targets should be explicitly described as a framework or language for describing R pipelines? Just as a ggplot2 is its own grammar of graphics and R's formula syntax describes linear models, targets provides a set of functions that provide a language for describing a build pipeline. For example, you eventually to learn about upstream and downstream targets, dynamic files, and branching.

#### Kicking the tires (1)

For my review, I decided to convert my website to use targets. Each post is an Rmd file and they are knitted if the .Rmd file is newer than the .md file. I have been only using timestamps but I would like to try something smarter and more portable. The build pipeline in principle is easy to describe:

* Knit a new md whenever the rmd file changes
* Spellcheck the rmd files whenever the rmd file changes

[The following comments predate the first review and Will's fixes of the issue.] Reading the package overview page, I do not see any indication for where to *get started quickly*. I see the manual linked as a resource for a deep dive. I see links to examples and talks. The vignettes only cover software design issues. Chapter 2 provides a walkthrough for an example analysis; here I figure out that I can use `tar_script()` to create a usethis-like placeholder file. This placeholder is well commented with sections and instructions. The only thing it's missing is that the user just needs to run `tar_make()` to run the build pipline. I think a Getting Started vignette would be useful, especially because getting started is pretty easy, as it turns out:

* Run `tar_script()` to create a `_targets.R` file.
* Edit that file. You can even use `tar_edit()` to open that file.
* Run `tar_make()`.

[Okay, based on other comments, Will thinks that vignettes are not the best way to document targets. I do think that it worthwhile to highlight how easy it is to add targets to a project using the three steps above.]

I knew from drake that I would need branching to work on my files and I know that files are something drake can struggle with. [Dynamic branching over files](https://wlandau.github.io/targets-manual/dynamic.html#dynamic-branching-over-files) provides the steps for registering dynamic input files. The manual notes that

> Dynamic branching over files is tricky. A target with `format = "file"` treats the entire set of files as an irreducible bundle. That means in order to branch over files downstream, each file must already have its own branch.

tarchetypes provides some sugar for registering file inputs with `tarchetypes::tar_files_input()`. But this way imposes dynamic branching where objects come in batches with uninformative names. My list of posts, for example, shows up as:

```
* run branch posts_31d3b445
* run branch posts_669f285f
* run branch posts_b4b851e3
* run branch posts_ae29f28f
```

To get better names, I switched to static branching and now find targets with meaningful names:

```
v skip target in_current_.._R.2016.08.04.fixing.apa.citations.from.pandoc.Rmd
v skip target in_current_.._R.2016.03.23.setnames.dots.Rmd
v skip target in_current_.._R.2016.08.15.recent.adventures.with.lazyeval.Rmd
[...]
```

Drake provided a way to customize names for branched-over objects but targets lacks this feature.

With a knitting workflow in place, I decided to add automatic spellchecking to the pipeline. This is an easy addition, and `tar_force()` allows me to print the spell check results only when there is a spelling mistake. 👍

```r
  tar_force(
    spellcheck_current_results,
    command = print(spellcheck_current),
    force = nrow(spellcheck_current) > 0
  )
```

### Kicking the tires (2)

I also tested targets by converting an existing analysis project from drake to targets. This involved reading from a database, cleaning the data, fitting models, doing a parametric bootstrap on the model coefficients and estimating bootstrap intervals on quantiles. It was easy to mechanically convert drake statements of the form `target = command` to `tar_target(target, command)`. The style of drake plans is more readable and interactive (you can highlight part of a plan and run it to debug a target). tarchetypes provides a `tar_plan()` for recreating that style but this review I stuck with the vanilla style.

Working on regular lists (instead of some `tar_list()` or some specialized function) made it easy for me to break up my build pipeline with a series of target lists. First, `targets_data = list(...)`, then `targets_models = list(...)`, then ending the file with `list(targets_data, target_models)`.

Again, I preferred static branching because I had a small number of models and wanted predictable names. The porting was straightforward, although a couple of times I pasted a `map()` from a drake target into a `tar_target(..., pattern)` triggering dynamic branching instead of dynamic branching.

***

### Some themes from testing

* I expect that files will be a hurdle for learners because of the need to target-ify names to track and target-file-ify files named by both names.
* Static branching is better at creating a small number of targets or targets where we are interested the target names, but dynamic branching is introduced first. I expect users to experiment with dynamic branching but run into issues with the names.
* I expect drake issues to trip over themselves by thinking that they know how this package works 😅.

### Comments as a drake user

I first tried targets in summer 2020 to overcome a strange issue in drake. On Windows, filenames have a [maximum length](https://www.howtogeek.com/266621/how-to-make-windows-10-accept-file-paths-over-260-characters/). drake's data-store produces files with very long names that would sometimes not work on Windows. I did not want to mess with the data-store or hashing configuration, so I tried targets and the problem was solved. This change in data storage is documented in the package's statement of need under ["lighter, friendlier data management"](https://wlandau.github.io/targets/articles/need.html#lighter-friendlier-data-management). In this case, switching to targets provided me with an immediate win.

* Switching to stricter conventions, like assuming a `_targets.R` file is a huge win for getting started and for studying other target-based projects.
* `tar_target(name, path)` and `tar_target(file, name, format = "file")` is better than having implicit file targets via `file_in()` or `file_out()`. It never sat well with having to do `file_in(!! variable)` for file names stored in variables.
