### 2021-01-07 637c1ec
Clean old vignettes
