### 2021-01-07 #401 @wlandau
# Reply to @tjmahr's review

Thank you so much for your review, TJ. You raised some really interesting points, and I enjoyed responding.

>  There is a bad argument in tar_map()'s call to tar_manifest().

Fixed in https://github.com/wlandau/tarchetypes/commit/01660c18edd4fa997c2001e1ccd657e3be26110a.

> The documentation has this trick where it skips examples unless a system environment variable is set, instead of using \dontrun or \donttest directives. Would it be better to use these conventions?

I definitely would prefer `\dontrun{}` and `\donttest{}`, but CRAN now routinely rejects packages that frequently use these directives. I just went through this with [`rfacts`](https://github.com/EliLillyCo/rfacts) late last year. CRAN rejected `\dontrun()` but accepted [this workaround](https://github.com/EliLillyCo/rfacts/blob/86f0b694e3aaffd13594d4d584a24ffe6cff85ba/R/run_facts.R#L16-L30) instead.

> Estimated hours spent reviewing: 24

Knowing what you know now, would it have made sense to split up this monolithic submission into 3 different ones? As I mentioned to @limnoliver, although I still view the 3 ([now 4](https://github.com/wlandau/targets-design)) repositories as one unified product, I am still concerned I judged poorly when I plowed ahead with everything at once.

> Chapter 3 describes best practices. Here the focus abruptly shifts from describing a style of programming (away from imperative, using functions instead of scripts) to describing how to use targets. I think there are two sets of best practices. One is a change in mindset and programming style (from a sequence of files that run R code to a sequence of function calls that run data through a pipeline), and the other is a set of tips for using targets. I recall that drake had a bunch of documentation (and even functions) for how to make this conceptual shift, but some more guidance here for users might be helpful. Currently, the "clean, organized, modular, function-oriented style of programming" is mentioned but not much guidance is given about it.

In https://github.com/wlandau/targets-manual/commit/986f71daff46cd6b6b8d64b97e9bad1d4ad86c73, I add a new chapter on functions (based the existing walkthrough chapter and `drake`'s functions chapter).

> The debugging section is really helpful and informative and I had missed these originally. It might make sense to promote debugging features to their own chapter.

Thanks. In https://github.com/wlandau/targets-manual/commit/986f71daff46cd6b6b8d64b97e9bad1d4ad86c73, I moved the debugging section to its own chapter.

> Because Chapter 3 provides miscellaneous guidance about using targets, there should be a section on how to interface with other tools out of R. For example, how might I run a nodejs script to spellcheck or lint my markdown files produced by RMarkdown whenever they change? How might I perform some basic python task on a csv produced within the pipeline?

Now discussed in https://github.com/wlandau/targets-manual/commit/6f5c3af901ff6b309ca6e7666db6b77e712a3df1.

> One thing I would impress upon users is that they might think of "data.csv" as a file. But it's just a string of characters. We have to use tar_target(data, "data.csv", format = "file") to tell targets that this string names a file that needs to be tracked.

I tried to make this clearer in https://github.com/wlandau/targets-manual/commit/92200ce201f8d11c6a8e8cb5cd567fe96c993c76.

> Chapters 5 and Chapters 6 cover dynamic and static branching. I go back and forth on what I would like to see here. I like that each chapter starts with a summary of when to use each kind of branching, but the basic problem that branching is meant to solve--creating several related targets at once--is not fleshed out in either spot. This is just an observation and not a request.

I took that for granted. As of https://github.com/wlandau/targets-manual/commit/b7654703d2423a847aa23bdd0106903fc83b2f6a, the dynamic branching chapter starts off by motivating the idea of branching in general..

> In both my test projects, I went with static branching because I wanted the created targets to have meaningful names. Am I missing something about how to get good names with dynamic branches?

Sounds like static branching fits your use cases much better, and that's totally fine. As I explain in https://github.com/wlandau/targets-manual/commit/b7654703d2423a847aa23bdd0106903fc83b2f6a, different kinds of branching fit different preferences.

There are many real-world scenarios where dynamic branching wins out: for example, the USGS use case @limnoliver described in her review using `tar_group()` to branch over a dataset. In that example, we don't even know what branches we are going to need (or even how many) until `tar_make()` reaches that point in the pipeline.

Personally, I use dynamic branching in the life sciences for simulating thousands of reps of clinical trials, as I describe in the second half of my [R/Pharma 2020 talk](https://www.youtube.com/watch?v=GRqKJBaC5g4) ([slides](https://wlandau.github.io/rpharma2020/), [source](https://github.com/wlandau/rpharma2020)). And I routinely use dynamic branching for [simulation-based calibration](https://arxiv.org/abs/1804.06788) to validate Bayesian models, which motivated examples [here](https://github.com/wlandau/targets-stan) and [here](https://wlandau.github.io/stantargets/articles/mcmc_rep.html.

As you saw, both kinds of branching are difficult to learn. One of my major goals with the [R Targetopia](https://wlandau.github.io/targetopia) is to completely abstract it away so users do not need to think about it for the most common domains of data science. For example, [`jagstargets`](https://github.com/wlandau/jagstargets) and [`stantargets`](https://github.com/wlandau/stantargets) use both forms of branching, but only behind the scenes.

Now for names. With dynamic branching, it is just not feasible to generate friendly names. A common first assumption is that if target `x` branches over target `y` and `y` has the value `c("a", "b", "c")`, then the branch names of `x` should be `x_a`, `x_b`, and `x_c`. But what if `y` is a grouped data frame? Or a list of Keras models? There is no fully automated way to get unique reproducible friendly informative names from any kind of arbitrary object. The remaining choices are:

1. Use a simple indexing scheme like 1, 2, 3, ..., or
2. Construct the suffixes from the hashes of the input values.

(1) does not work for `targets` because it invalidates too may branches. To demonstrate, let's say `x` has branches `x_1`, `x_2`, and `x_3` corresponding to inputs in `y <- c("a", "b", "c")`. But what if the user changes the input vector to `y <- c("a", "new_value", "b", "c")`? Not only do we need to create a fourth branch, we also need to invalidate `x_2` and `x_3` because they no longer match the corresponding indexes of `y`.

(2) is the only option I see because hashes are reproducible and the order does not matter. Unfortunately, hashes are unappealing to look at, but all I can do is choose a short hash function (xxhash32).

> The examples all seem to use trivial examples instead of fitting multiple models or loading different datasets.

I often find it difficult to cover features this complicated while expressing them through a sufficiently motivating yet pedagogically simple example. Hopefully example projects like https://github.com/wlandau/targets-stan make up for some of this.

> Chapter 7 covers high performance computing. I am able to use tar_make_future() successfully. Does the default worker value (1L) mean that tar_make_future() by default is not parallel?

`tar_make_future(workers = 1L)` spawns one remote worker and farms everything out to it. So yes, it's not parallel, but it's not the same as running locally either. But I still believe a default of `workers = 1L` is good practice because it consumes as few resources as possible, and I do not want to make assumptions about what other users' computers can handle. And believe it or not, `tar_make_future(workers = 1L)` is sometimes exactly what you want. Last summer, I used it to submit one really slow Bayesian model to a cluster because I knew the compute node would be a lot more powerful than the login node.

> Is there any guidance about having two layers of parallelism?

I mostly defer to packages `future` and `clustermq` on that one. Users of `targets` need to be reasonably familiar with each because the configuration goes in `_targets.R` (e.g. `future::plan()`). The `future` README describes [nested futures](https://github.com/HenrikBengtsson/future#nested-futures-and-evaluation-topologies), and the same principles apply to `targets`. In the two-level nesting case, the outer future controls parallelism among targets, and the inner future controls parallelism within targets.

In traditional HPC settings, nested parallelism is most useful when you farm out different targets to different nodes on a cluster and each target uses multiple local cores on its node. The current HPC chapter walks through how to do that for both `tar_make_clustermq()` and `tar_make_future()`.

> And additional thought, just about teaching or describing this package: Should targets should be explicitly described as a framework or language for describing R pipelines? Just as a ggplot2 is its own grammar of graphics and R's formula syntax describes linear models, targets provides a set of functions that provide a language for describing a build pipeline. For example, you eventually to learn about upstream and downstream targets, dynamic files, and branching.

I think of `targets` as more of a framework than a language, and I am on the fence about whether most users actually need to know that. The framework perspective is most relevant for package developers who write [target factories](https://wlandau.github.io/posts/2020-12-14-targetopia/) in [R Targetopia](https://wlandau.github.io/targetopia/) packages like [`jagstargets`](https://wlandau.github.io/jagstargets/) and [`stantargets`](https://wlandau.github.io/stantargets/). The reason why `targets` succeeds as a framework is precisely because it is *not* a language: i.e., it does not force users into a [DSL](http://adv-r.had.co.nz/dsl.html) for pipeline construction. `drake` is the opposite: `drake_plan()` is a [DSL](http://adv-r.had.co.nz/dsl.html) with maximum user convenience and minimum developer convenience. (It is really hard to build `drake` plan factories.) That is why we have no `drake`-topia.

> tar_script() to create a usethis-like placeholder file. This placeholder is well commented with sections and instructions. The only thing it's missing is that the user just needs to run tar_make() to run the build pipline.

I recently  added comments to that placeholder file to encourage users to try out tar_make() and tar_read().

> I think a Getting Started vignette would be useful...[Okay, based on other comments, Will thinks that vignettes are not the best way to document targets. I do think that it worthwhile to highlight how easy it is to add targets to a project using the three steps above.]

I am trying to use the [walkthorugh chapter](https://wlandau.github.io/targets-manual/walkthrough.html) as a "get started" vignette, and it is now linked from the "how to get started" section of the README. I also added these comments about tar_script() and tar_edit() to that chapter:

> Whereas files raw_data.csv and functions.R are typical user-defined components of a project-oriented workflow, _targets.R file is special. Every targets workflow needs a file called _targets.R in the project’s root directory. Functions tar_script() and tar_edit() can help you create one.

But to really get a handle on `_targets.R`, I think people will need hands-on practice. That's why I linked to the [targets-tutorial](https://github.com/wlandau/targets-tutorial) short course as part of the "how to get started" section of the README. The short course has lots of practice using tar_edit() and modifying `_targets.R`. And I tried to make the course as accessible as possible with [this RStudio Cloud workspace](https://rstudio.cloud/project/1699460).

> I knew from drake that I would need branching to work on my files and I know that files are something drake can struggle with. Dynamic branching over files provides the steps for registering dynamic input files. The manual notes that

For files, static branching is much easier, and it is free of all the perils identified in that section of the manual. For your use cases, I would stick with static-only branching.

> Drake provided a way to customize names for branched-over objects but targets lacks this feature.

`tar_map()` actually does have a `names` argument that lets you select elements from `values` to construct the names. If you want custom names, simply add another element to `values` and then identify it in `names`. Example:

``` r
library(targets)
tar_script({
  library(tarchetypes)
  tar_option_set(packages = "rmarkdown")
  tar_map(
    values = list(
      slug1 = c("file", "blog"),
      slug2 = c("new", "old"),
      file = c("fixing.apa.citations.from.pandoc.Rmd", "recent.adventures.with.lazyeval.Rmd")
    ),
    names = starts_with("slug"), # tidyselect is optional
    tar_target(out, render(file))
  )
})

tar_manifest()
#> # A tibble: 2 x 3
#>   name         command                                            pattern
#>   <chr>        <chr>                                              <chr>
#> 1 out_blog_old "render(\"recent.adventures.with.lazyeval.Rmd\")"  <NA>
#> 2 out_file_new "render(\"fixing.apa.citations.from.pandoc.Rmd\")" <NA>
```

`tar_eval()` may be an easier way to do this. It has fewer guardrails but more flexibility.

``` r
library(targets)
tar_script({
  library(tarchetypes)
  tar_option_set(packages = "rmarkdown")
  tar_eval(
    values = list(
      name = c("file_new", "blog_old"),
      file = c("fixing.apa.citations.from.pandoc.Rmd", "recent.adventures.with.lazyeval.Rmd")
    ),
    tar_target(name, render(file))
  )
})

tar_manifest()
#> # A tibble: 2 x 3
#>   name     command                                            pattern
#>   <chr>    <chr>                                              <chr>
#> 1 blog_old "render(\"recent.adventures.with.lazyeval.Rmd\")"  <NA>
#> 2 file_new "render(\"fixing.apa.citations.from.pandoc.Rmd\")" <NA>
```

> With a knitting workflow in place, I decided to add automatic spellchecking to the pipeline. This is an easy addition, and tar_force() allows me to print the spell check results only when there is a spelling mistake.

Glad to hear that was useful. [`triggers`](https://books.ropensci.org/drake/triggers.html) in `drake` were a struggle because I implemented them at a low level. This time around, `tarchetypes` can fortunately handle all that at the interface level (e.g. target factories like `tar_force()`).

> Again, I preferred static branching because I had a small number of models and wanted predictable names. The porting was straightforward, although a couple of times I pasted a map() from a drake target into a tar_target(..., pattern) triggering dynamic branching instead of dynamic branching.

Some people have proposed an automatic drake-to-targets converter, which I do not think is possible because of the advanced features. (We went through this same thing during the shift from remake to drake, but the converter never got developed.) I try to make up for this in the [drake chapter of the manual](https://wlandau.github.io/targets-manual/drake.html).

> I expect that files will be a hurdle for learners because of the need to target-ify names to track and target-file-ify files named by both names.

As I mentioned above, there is an extra step for names in `tar_map()`.

> Static branching is better at creating a small number of targets or targets where we are interested the target names, but dynamic branching is introduced first. I expect users to experiment with dynamic branching but run into issues with the names.

Conceptually, static branching is an outer layer on top of dynamic branching, and the former operates on the latter during [dynamic-within-static](https://wlandau.github.io/targets-manual/static.html#dynamic-within-static-branching). That is why I put the dynamic branching chapter first. [dynamic-within-static](https://wlandau.github.io/targets-manual/static.html#dynamic-within-static-branching) is really a static branching technique, but it assumes knowledge of dynamic branching.

The top of the dynamic branching chapter now states that with the exception of [dynamic-within-static](https://wlandau.github.io/targets-manual/static.html#dynamic-within-static-branching), those chapters can be read in any order.
