+++ R/tar_sitrep.R
@@ -112,7 +112,6 @@ tar_sitrep_inner <- function(
   fields_quosure,
   reporter
 ) {
-  pipeline_validate_lite(pipeline)
   names_all <- pipeline_get_names(pipeline)
   names <- eval_tidyselect(names_quosure, names_all)
   meta <- meta_init()
