+++ R/tar_make_clustermq.R
@@ -71,7 +71,6 @@ tar_make_clustermq_inner <- function(
   workers,
   log_worker
 ) {
-  pipeline_validate_lite(pipeline)
   names <- eval_tidyselect(names_quosure, pipeline_get_names(pipeline))
   clustermq_init(
     pipeline = pipeline,
