+++ R/utils_callr.R
@@ -24,6 +24,7 @@ callr_outer <- function(
 callr_inner <- function(targets_script, targets_function, targets_arguments) {
   value <- source(targets_script)$value
   targets_arguments$pipeline <- targets::as_pipeline(value)
+  targets::pipeline_validate_lite(targets_arguments$pipeline)
   do.call(targets_function, targets_arguments)
 }
 
