+++ R/tar_visnetwork.R
@@ -93,7 +93,6 @@ tar_visnetwork_inner <- function(
   level_separation,
   reporter
 ) {
-  pipeline_validate_lite(pipeline)
   allow <- eval_tidyselect(allow_quosure, pipeline_get_names(pipeline))
   exclude <- eval_tidyselect(exclude_quosure, pipeline_get_names(pipeline))
   network <- inspection_init(
