### 2021-01-08 af04492
Fix bug in deps and string in tar_target_raw()
