+++ tests/testthat/test-tar_target_raw.R
@@ -23,3 +23,35 @@ tar_test("tar_target_raw() receives options", {
   x <- tar_target_raw("x", "y")
   expect_equal(x$settings$format, "file")
 })
+
+tar_test("can set deps", {
+  out <- tar_target_raw(
+    "notebook",
+    command = quote(abc)
+  )
+  expect_equal(out$command$deps, "abc")
+  expect_true(grepl("abc", out$command$string))
+  out <- tar_target_raw(
+    "notebook",
+    command = quote(abc),
+    deps = "xyz"
+  )
+  expect_equal(out$command$deps, "xyz")
+  expect_true(grepl("abc", out$command$string))
+})
+
+tar_test("can set string", {
+  out <- tar_target_raw(
+    "notebook",
+    command = quote(abc)
+  )
+  expect_equal(out$command$deps, "abc")
+  expect_true(grepl("abc", out$command$string))
+  out <- tar_target_raw(
+    "notebook",
+    command = quote(abc),
+    string = "xyz"
+  )
+  expect_equal(out$command$deps, "abc")
+  expect_equal(out$command$string, "xyz")
+})
