### 2021-01-08 fd07fad
Restore igraph opt in adjacent vertices fn
