### 2021-01-08 #401 @maurolepore
> And I believe that completes my first round of replies. Please let me know if I missed anything.
-- @wlandau

> Upon changes being made, change the review status tag to 5/awaiting-reviewer-response, and request that reviewers indicate approval with the reviewer approval template.
-- Section 8.1.3 of <https://devguide.ropensci.org/editorguide.html>

--

@limnoliver and @tjmahr,

Thanks a lot for your amazing reviews. I understand @wlandau feels the changes he made address all reviewer's comments. Still I see one reviewer box unchecked: "Vignette(s) demonstrating major functionality that runs successfully locally".

@wlandau explained why he excluded vignettes. But such exclusion rose the reviewer's attention and fails to meet rOpenSci's guidelines. As a fresh editor, I feel uncomfortable accepting the exclusion just now and I propose:

* @limnoliver and @tjmahr, please share your final opinion about including/excluding vignettes.
* @wlandau please let me know if you prefer to (1) meet this guideline or (2) engage the chief editor. I'm totally happy either way.
