+++ vignettes/overview.Rmd
@@ -1,0 +1,58 @@
+---
+title: "An overview of targets"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{An overview of targets}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+This vignette is a high-level overview of `targets` and its educational materials. The goal is to summarize the major features of `targets` and direct users to the appropriate resources. It explains how to get started, and then it briefly describes each chapter of the [user manual](https://wlandau.github.io/targets-manual/).
+
+## What is `targets`?
+
+The `targets` R package is a [Make](https://www.gnu.org/software/make/)-like pipeline toolkit for Statistics and data science in R. With `targets`, you can maintain a reproducible workflow without repeating yourself. `targets` skips costly runtime for tasks that are already up to date, runs the necessary computation with implicit parallel computing, and abstracts files as R objects. A fully up-to-date `targets` pipeline is tangible evidence that the output aligns with the code and data, which substantiates trust in the results.
+
+## How to get started
+
+The top of the [reference website](https://wlandau.github.io/targets/) links to a number of materials to help new users start learning `targets`. It lists online talks, tutorials, books, and workshops in the order that a new user should consume them. The rest of the main page outlines a more comprehensive list of resources.
+
+## The walkthrough
+
+The [user manual](https://wlandau.github.io/targets-manual/) starts with a  [walkthrough](https://wlandau.github.io/targets-manual/walkthrough.html) chapter, a short tutorial to quickly started with `targets` using a simple example project. That project also has a [repository](https://github.com/wlandau/targets-minimal) with the source code and an [RStudio Cloud workspace](https://rstudio.cloud/project/1430691) that lets you try out the workflow in a web browser. Sign up for a free RStudio Cloud account, click on the link, and try out functions `tar_make()` and `tar_read()` in the R console.
+
+## Debugging
+
+The [debugging chapter](https://wlandau.github.io/targets-manual/debugging.html) describes two alternative built-in systems for troubleshooting errors. The first system uses workspaces, which let you load a target's dependencies into you R session. This way is usually preferred, especially with large pipelines on computing clusters, but it still may require some manual work. The second system launches an interactive debugger while the pipeline is actually running, which may not be feasible in some situations, but can often help you reach the problem more quickly.
+
+## Functions
+
+`targets` expects users to adopt a function-oriented style of programming. User-defined R functions are essential to express the complexities of data generation, analysis, and reporting. The [user manual](https://wlandau.github.io/targets-manual/) has a [whole chapter](https://wlandau.github.io/targets-manual/functions.html) dedicated to user-defined functions for data science, and it explains why they are important and how to use them in `targets`-powered pipelines.
+
+## Best practices
+
+The [best practices chapter](https://wlandau.github.io/targets-manual/practices.html) contains general recommendations about `targets` pipelines in the real world, including how to define good targets, what to do if your workflow is itself an R package, and how to diagnose performance issues.
+
+## External files and literate programming
+
+`targets` has special ways to include data files and literate programming reports in a pipeline. This functionality is optional in the general case, but it is necessary if you want a target to rerun in response to a change in a data file, or if you want an R Markdown report to re-render when an upstream target changes. The [files chapter](https://wlandau.github.io/targets-manual/files.html) walks through this functionality, from input data to parameterized R Markdown.
+
+## Dynamic branching
+
[…]
