### 2021-01-08 #401 @wlandau
Looks like vignettes are a more strict requirement than I originally assumed. I still struggle to use the vignette system for this, but I gave it another try today. The new [overview vignette](https://github.com/wlandau/targets/blob/main/vignettes/overview.Rmd) has a summary of `targets` as a whole and a short synopsis of each chapter in the user manual. The goal is to condense down that lengthy book and give people a high-level view of all the major features. This way, we have a vignette with a purpose, but we do not introduce new material that should be in the manual or the README.
