### 2021-01-08 #401 @maurolepore
Thanks @wlandau. I pinged other editors to check how strict the guideline is. I value critical thinking over box-ticking but I need to pick more brains. Thanks for your patience.
