### 2021-01-09 #401 @maurolepore
@wlandau, following a discussion with other editors I am now think your new  vignette "Overview" does meet the guidelines. I believe the guideline's goal is to ensure the package has enough documentation and your package does (via the manual). Yet this vignette "Overview" is useful because of what you wrote above and because some users may skip the website; the vignette can then point users to the website, ensuring they don't miss the great documentation of the package.

I am now waiting for @limnoliver and @tjmahr to say if they want to either (a) request more changes or (b) approve.
