### 2021-01-12 #401 @wlandau
Before I forget: @limnoliver, @tjmahr, and @maurolepore, would you like to be [listed as "reviewers"](https://github.com/ropensci/drake/blob/0ea2998ce8d2bbb822cf5a78d2466e67e18da997/DESCRIPTION#L73-L84) in the `DESCRIPTION` files of the 4 repos under review? If so, would you like me to include your ORCIDs?

It is amazing how much better the documentation has become as a result of your input. I am more optimistic now that new users will find `targets` accessible. Even though I feel like I already went down this road with `drake`, there is always much more to learn.
