+++ inst/paper.bib
@@ -3,7 +3,7 @@
   author = {William Michael Landau},
   title = {{targets: dynamic function-oriented {M}ake-like declarative pipelines for {R}}},
   year = {2020},
-  howpublished = {https://github.com/wlandau/targets},
+  howpublished = {https://github.com/ropensci/targets},
   DOI = {}
 }
 
