### 2021-01-12 f5de894
As requested (https://github.com/ropensci/software-review/issues/401#issuecomment-758849974) delete code of conduct file
