### 2021-01-12 #401 @maurolepore
DONE
> * Create a two-person team in rOpenSci’s “ropensci” GitHub organization, named for the package, with yourself and the package author as members.
> * Go to the repository settings in rOpenSci’s “ropensci” GitHub organization and give the author “Admin” access to the repository.

----
> * If authors maintain a gitbook that is at least partly about their package, contact an rOpenSci staff member so they might contact the authors about transfer to the ropensci-books GitHub organisation.`**`

`**` checking about this.

----

> * Nominate a package to be featured in an rOpenSci blog post or tech note if you think it might be of high interest. Please note in the software review issue one or two things the author could highlight, and tag @ropensci/blog-editors

 @ropensci/blog-editors, I think this package might be of high interest. I suspect many readers will be familiar with the drake package and may wonder how it compares to targets and how to migrate.
