### 2021-01-12 3ecbeaa
Link to rOpenSci code of conduct
https://github.com/ropensci/software-review/issues/401#issuecomment-758849974
