+++ README.md
@@ -137,14 +137,10 @@ and [`tarchetypes`](https://docs.ropensci.org/tarchetypes).
     entire open source community. Use the [`targets-r-package`
     tag](https://stackoverflow.com/questions/tagged/targets-r-package).
 
-## Participation
+## Code of conduct
 
-Development is a community effort, and we welcome discussion and
-contribution. By participating in this project, you agree to abide by
-the [code of
-conduct](https://github.com/ropensci/targets/blob/main/CODE_OF_CONDUCT.md)
-and the [contributing
-guide](https://github.com/ropensci/targets/blob/main/CONTRIBUTING.md).
+Please note that this package is released with a [Contributor Code of
+Conduct](https://ropensci.org/code-of-conduct/).
 
 ## Citation
 
