+++ README.Rmd
@@ -84,9 +84,9 @@ The R Targetopia consists of specialized workflow frameworks tailored to individ
 * The [RStudio Community](https://community.rstudio.com/) forum is full of friendly enthusiasts of R and the tidyverse. Use the [`targets` tag](https://community.rstudio.com/tag/targets).
 * [Stack Overflow](https://stackoverflow.com/) broadcasts to the entire open source community. Use the [`targets-r-package` tag](https://stackoverflow.com/questions/tagged/targets-r-package).
 
-## Participation
+## Code of conduct
 
-Development is a community effort, and we welcome discussion and contribution. By participating in this project, you agree to abide by the [code of conduct](https://github.com/ropensci/targets/blob/main/CODE_OF_CONDUCT.md) and the [contributing guide](https://github.com/ropensci/targets/blob/main/CONTRIBUTING.md).
+Please note that this package is released with a [Contributor Code of Conduct](https://ropensci.org/code-of-conduct/).
 
 ## Citation
 
