+++ .github/ISSUE_TEMPLATE/bottleneck.md
@@ -9,7 +9,7 @@ assignees: wlandau
 
 ## Prework
 
-* [ ] Read and agree to the [code of conduct](https://github.com/ropensci/targets/blob/main/CODE_OF_CONDUCT.md) and [contributing guidelines](https://github.com/ropensci/targets/blob/main/CONTRIBUTING.md).
+* [ ] Read and agree to the [code of conduct](https://ropensci.org/code-of-conduct/) and [contributing guidelines](https://github.com/ropensci/targets/blob/main/CONTRIBUTING.md).
 * [ ] If there is [already a relevant issue](https://github.com/ropensci/targets/issues), whether open or closed, comment on the existing thread instead of posting a new issue.
 * [ ] Post a [minimal reproducible example](https://www.tidyverse.org/help/) like [this one](https://github.com/ropensci/targets/issues/256#issuecomment-754229683) so the maintainer can troubleshoot the problems you identify. A reproducible example is:
     * [ ] **Runnable**: post enough R code and data so any onlooker can create the error on their own computer.
