+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,6 +1,6 @@
 # Prework
 
-* [ ] I understand and agree to the [code of conduct](https://github.com/ropensci/targets/blob/main/CODE_OF_CONDUCT.md) and the [contributing guidelines](https://github.com/ropensci/targets/blob/main/CONTRIBUTING.md).
+* [ ] I understand and agree to the [code of conduct](https://ropensci.org/code-of-conduct/) and the [contributing guidelines](https://github.com/ropensci/targets/blob/main/CONTRIBUTING.md).
 * [ ] I have already submitted an issue to the [issue tracker](http://github.com/ropensci/targets/issues) to discuss my idea with the maintainer.
 * [ ] This pull request is not a [draft](https://github.blog/2019-02-14-introducing-draft-pull-requests).
 
