+++ NEWS.md
@@ -1,4 +1,4 @@
-# targets 0.0.0.9004
+# targets 0.0.1
 
 ## Enhancements
 
