+++ R/class_builder.R
@@ -214,7 +214,7 @@ builder_handle_error <- function(target, pipeline, scheduler, meta) {
   if (identical(target$settings$error, "workspace")) {
     builder_save_workspace(target, pipeline, scheduler)
   }
-  trn (
+  trn(
     identical(target$settings$error, "continue"),
     msg_run("Error: ", target$metrics$error, "\n"),
     throw_run(target$metrics$error)
