### 2021-01-13 #401 @wlandau
My apologies @tjmahr and @limnoliver, somehow I managed to miss this from your reviews:

> [x] Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

I will add you both.
