### 2021-01-13 #401 @wlandau
`"rev"` roles now added.
