### 2021-01-14 #401 @wlandau
Thanks so much, @maelle! As I discussed with @maurolepore, I transferred `wlandau/targets-manual` and `wlandau/targets-design` to `ropensci-books`.

Just one last thing: would you grant me access to the settings? I would like to rename `ropensci-books/targets-manual` to the more concise `ropensci-books/targets`.
