### 2021-01-14 #401 @wlandau
@maurolepore, I believe I addressed the items in https://github.com/ropensci/software-review/issues/401#issuecomment-758849974 (except the URL redirects, which are currently not working for the `ropensci-books` repos). Please let me know if there is anything I missed.
