+++ .github/workflows/check.yaml
@@ -13,7 +13,7 @@ jobs:
       matrix:
         config:
           - {os: windows-latest, r: '4.0'}
-          - {os: macOS-latest, r: 'release'}
+          - {os: macOS-latest, r: '4.0'}
           - {os: ubuntu-16.04,   r: '3.5', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
           - {os: ubuntu-16.04,   r: '3.6', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
           - {os: ubuntu-16.04, r: 'release', rspm: "https://packagemanager.rstudio.com/cran/__linux__/xenial/latest"}
