### 2021-01-14 #401 @wlandau
I just submitted `targets` to CRAN, and I will submit `tarchetypes` as soon as `targets` is accepted and the binaries are available. My colleagues and I look forward to the upcoming cascade of internal production releases that will follow.

In the near future, I would like to post an rOpenSci submission for [`jagstargets`](https://github.com/wlandau/jagstargets), an [R Targetopia](https://wlandau.github.io/targetopia/) package for Bayesian data analysis with JAGS. [`jagstargets`](https://github.com/wlandau/jagstargets) should be a nice warmup for [`stantargets`](https://github.com/wlandau/stantargets), a similar but larger package built on `targets` and Stan.

This review process was incredibly rewarding, and `targets` is much more accessible as a result. I cannot thank you enough!
